### 2023-07-08 #5633 @editorialbot
[PRE REVIEW]: Raphtory: The temporal graph engine for Rust and Python
**Submitting author:** @narnolddd (<a href="http://orcid.org/0000-0001-6396-4788">Naomi Arnold</a>)
**Repository:** https://github.com/Pometry/Raphtory
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.2
