+++ raphtory-graphql/src/main.rs
@@ -12,5 +12,5 @@ mod server;
 async fn main() {
     dotenv().ok();
     let graph_directory = env::var("GRAPH_DIRECTORY").unwrap_or("/tmp/graphs".to_string());
-    RaphtoryServer::new(&graph_directory).run().await.unwrap()
+    RaphtoryServer::from_directory(&graph_directory).run().await.unwrap()
 }
