+++ raphtory-graphql/src/data.rs
@@ -14,8 +14,25 @@ pub(crate) struct Data {
 }
 
 impl Data {
-    pub fn new(graphs: HashMap<String, DynamicGraph>) -> Self {
-        let graphs: HashMap<String, IndexedGraph<DynamicGraph>> = graphs
+    pub fn from_map(graphs: HashMap<String, DynamicGraph>) -> Self {
+        let graphs = Self::convert_graphs(graphs);
+        Self { graphs }
+    }
+
+    pub fn from_directory(directory_path: &str) -> Self {
+        let graphs = Self::load_from_file(directory_path);
+        Self { graphs }
+    }
+
+    pub fn from_map_and_directory(graphs: HashMap<String, DynamicGraph>, directory_path: &str) -> Self {
+        let mut graphs = Self::convert_graphs(graphs);
+        let mut graphs_from_files = Self::load_from_file(directory_path);
+        graphs_from_files.extend(graphs);
+        Self { graphs: graphs_from_files }
+    }
+
+    fn convert_graphs(graphs: HashMap<String, DynamicGraph>) -> HashMap<String, IndexedGraph<DynamicGraph>> {
+        graphs
             .into_iter()
             .map(|(name, g)| {
                 (
@@ -23,14 +40,13 @@ impl Data {
                     IndexedGraph::from_graph(&g).expect("Unable to index graph"),
                 )
             })
-            .collect();
-        Self { graphs }
+            .collect()
     }
 
-    pub fn load(directory_path: &str) -> Self {
+    fn load_from_file(path: &str) -> HashMap<String, IndexedGraph<DynamicGraph>> {
         let mut valid_paths = HashSet::<String>::new();
 
-        for entry in WalkDir::new(directory_path)
+        for entry in WalkDir::new(path)
             .into_iter()
             .filter_map(|e| e.ok())
         {
@@ -73,7 +89,6 @@ impl Data {
                 };
             }).map(|(name, g)| (name, IndexedGraph::from_graph(&g.into_dynamic()).expect("Unable to index graph")))
             .collect();
-
-        Self { graphs }
+        graphs
     }
 }
