### 2023-07-10 2fe935c
add the slight improvement to add_edge
