+++ raphtory/src/algorithms/local_clustering_coefficient.rs
@@ -38,7 +38,7 @@
 //! ];
 //!
 //! for (t, src, dst) in &vs {
-//!     g.add_edge(*t, *src, *dst, [], None);
+//!     g.add_edge(*t, *src, *dst, EMPTY, None);
 //! }
 //!
 //! let actual = (1..=5)
@@ -79,10 +79,10 @@ pub fn local_clustering_coefficient<G: GraphViewOps, V: Into<VertexRef>>(
 #[cfg(test)]
 mod clustering_coefficient_tests {
     use super::local_clustering_coefficient;
-    use crate::db::{
+    use crate::{db::{
         api::{mutation::AdditionOps, view::*},
         graph::graph::Graph,
-    };
+    }, prelude::EMPTY};
 
     #[test]
     fn clusters_of_triangles() {
@@ -98,7 +98,7 @@ mod clustering_coefficient_tests {
         ];
 
         for (t, src, dst) in &vs {
-            g.add_edge(*t, *src, *dst, [], None).unwrap();
+            g.add_edge(*t, *src, *dst, EMPTY, None).unwrap();
         }
 
         let expected = vec![0.33333334, 1.0, 1.0, 0.0, 0.0];
