### 2023-07-10 10ddf8f
add properties object for edges and some other improvements
