+++ raphtory/src/db/graph/edge.rs
@@ -5,7 +5,10 @@
 //! and can have properties associated with them.
 //!
 
-use crate::db::api::properties::internal::{StaticProperties, StaticPropertiesOps};
+use crate::db::api::properties::internal::{
+    StaticProperties, StaticPropertiesOps, TemporalProperties, TemporalPropertiesOps,
+    TemporalPropertyViewOps,
+};
 use crate::db::api::view::internal::{GraphPropertiesOps, Static};
 use crate::{
     core::{
@@ -69,6 +72,34 @@ impl<G: GraphViewOps> StaticPropertiesOps for EdgeView<G> {
 
     fn get_static_property(&self, key: &str) -> Option<Prop> {
         self.graph.static_edge_prop(self.edge, key)
+    }
+}
+
+impl<G: GraphViewOps> TemporalPropertyViewOps for EdgeView<G> {
+    fn temporal_history(&self, id: &String) -> Vec<i64> {
+        self.graph
+            .temporal_edge_prop_vec(self.edge, id)
+            .into_iter()
+            .map(|(t, _)| t)
+            .collect()
+    }
+
+    fn temporal_values(&self, id: &String) -> Vec<Prop> {
+        self.graph
+            .temporal_edge_prop_vec(self.edge, id)
+            .into_iter()
+            .map(|(_, v)| v)
+            .collect()
+    }
+}
+
+impl<G: GraphViewOps> TemporalPropertiesOps for EdgeView<G> {
+    fn temporal_property_keys(&self) -> Vec<String> {
+        self.graph.temporal_edge_prop_names(self.edge)
+    }
+
+    fn get_temporal_property(&self, key: &str) -> Option<String> {
+        (!self.graph.temporal_edge_prop_vec(self.edge, key).is_empty()).then_some(key.to_owned())
     }
 }
 
@@ -143,40 +174,12 @@ impl<G: GraphViewOps> EdgeListOps for BoxedIter<EdgeView<G>> {
     /// Specifies the associated type for the iterator over edges.
     type IterType<T> = Box<dyn Iterator<Item = T> + Send>;
 
-    fn has_property(self, name: String, include_static: bool) -> BoxedIter<bool> {
-        Box::new(self.map(move |e| e.has_property(&name, include_static)))
-    }
-
-    fn property(self, name: String, include_static: bool) -> BoxedIter<Option<Prop>> {
-        Box::new(self.map(move |e| e.property(&name, include_static)))
-    }
-
-    fn properties(self, include_static: bool) -> BoxedIter<HashMap<String, Prop>> {
-        Box::new(self.map(move |e| e.properties(include_static)))
-    }
-
-    fn property_names(self, include_static: bool) -> BoxedIter<Vec<String>> {
-        Box::new(self.map(move |e| e.property_names(include_static)))
-    }
-
-    fn has_static_property(self, name: String) -> BoxedIter<bool> {
-        Box::new(self.map(move |e| e.has_static_property(&name)))
-    }
-
-    fn static_property(self, name: String) -> BoxedIter<Option<Prop>> {
-        Box::new(self.map(move |e| e.static_property(&name)))
+    fn properties(self) -> Self::IterType<TemporalProperties<Self::Edge>> {
+        Box::new(self.map(move |e| e.properties()))
     }
 
     fn static_properties(self) -> Self::IterType<StaticProperties<EdgeView<G>>> {
         Box::new(self.map(move |e| e.static_properties()))
-    }
-
-    fn property_history(self, name: String) -> BoxedIter<Vec<(i64, Prop)>> {
-        Box::new(self.map(move |e| e.property_history(&name)))
-    }
-
[…]
