### 2023-07-10 b20a6fa
remove trait Properties
