### 2023-07-10 dd42e42
initial function to load from csv
