### 2023-07-10 4750496
refactor the trait by bringing Properties back
