+++ raphtory/src/db/graph/views/window_graph.rs
@@ -29,9 +29,9 @@
 //! use raphtory::db::api::view::*;
 //!
 //! let graph = Graph::new();
-//! graph.add_edge(0, 1, 2, EMPTY, None).unwrap();
-//! graph.add_edge(1, 1, 3, EMPTY, None).unwrap();
-//! graph.add_edge(2, 2, 3, EMPTY, None).unwrap();
+//! graph.add_edge(0, 1, 2, NO_PROPS, None).unwrap();
+//! graph.add_edge(1, 1, 3, NO_PROPS, None).unwrap();
+//! graph.add_edge(2, 2, 3, NO_PROPS, None).unwrap();
 //!
 //!  let wg = graph.window(0, 1);
 //!  assert_eq!(wg.edge(1, 2, None).unwrap().src().id(), 1);
@@ -434,8 +434,8 @@ impl<G: GraphViewOps> GraphOps for WindowedGraph<G> {
 /// use raphtory::prelude::*;
 ///
 /// let graph = Graph::new();
-/// graph.add_edge(0, 1, 2, EMPTY, None).unwrap();
-/// graph.add_edge(1, 2, 3, EMPTY, None).unwrap();
+/// graph.add_edge(0, 1, 2, NO_PROPS, None).unwrap();
+/// graph.add_edge(1, 2, 3, NO_PROPS, None).unwrap();
 /// let windowed_graph = graph.window(0, 1);
 /// ```
 impl<G: GraphViewOps> WindowedGraph<G> {
@@ -493,7 +493,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, src, dst) in &vs {
-            g.add_edge(*t, *src, *dst, EMPTY, None).unwrap();
+            g.add_edge(*t, *src, *dst, NO_PROPS, None).unwrap();
         }
 
         let wg = WindowedGraph::new(g, -1, 1);
@@ -523,7 +523,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, src, dst) in vs {
-            g.add_edge(t, src, dst, EMPTY, None).unwrap();
+            g.add_edge(t, src, dst, NO_PROPS, None).unwrap();
         }
 
         let wg = g.window(i64::MIN, i64::MAX);
@@ -545,7 +545,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, src, dst) in &vs {
-            g.add_edge(*t, *src, *dst, EMPTY, None).unwrap();
+            g.add_edge(*t, *src, *dst, NO_PROPS, None).unwrap();
         }
 
         let wg = WindowedGraph::new(g, -1, 1);
@@ -566,7 +566,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, v) in &vs {
-            g.add_vertex(*t, *v, EMPTY)
+            g.add_vertex(*t, *v, NO_PROPS)
                 .map_err(|err| println!("{:?}", err))
                 .ok();
         }
@@ -591,7 +591,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, v) in &vs {
-            g.add_vertex(*t, *v, EMPTY)
+            g.add_vertex(*t, *v, NO_PROPS)
                 .map_err(|err| println!("{:?}", err))
                 .ok();
         }
@@ -641,7 +641,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, e) in &edges {
-            g.add_edge(*t, e.0, e.1, EMPTY, None).unwrap();
+            g.add_edge(*t, e.0, e.1, NO_PROPS, None).unwrap();
         }
 
         let start = edges.get(rand_start_index).expect("start index in range").0;
@@ -724,7 +724,7 @@ mod views_test {
 
         for dst in dsts {
             let t = 1;
[…]
