### 2023-07-10 de93273
refactor the error handling for csv load
