### 2023-07-10 6037950
initial version of creating a polaris df from polars and fixing all the str& their dodgy prelude breaks
