### 2023-07-10 ab9e84c
Rusty apis (#1101)
* improve slightly the rust prop api

* add the slight improvement to add_edge

* remove trait Properties

* fix issue in benchmarks

* refactor the trait by bringing Properties back
