+++ raphtory/src/db/graph/views/window_graph.rs
@@ -29,9 +29,9 @@
 //! use raphtory::db::api::view::*;
 //!
 //! let graph = Graph::new();
-//! graph.add_edge(0, 1, 2, [], None).unwrap();
-//! graph.add_edge(1, 1, 3, [], None).unwrap();
-//! graph.add_edge(2, 2, 3, [], None).unwrap();
+//! graph.add_edge(0, 1, 2, NO_PROPS, None).unwrap();
+//! graph.add_edge(1, 1, 3, NO_PROPS, None).unwrap();
+//! graph.add_edge(2, 2, 3, NO_PROPS, None).unwrap();
 //!
 //!  let wg = graph.window(0, 1);
 //!  assert_eq!(wg.edge(1, 2, None).unwrap().src().id(), 1);
@@ -39,7 +39,7 @@
 
 use crate::{
     core::{
-        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, VID, EID},
+        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, EID, VID},
         utils::time::IntoTime,
         Direction, Prop,
     },
@@ -246,10 +246,12 @@ impl<G: GraphViewOps> TimeSemantics for WindowedGraph<G> {
 /// This trait provides operations to a `WindowedGraph` used internally by the `GraphWindowSet`.
 /// *Note: All functions in this are bound by the time set in the windowed graph.
 impl<G: GraphViewOps> GraphOps for WindowedGraph<G> {
-
     fn find_edge_id(&self, e_id: EID) -> Option<EdgeRef> {
         let e_ref = self.graph.find_edge_id(e_id)?;
-        if self.graph.include_edge_window(e_ref, self.t_start .. self.t_end) {
+        if self
+            .graph
+            .include_edge_window(e_ref, self.t_start..self.t_end)
+        {
             Some(e_ref)
         } else {
             None
@@ -432,8 +434,8 @@ impl<G: GraphViewOps> GraphOps for WindowedGraph<G> {
 /// use raphtory::prelude::*;
 ///
 /// let graph = Graph::new();
-/// graph.add_edge(0, 1, 2, [], None).unwrap();
-/// graph.add_edge(1, 2, 3, [], None).unwrap();
+/// graph.add_edge(0, 1, 2, NO_PROPS, None).unwrap();
+/// graph.add_edge(1, 2, 3, NO_PROPS, None).unwrap();
 /// let windowed_graph = graph.window(0, 1);
 /// ```
 impl<G: GraphViewOps> WindowedGraph<G> {
@@ -491,7 +493,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, src, dst) in &vs {
-            g.add_edge(*t, *src, *dst, [], None).unwrap();
+            g.add_edge(*t, *src, *dst, NO_PROPS, None).unwrap();
         }
 
         let wg = WindowedGraph::new(g, -1, 1);
@@ -521,7 +523,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, src, dst) in vs {
-            g.add_edge(t, src, dst, [], None).unwrap();
+            g.add_edge(t, src, dst, NO_PROPS, None).unwrap();
         }
 
         let wg = g.window(i64::MIN, i64::MAX);
@@ -543,7 +545,7 @@ mod views_test {
         let g = Graph::new();
 
         for (t, src, dst) in &vs {
-            g.add_edge(*t, *src, *dst, [], None).unwrap();
+            g.add_edge(*t, *src, *dst, NO_PROPS, None).unwrap();
         }
 
         let wg = WindowedGraph::new(g, -1, 1);
@@ -564,7 +566,7 @@ mod views_test {
[…]
