+++ raphtory/src/db/api/mutation/mod.rs
@@ -3,22 +3,23 @@ mod deletion_ops;
 pub mod internal;
 mod property_addition_ops;
 
-use crate::core::Prop;
 pub use addition_ops::AdditionOps;
 pub use deletion_ops::DeletionOps;
 pub use property_addition_ops::PropertyAdditionOps;
+
+use crate::prelude::Prop;
 
 pub trait Properties {
     fn collect_properties(self) -> Vec<(String, Prop)>;
 }
 
-impl<T> Properties for T
+impl<S:AsRef<str>, P:Into<Prop>, PI> Properties for PI
 where
-    T: IntoIterator<Item = (String, Prop)>,
+    PI: IntoIterator<Item = (S, P)>,
 {
     fn collect_properties(self) -> Vec<(String, Prop)> {
         self.into_iter()
-            .map(|(k, v)| (k.into(), v.into()))
+            .map(|(k, v)| (k.as_ref().to_string(), v.into()))
             .collect()
     }
 }
