+++ raphtory/src/db/graph/vertex.rs
@@ -473,9 +473,9 @@ mod vertex_test {
     #[test]
     fn test_earliest_time() {
         let g = Graph::new();
-        g.add_vertex(0, 1, []).unwrap();
-        g.add_vertex(1, 1, []).unwrap();
-        g.add_vertex(2, 1, []).unwrap();
+        g.add_vertex(0, 1, NO_PROPS).unwrap();
+        g.add_vertex(1, 1, NO_PROPS).unwrap();
+        g.add_vertex(2, 1, NO_PROPS).unwrap();
         let mut view = g.at(1);
         assert_eq!(view.vertex(1).expect("v").earliest_time().unwrap(), 0);
         assert_eq!(view.vertex(1).expect("v").latest_time().unwrap(), 1);
@@ -488,13 +488,13 @@ mod vertex_test {
     #[test]
     fn test_properties() {
         let g = Graph::new();
-        let props = [("test".to_string(), Prop::Str("test".to_string()))];
-        g.add_vertex(0, 1, []).unwrap();
+        let props = [("test", "test")];
+        g.add_vertex(0, 1, NO_PROPS).unwrap();
         g.add_vertex(2, 1, props.clone()).unwrap();
 
         let v1 = g.vertex(1).unwrap();
         let v1_w = g.window(0, 1).vertex(1).unwrap();
-        assert_eq!(v1.properties(false), props.into());
+        assert_eq!(v1.properties(false), props.into_iter().map(|(k, v)| (k.to_string(), v.as_prop())).collect());
         assert_eq!(v1_w.properties(false), HashMap::default())
     }
 }
