### 2023-07-10 b9034bc
add Rust Benchmark (cargo) benchmark result for 32cec73c84bbfa7f5773d76d139bd4f7212be7a5
