### 2023-07-10 80ae2b4
add parsing of src dst and time
