+++ raphtory/src/python/utils/errors.rs
@@ -1,6 +1,5 @@
 use crate::{
-    core::utils::{errors::GraphError, time::error::ParseTimeError},
-    graph_loader::source::csv_loader::CsvErr,
+    core::utils::{errors::{GraphError, CsvErr}, time::error::ParseTimeError},
 };
 use pyo3::{exceptions::PyException, PyErr};
 use std::error::Error;
