### 2023-07-10 d217406
Running the server in the background
