+++ python/Cargo.toml
@@ -24,6 +24,7 @@ raphtory_core = {path = "../raphtory", version = "0.4.3", features=["python"], p
 raphtory-graphql = {path = "../raphtory-graphql", version = "0.4.3"}
 openssl = { version = "0.10", features = ["vendored"] } # DO NOT REMOVE IT BREAKS PYTHON RELEASE
 
+
 [features]
 extension-module = ["pyo3/extension-module"]
 default = ["extension-module"]
