+++ raphtory/Cargo.toml
@@ -36,7 +36,6 @@ rustc-hash = "1.1.0"
 serde = { version = "1", features = ["derive","rc"] }
 sorted_vector_map = "0.1"
 tempdir = "0.3"
-# tokio = { version = "1.27.0", features = ["full"] }
 thiserror = "1"
 twox-hash = "1.6.3"
 uuid = { version = "1.3.0", features = ["v4"] }
