### 2023-07-11 be67aed
implement properties for Graph
