+++ raphtory/src/db/api/properties/internal.rs
@@ -2,6 +2,7 @@ use crate::core::entities::properties::props::DictMapper;
 use crate::core::entities::properties::tprop::TProp;
 use crate::core::{Prop, PropUnwrap};
 use crate::db::api::mutation::Properties;
+use crate::db::api::view::internal::Base;
 use crate::prelude::Graph;
 use chrono::NaiveDateTime;
 use std::iter::Zip;
@@ -14,7 +15,9 @@ pub trait CorePropertiesOps {
     fn static_prop(&self, id: usize) -> Option<&Prop>;
 }
 
-pub trait TemporalPropertyViewOps<Key = String> {
+pub type Key = String; //Fixme: This should really be the internal usize index but that means more reworking of the low-level api
+
+pub trait TemporalPropertyViewOps {
     fn temporal_value(&self, id: &Key) -> Option<Prop> {
         self.temporal_values(id).last().cloned()
     }
@@ -40,7 +43,7 @@ pub trait StaticPropertiesOps {
     fn get_static_property(&self, key: &str) -> Option<Prop>;
 }
 
-pub trait TemporalPropertiesOps<Key: 'static = String>: TemporalPropertyViewOps<Key> {
+pub trait TemporalPropertiesOps {
     fn temporal_property_keys(&self) -> Vec<String>;
     fn temporal_property_values(&self) -> Box<dyn Iterator<Item = Key> + '_> {
         Box::new(
@@ -52,13 +55,78 @@ pub trait TemporalPropertiesOps<Key: 'static = String>: TemporalPropertyViewOps<
     fn get_temporal_property(&self, key: &str) -> Option<Key>;
 }
 
-pub struct TemporalPropertyView<P: TemporalPropertyViewOps<K>, K = String> {
-    pub(crate) id: K,
+pub trait InheritTempralPropertyViewOps: Base {}
+pub trait InheritTemporalPropertiesOps: Base {}
+pub trait InheritStaticPropertiesOps: Base {}
+pub trait InheritPropertiesOps: Base {}
+
+impl<P: InheritPropertiesOps> InheritStaticPropertiesOps for P {}
+impl<P: InheritPropertiesOps> InheritTemporalPropertiesOps for P {}
+
+impl<P: InheritTempralPropertyViewOps> TemporalPropertyViewOps for P
+where
+    P::Base: TemporalPropertyViewOps,
+{
+    fn temporal_value(&self, id: &Key) -> Option<Prop> {
+        self.base().temporal_value(id)
+    }
+
+    fn temporal_history(&self, id: &Key) -> Vec<i64> {
+        self.base().temporal_history(id)
+    }
+
+    fn temporal_values(&self, id: &Key) -> Vec<Prop> {
+        self.base().temporal_values(id)
+    }
+
+    fn temporal_value_at(&self, id: &Key, t: i64) -> Option<Prop> {
+        self.base().temporal_value_at(id, t)
+    }
+}
+
+impl<P: InheritTemporalPropertiesOps> InheritTempralPropertyViewOps for P {}
+
+impl<P: InheritTemporalPropertiesOps> TemporalPropertiesOps for P
+where
+    P::Base: TemporalPropertiesOps,
+{
+    fn temporal_property_keys(&self) -> Vec<String> {
+        self.base().temporal_property_keys()
+    }
+
+    fn temporal_property_values(&self) -> Box<dyn Iterator<Item = Key> + '_> {
+        self.base().temporal_property_values()
+    }
+
+    fn get_temporal_property(&self, key: &str) -> Option<Key> {
+        self.base().get_temporal_property(key)
+    }
+}
+
[…]
