+++ raphtory/src/python/graph/properties.rs
@@ -1,7 +1,8 @@
 use crate::core::Prop;
 use crate::db::api::properties::internal::{
-    StaticProperties, StaticPropertiesOps, TemporalProperties, TemporalPropertiesOps,
-    TemporalPropertyView, TemporalPropertyViewOps,
+    BoxableTemporalProperties, InheritStaticPropertiesOps, InheritTemporalPropertiesOps,
+    InheritTempralPropertyViewOps, StaticProperties, StaticPropertiesOps, TemporalProperties,
+    TemporalPropertiesOps, TemporalPropertyView, TemporalPropertyViewOps,
 };
 use crate::db::api::view::internal::{DynamicGraph, IntoDynamic, Static};
 use crate::db::graph::vertex::VertexView;
@@ -13,68 +14,25 @@ use std::borrow::Borrow;
 use std::ops::Deref;
 use std::sync::Arc;
 
-impl<T: Deref> TemporalPropertyViewOps<String> for T
-where
-    T::Target: TemporalPropertyViewOps<String>,
-{
-    fn temporal_value(&self, id: &String) -> Option<Prop> {
-        self.deref().temporal_value(id)
-    }
+pub type DynTemporalProperties =
+    TemporalProperties<Arc<dyn BoxableTemporalProperties + Send + Sync>>;
 
-    fn temporal_history(&self, id: &String) -> Vec<i64> {
-        self.deref().temporal_history(id)
-    }
+impl InheritTemporalPropertiesOps for Arc<dyn BoxableTemporalProperties + Send + Sync> {}
 
-    fn temporal_values(&self, id: &String) -> Vec<Prop> {
-        self.deref().temporal_values(id)
-    }
+impl InheritTempralPropertyViewOps for Arc<dyn TemporalPropertyViewOps + Send + Sync> {}
 
-    fn temporal_value_at(&self, id: &String, t: i64) -> Option<Prop> {
-        self.deref().temporal_value_at(id, t)
-    }
-}
-
-impl<T: Deref> TemporalPropertiesOps for T
-where
-    T::Target: TemporalPropertiesOps<String>,
-{
-    fn temporal_property_keys(&self) -> Vec<String> {
-        self.deref().temporal_property_keys()
-    }
-
-    fn temporal_property_values(&self) -> Box<dyn Iterator<Item = String> + '_> {
-        self.deref().temporal_property_values()
-    }
-
-    fn get_temporal_property(&self, key: &str) -> Option<String> {
-        self.deref().get_temporal_property(key)
-    }
-}
-
-impl<T: Deref> StaticPropertiesOps for T
-where
-    T::Target: StaticPropertiesOps,
-{
-    fn static_property_keys(&self) -> Vec<String> {
-        self.deref().static_property_keys()
-    }
-
-    fn static_property_values(&self) -> Vec<Prop> {
-        self.deref().static_property_values()
-    }
-
-    fn get_static_property(&self, key: &str) -> Option<Prop> {
-        self.deref().get_static_property(key)
-    }
-}
-
-pub type DynTemporalProperties = TemporalProperties<Arc<dyn TemporalPropertiesOps + Send + Sync>>;
-
-impl<P: TemporalPropertiesOps + Clone + Send + Sync + Static + 'static> From<TemporalProperties<P>>
-    for DynTemporalProperties
+impl<P: BoxableTemporalProperties + Clone + Send + Sync + Static + 'static>
+    From<TemporalProperties<P>> for DynTemporalProperties
 {
     fn from(value: TemporalProperties<P>) -> Self {
         TemporalProperties::new(Arc::new(value.props))
+    }
+}
+
[…]
