### 2023-07-11 3d9aee2
Added a proper health check
