+++ docs/source/api/raphtory.rst
@@ -60,6 +60,17 @@ Graph Generation
    :private-members:
    :inherited-members:
 
+GraphQL Server
+--------------------------
+
+.. automodule:: raphtory.graphql
+   :autosummary:
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+
 
 
 
