### 2023-07-11 0fecbca
Fix errors and warnings after merge
