+++ raphtory-graphql/src/model/filters/nodefilter.rs
@@ -6,7 +6,7 @@ use crate::model::{
     graph::node::Node,
 };
 use dynamic_graphql::InputObject;
-use raphtory::{core::Prop, db::api::view::VertexViewOps};
+use raphtory::db::api::view::VertexViewOps;
 
 #[derive(InputObject)]
 pub struct NodeFilter {
