+++ raphtory/src/search/mod.rs
@@ -6,8 +6,7 @@ use rayon::{prelude::ParallelIterator, slice::ParallelSlice};
 use tantivy::{
     collector::TopDocs,
     schema::{Field, Schema, SchemaBuilder, FAST, INDEXED, STORED, TEXT},
-    Document, Index, IndexReader, IndexSettings, IndexWriter,
-    TantivyError,
+    Document, Index, IndexReader, IndexSettings, IndexWriter, TantivyError,
 };
 
 use crate::{
@@ -22,15 +21,13 @@ use crate::{
     prelude::*,
 };
 
-use self::fields::{EDGE_ID, NAME, TIME, VERTEX_ID, VERTEX_ID_REV};
-
 #[derive(Clone)]
 pub struct IndexedGraph<G> {
     graph: G,
     vertex_index: Arc<Index>,
     edge_index: Arc<Index>,
-    reader: tantivy::IndexReader,
-    edge_reader: tantivy::IndexReader,
+    reader: IndexReader,
+    edge_reader: IndexReader,
 }
 
 impl<G> Deref for IndexedGraph<G> {
@@ -159,18 +156,16 @@ impl<G: GraphViewOps> IndexedGraph<G> {
 
             for prop in prop_names_set.iter() {
                 // load temporal props
-                for (_, prop_value) in vertex.property_history(prop.to_string()) {
+                if let Some(prop_value) = vertex.properties().get(prop).and_then(|p| p.value()) {
                     if found_props.contains(prop) {
                         continue;
                     }
-
                     Self::set_schema_field_from_prop(&mut schema, prop, prop_value);
-
                     found_props.insert(prop.to_string());
                 }
                 // load static props
-                if let Some(prop_value) = vertex.static_property(prop.to_string()) {
-                    let name = if prop == "_id" { NAME } else { prop };
+                if let Some(prop_value) = vertex.static_properties().get(prop) {
+                    let name = if prop == "_id" { fields::NAME } else { prop };
                     if !found_props.contains(name) {
                         Self::set_schema_field_from_prop(&mut schema, name, prop_value);
                         found_props.insert(prop.to_string());
@@ -207,19 +202,18 @@ impl<G: GraphViewOps> IndexedGraph<G> {
 
             for prop in prop_names_set.iter() {
                 // load temporal props
-                for (_, prop_value) in edge.property_history(prop) {
+                if let Some(prop_value) = edge.properties().get(prop).and_then(|p| p.value()) {
                     if found_props.contains(prop) {
                         continue;
                     }
-
                     Self::set_schema_field_from_prop(&mut schema, prop, prop_value);
-
                     found_props.insert(prop.to_string());
                 }
                 // load static props
-                if let Some(prop_value) = edge.static_property(prop) {
-                    if !found_props.contains(prop) {
-                        Self::set_schema_field_from_prop(&mut schema, prop, prop_value);
+                if let Some(prop_value) = edge.static_properties().get(prop) {
+                    let name = if prop == "_id" { fields::NAME } else { prop };
+                    if !found_props.contains(name) {
+                        Self::set_schema_field_from_prop(&mut schema, name, prop_value);
                         found_props.insert(prop.to_string());
                     }
                 }
@@ -299,7 +293,7 @@ impl<G: GraphViewOps> IndexedGraph<G> {
 
             let mut writer_guard = writer_lock.write();
             writer_guard.commit()?;
-            Ok::<(), tantivy::TantivyError>(())
+            Ok::<(), TantivyError>(())
         })?;
 
         reader.reload()?;
[…]
