### 2023-07-11 333b8ff
minor cleanup of properties
