### 2023-07-11 f2155c1
add test to load from polars
