### 2023-07-11 b1e0319
add props for string vertex type too
