+++ raphtory/src/algorithms/algorithm_result.rs
@@ -1,0 +1,200 @@
+use polars::prelude::*;
+use std::collections::HashMap;
+use polars::export::num::ToPrimitive;
+
+pub struct AlgorithmResult<T> {
+    result: HashMap<String, T>,
+    is_categorical: bool,
+}
+
+pub trait AsValueSeries<T> {
+    fn as_series(value_vector: Vec<T>) -> Series;
+}
+
+impl AsValueSeries<u64> for u64 {
+    fn as_series(value_vector: Vec<u64>) -> Series {
+        Series::new("value", value_vector)
+    }
+}
+
+impl AsValueSeries<u32> for u32 {
+    fn as_series(value_vector: Vec<u32>) -> Series {
+        Series::new("value", value_vector)
+    }
+}
+
+impl AsValueSeries<i64> for i64 {
+    fn as_series(value_vector: Vec<i64>) -> Series {
+        Series::new("value", value_vector)
+    }
+}
+
+impl AsValueSeries<i32> for i32 {
+    fn as_series(value_vector: Vec<i32>) -> Series {
+        Series::new("value", value_vector)
+    }
+}
+
+impl AsValueSeries<f64> for f64 {
+    fn as_series(value_vector: Vec<f64>) -> Series {
+        Series::new("value", value_vector)
+    }
+}
+
+impl AsValueSeries<f32> for f32 {
+    fn as_series(value_vector: Vec<f32>) -> Series {
+        Series::new("value", value_vector)
+    }
+}
+
+impl AsValueSeries<bool> for bool {
+    fn as_series(value_vector: Vec<bool>) -> Series {
+        Series::new("value", value_vector)
+    }
+}
+
+impl<T> AlgorithmResult<T> where T: AsValueSeries<T> + Clone + Ord + Eq + std::hash::Hash + Clone {
+    pub fn new(result: HashMap<String, T>, is_categorical: bool) -> Self {
+        AlgorithmResult {
+            result,
+            is_categorical,
+        }
+    }
+
+    pub fn get(&self, key: &str) -> Option<&T> {
+        self.result.get(key)
+    }
+
+    pub fn to_polars_df(&self) -> DataFrame {
+        let keys: Vec<&str> = self.result.keys().map(|s| s.as_str()).collect();
+        let y = self.result.values().cloned().collect();
+        let values: Series = T::as_series(y);
+
+        let series_keys = Series::new("key", keys);
+
+        let columns = vec![series_keys, values];
+        DataFrame::new(columns).unwrap()
+    }
+
+    // to pandas dataframe
+
+
+    pub fn sort(&self, reverse: bool) -> Vec<(String, T)> {
+        let mut sorted: Vec<(String, T)> = self.result.clone().into_iter().collect();
+        sorted.sort_by(|(_, a), (_, b)| {
+            if reverse {
+                b.cmp(a)
+            } else {
+                a.cmp(b)
+            }
+        });
+        sorted
+    }
+
+    pub fn top_k(&self, k: usize, percentage: bool, reverse: bool) -> Option<Vec<(String, T)>> {
+        if !self.is_categorical {
+            if percentage {
+                let total_count = self.result.len();
+                let k = (total_count as f64 * (k as f64 / 100.0)) as usize;
+                let sorted_result = self.sort(reverse);
+                Some(sorted_result.iter().cloned().take(k).collect())
+            } else {
+                let sorted_result = self.sort(reverse);
+                Some(sorted_result.iter().cloned().take(k).collect())
+            }
+        } else {
+            println!("Cannot run top(k) on non-categorical data");
+            None
+        }
+    }
+
+    pub fn group_by(
+        &self
+    ) -> HashMap<T, Vec<String>> {
[…]
