+++ raphtory/Cargo.toml
@@ -44,6 +44,7 @@ lock_api = { version = "0.4", features = ["arc_lock", "serde"] }
 dashmap = {version ="5", features = ["serde"] }
 serde_with = "1.12.0"
 enum_dispatch = "0.3"
+polars = { version = "0.30.0" }
 
 # io optional dependencies
 csv = {version="1.1.6", optional=true}
@@ -62,7 +63,6 @@ tantivy = {version="0.20", optional=true}
 pyo3 =  {version= "0.19.0", features=["multiple-pymethods", "chrono"], optional=true}
 num = {version="0.4.0", optional=true}
 display-error-chain = {version="0.1.1", optional=true}
-polars = {version = "0.30.0", optional=true }
 pyo3-polars = {version = "0.4.1", optional=true }
 
 [dev-dependencies]
@@ -75,8 +75,8 @@ quickcheck_macros = "1"
 [features]
 default = ["search"]
 # Enables the graph loader io module
-io = ["dep:zip","dep:neo4rs", "dep:bzip2", "dep:flate2", "dep:csv", "dep:serde_json", "dep:reqwest", "dep:tokio"]
+io = ["dep:zip", "dep:neo4rs", "dep:bzip2", "dep:flate2", "dep:csv", "dep:serde_json", "dep:reqwest", "dep:tokio"]
 # Enables generating the pyo3 python bindings
-python = ["io", "dep:pyo3", "dep:num", "dep:display-error-chain", "dep:polars", "dep:pyo3-polars"]
+python = ["io", "dep:pyo3", "dep:num", "dep:display-error-chain", "dep:pyo3-polars"]
 # search
 search = ["dep:tantivy"]
