### 2023-07-12 ef39469
change graphql property function to return values
