+++ raphtory/src/algorithms/reciprocity.rs
@@ -111,41 +111,41 @@ pub fn global_reciprocity<G: GraphViewOps>(g: &G, threads: Option<usize>) -> f64
 }
 
 /// returns the reciprocity of every vertex in the graph as a tuple of
-pub fn all_local_reciprocity_ar<G: GraphViewOps>(
-    g: &G,
-    threads: Option<usize>,
-) -> AlgorithmResult<f64> {
-    let mut ctx: Context<G, ComputeStateVec> = g.into();
-
-    let min = sum(0);
-    ctx.agg(min);
-
-    let step1 = ATask::new(move |evv| {
-        let edge_counts = get_reciprocal_edge_count(&evv);
-        let res = (2.0 * edge_counts.2 as f64) / (edge_counts.1 as f64 + edge_counts.0 as f64);
-        if res.is_nan() {
-            evv.global_update(&min, 0.0);
-        } else {
-            evv.update(&min, res);
-        }
-        Step::Continue
-    });
-
-    let mut runner: TaskRunner<G, _> = TaskRunner::new(ctx);
-
-
-    AlgorithmResult::new(
-        runner.run(
-        vec![],
-        vec![Job::new(step1)],
-        (),
-        |_, ess, _, _| ess.finalize(&min, |min| min),
-        threads,
-        1,
-        None,
-        None,
-    ), false)
-}
+// pub fn all_local_reciprocity_ar<G: GraphViewOps>(
+//     g: &G,
+//     threads: Option<usize>,
+// ) -> AlgorithmResult<f64> {
+//     let mut ctx: Context<G, ComputeStateVec> = g.into();
+//
+//     let min = sum(0);
+//     ctx.agg(min);
+//
+//     let step1 = ATask::new(move |evv| {
+//         let edge_counts = get_reciprocal_edge_count(&evv);
+//         let res = (2.0 * edge_counts.2 as f64) / (edge_counts.1 as f64 + edge_counts.0 as f64);
+//         if res.is_nan() {
+//             evv.global_update(&min, 0.0);
+//         } else {
+//             evv.update(&min, res);
+//         }
+//         Step::Continue
+//     });
+//
+//     let mut runner: TaskRunner<G, _> = TaskRunner::new(ctx);
+//
+//
+//     AlgorithmResult::new(
+//         runner.run(
+//         vec![],
+//         vec![Job::new(step1)],
+//         (),
+//         |_, ess, _, _| ess.finalize(&min, |min| min),
+//         threads,
+//         1,
+//         None,
+//         None,
+//     ), false)
+// }
 
 pub fn all_local_reciprocity<G: GraphViewOps>(
     g: &G,
@@ -190,6 +190,7 @@ mod reciprocity_test {
     };
     use pretty_assertions::assert_eq;
     use std::collections::HashMap;
+    // use crate::algorithms::reciprocity::all_local_reciprocity_ar;
 
     #[test]
     fn test_global_recip() {
@@ -223,14 +224,14 @@ mod reciprocity_test {
 
         let map_names_by_id: HashMap<String, f64> = expected_vec.into_iter().collect();
 
-        // let mut hash_map_result: HashMap<String, f64> = HashMap::new();
-        // hash_map_result.insert("1".to_string(), 0.4);
-        // hash_map_result.insert("2".to_string(), 2.0 / 3.0);
-        // hash_map_result.insert("3".to_string(), 0.5);
-        // hash_map_result.insert("4".to_string(), 2.0 / 3.0);
-        // hash_map_result.insert("5".to_string(), 0.0);
-        //
-        // let actual = all_local_reciprocity(&graph, None);
+        let mut hash_map_result: HashMap<String, f64> = HashMap::new();
[…]
