+++ .github/workflows/test_rust_workflow.yml
@@ -53,7 +53,9 @@ jobs:
       - name: Checkout code
         run: |
           mkdir C:\raphtory
-          git checkout ${{ github.sha }} C:\raphtory
+          git clone https://github.com/${{ github.repository }} --branch ${{ github.ref }} C:\raphtory
+          cd C:\raphtory
+          git checkout ${{ github.sha }}
       - name: Set up cargo cache
         uses: actions/cache@v3
         continue-on-error: false
