+++ .github/workflows/test_rust_workflow.yml
@@ -53,7 +53,7 @@ jobs:
       - name: Checkout code
         run: |
           mkdir C:\raphtory
-          git clone https://github.com/${{ github.repository }} --branch ${{ github.ref }} C:\raphtory
+          git clone https://github.com/${{ github.repository }} --branch ${{ github.ref_name }} C:\raphtory
           cd C:\raphtory
           git checkout ${{ github.sha }}
       - name: Set up cargo cache
