+++ .github/workflows/test_rust_workflow.yml
@@ -53,7 +53,7 @@ jobs:
       - run : mkdir C:\raphtory
       - name: Checkout code
         uses: actions/checkout@v3
-      - run : move "D:\a\Raphtory\Raphtory" "C:\raphtory"
+      - run : Xcopy "D:\a\Raphtory\Raphtory" "C:\raphtory" /E/H/C/l
       - name: Set up cargo cache
         uses: actions/cache@v3
         continue-on-error: false
