+++ raphtory/src/algorithms/motifs/three_node_local.rs
@@ -1,3 +1,38 @@
+/// This class regards the counting of the number of three edge, up-to-three node delta-temporal motifs in the graph, using the algorithm of Paranjape et al, Motifs in Temporal Networks (2017).
+/// We point the reader to this reference for more information on the algorithm and background, but provide a short summary below.
+/// 
+///  ## Motifs included
+/// 
+///  ### Stars
+///
+///  There are three classes (in the order they are outputted) of star motif on three nodes based on the switching behaviour of the edges between the two leaf nodes.
+///
+///   - PRE: Stars of the form i<->j, i<->j, i<->k (ie two interactions with leaf j followed by one with leaf k)
+///   - MID: Stars of the form i<->j, i<->k, i<->j (ie switching interactions from leaf j to leaf k, back to j again)
+///   - POST: Stars of the form i<->j, i<->k, i<->k (ie one interaction with leaf j followed by two with leaf k)
+///
+///  Within each of these classes is 8 motifs depending on the direction of the first to the last edge -- incoming "I" or outgoing "O".
+///  These are enumerated in the order III, IIO, IOI, IOO, OII, OIO, OOI, OOO (like binary with "I"-0 and "O"-1).
+///
+///  ### Two node motifs
+///
+///  Also included are two node motifs, of which there are 8 when counted from the perspective of each vertex. These are characterised by the direction of each edge, enumerated
+///  in the above order. Note that for the global graph counts, each motif is counted in both directions (a single III motif for one vertex is an OOO motif for the other vertex).
+///
+///  ### Triangles
+///
+///  There are 8 triangle motifs:
+///
+///   1. i --> j, k --> j, i --> k
+///   2. i --> j, k --> i, j --> k
+///   3. i --> j, j --> k, i --> k
+///   4. i --> j, i --> k, j --> k
+///   5. i --> j, k --> j, k --> i
+///   6. i --> j, k --> i, k --> j
+///   7. i --> j, j --> k, k --> i
+///   8. i --> j, i --> k, k --> j
+///
+
 use crate::{algorithms::motifs::three_node_motifs::*, db::api::view::*};
 use std::collections::HashMap;
 
@@ -201,37 +236,7 @@ fn triangle_motif_count<G: GraphViewOps>(graph: &G, delta: i64) -> HashMap<u64, 
 }
 
 
-/// Computes the number of triangles in a graph using a fast algorithm
-/// 
-///   *  ## Motifs
-///  ### Stars
-///
-///  There are three classes (in the order they are outputted) of star motif on three nodes based on the switching behaviour of the edges between the two leaf nodes.
-///
-///   - PRE: Stars of the form i<->j, i<->j, i<->k (ie two interactions with leaf j followed by one with leaf k)
-///   - MID: Stars of the form i<->j, i<->k, i<->j (ie switching interactions from leaf j to leaf k, back to j again)
-///   - POST: Stars of the form i<->j, i<->k, i<->k (ie one interaction with leaf j followed by two with leaf k)
-///
-///  Within each of these classes is 8 motifs depending on the direction of the first to the last edge -- incoming "I" or outgoing "O".
-///  These are enumerated in the order III, IIO, IOI, IOO, OII, OIO, OOI, OOO (like binary with "I"-0 and "O"-1).
-///
-///  ### Two node motifs
-///
-///  Also included are two node motifs, of which there are 8 when counted from the perspective of each vertex. These are characterised by the direction of each edge, enumerated
-///  in the above order. Note that for the global graph counts, each motif is counted in both directions (a single III motif for one vertex is an OOO motif for the other vertex).
-///
-///  ### Triangles
-///
[…]
