+++ .github/workflows/test_rust_workflow.yml
@@ -53,7 +53,7 @@ jobs:
       - run : mkdir C:\raphtory
       - name: Checkout code
         uses: actions/checkout@v3
-      - run : Xcopy "D:\a\Raphtory\Raphtory" "C:\raphtory" /E/H/C/l
+      - run : Xcopy "D:\a\Raphtory\Raphtory" "C:\raphtory" /E/I/H/Y
       - name: Set up cargo cache
         uses: actions/cache@v3
         continue-on-error: false
@@ -74,7 +74,7 @@ jobs:
           components: rustfmt, clippy
       - name: manual rust test
         run: |
-          cd c:\raphtory\raphtory && C:\Users\runneradmin\.cargo\bin\cargo.exe test --all --no-default-features --features 'raphtory/io raphtory/python'
+          cd c:\raphtory && C:\Users\runneradmin\.cargo\bin\cargo.exe test --all --no-default-features --features 'raphtory/io raphtory/python'
 #      - name: Run Tests
 #        uses: actions-rs/cargo@v1
 #        env:
