### 2023-07-12 aadeca4
start changing the properties so that their is a top-level properties object with temporal and static as sub-parts
