### 2023-07-12 cb338be
remove polars and read edge list from pandas
