+++ raphtory/src/python/graph/graph.rs
@@ -5,14 +5,21 @@
 //! It is a wrapper around a set of shards, which are the actual graph data structures.
 //! In Python, this class wraps around the rust graph.
 use crate::{
-    core::utils::errors::GraphError,
+    core::{entities::vertices::input_vertex::InputVertex, utils::errors::GraphError},
     prelude::*,
     python::{
         graph::views::graph_view::PyGraphView,
         utils::{PyInputVertex, PyTime},
     },
 };
-use pyo3::prelude::*;
+use arrow2::{
+    array::{Array, BooleanArray, PrimitiveArray, Utf8Array},
+    ffi,
+    types::NativeType,
+};
+use pyo3::{
+    create_exception, exceptions::PyException, ffi::Py_uintptr_t, prelude::*, types::PyDict,
+};
 // use pyo3_polars::PyDataFrame;
 use std::{
     collections::HashMap,
@@ -20,6 +27,7 @@ use std::{
     path::{Path, PathBuf},
 };
 
+pub type ArrayRef = Box<dyn Array>;
 /// A temporal graph.
 #[derive(Clone)]
 #[pyclass(name="Graph", extends=PyGraphView)]
@@ -214,7 +222,6 @@ impl PyGraph {
         self.graph.save_to_file(Path::new(path))
     }
 
-
     #[staticmethod]
     #[pyo3(signature = (df, src = "source", dst = "destination", time = "time", props = None))]
     fn load_from_pandas(
@@ -225,97 +232,130 @@ impl PyGraph {
         props: Option<Vec<&str>>,
     ) -> Result<Graph, GraphError> {
         let graph = Graph::new();
+        Python::with_gil(|py| {
+            // let pyarrow = py.import("pyarrow")?;
+            let globals = PyDict::new(py);
+
+            globals.set_item("df", df)?;
+            let locals = PyDict::new(py);
+            py.run(
+                r#"import pyarrow as pa; pa_table = pa.Table.from_pandas(df)"#,
+                Some(globals),
+                Some(locals),
+            )?;
+
+            if let Some(table) = locals.get_item("pa_table") {
+                let rb = table.call_method0("to_batches")?.extract::<Vec<&PyAny>>()?;
+                let names = if let Some(batch0) = rb.get(0) {
+                    let schema = batch0.getattr("schema")?;
+                    schema.getattr("names")?.extract::<Vec<String>>()?
+                } else {
+                    vec![]
+                };
+
+                let arrays = rb
+                    .iter()
+                    .map(|rb| {
+                        (0..names.len())
+                            .map(|i| {
+                                let array = rb.call_method1("column", (i,))?;
+                                let arr = array_to_rust(array)?;
+                                Ok::<Box<dyn Array>, PyErr>(arr)
+                            })
+                            .collect::<Result<Vec<_>, PyErr>>()
+                    })
+                    .collect::<Result<Vec<_>, PyErr>>()?;
+
+                let df = PretendDF { names, arrays };
+                load_from_df(&df, src, dst, time, props, &graph).expect("Failed to load graph!"); // FIXME: handle this correctly
+            }
+
+            Ok::<(), PyErr>(())
+        })
+        .map_err(|e| GraphError::LoadFailure(format!("Failed to load graph {e:?}")))?;
         Ok(graph)
     }
-
-    // #[staticmethod]
-    // #[pyo3(signature = (df, src = "source", dst = "destination", time = "time", props = None))]
-    // fn load_from_polars(
[…]
