+++ raphtory/src/db/api/mutation/addition_ops.rs
@@ -6,10 +6,11 @@ use crate::{
             time::{IntoTimeWithFormat, TryIntoTime},
         },
     },
-    db::api::mutation::internal::InternalAdditionOps, prelude::NO_PROPS,
+    db::api::mutation::internal::InternalAdditionOps,
+    prelude::NO_PROPS,
 };
 
-use super::Properties;
+use super::CollectProperties;
 
 pub trait AdditionOps {
     // TODO: Probably add vector reference here like add
@@ -33,14 +34,14 @@ pub trait AdditionOps {
     /// let v = g.add_vertex(0, "Alice", NO_PROPS);
     /// let v = g.add_vertex(0, 5, NO_PROPS);
     /// ```
-    fn add_vertex<V: InputVertex, T: TryIntoTime, PI: Properties>(
+    fn add_vertex<V: InputVertex, T: TryIntoTime, PI: CollectProperties>(
         &self,
         t: T,
         v: V,
         props: PI,
     ) -> Result<(), GraphError>;
 
-    fn add_vertex_with_custom_time_format<V: InputVertex, PI: Properties>(
+    fn add_vertex_with_custom_time_format<V: InputVertex, PI: CollectProperties>(
         &self,
         t: &str,
         fmt: &str,
@@ -71,7 +72,7 @@ pub trait AdditionOps {
     /// graph.add_vertex(2, "Bob", NO_PROPS).unwrap();
     /// graph.add_edge(3, "Alice", "Bob", NO_PROPS, None).unwrap();
     /// ```    
-    fn add_edge<V: InputVertex, T: TryIntoTime, PI: Properties>(
+    fn add_edge<V: InputVertex, T: TryIntoTime, PI: CollectProperties>(
         &self,
         t: T,
         src: V,
@@ -80,7 +81,7 @@ pub trait AdditionOps {
         layer: Option<&str>,
     ) -> Result<(), GraphError>;
 
-    fn add_edge_with_custom_time_format<V: InputVertex, PI: Properties>(
+    fn add_edge_with_custom_time_format<V: InputVertex, PI: CollectProperties>(
         &self,
         t: &str,
         fmt: &str,
@@ -95,24 +96,18 @@ pub trait AdditionOps {
 }
 
 impl<G: InternalAdditionOps> AdditionOps for G {
-
-    fn add_vertex<V: InputVertex, T: TryIntoTime, PI: Properties>(
+    fn add_vertex<V: InputVertex, T: TryIntoTime, PI: CollectProperties>(
         &self,
         t: T,
         v: V,
         props: PI,
     ) -> Result<(), GraphError> {
         let properties = props.collect_properties();
-        self.internal_add_vertex(
-            t.try_into_time()?,
-            v.id(),
-            v.id_str(),
-            properties,
-        )?;
+        self.internal_add_vertex(t.try_into_time()?, v.id(), v.id_str(), properties)?;
         Ok(())
     }
 
-    fn add_edge<V: InputVertex, T: TryIntoTime, PI: Properties>(
+    fn add_edge<V: InputVertex, T: TryIntoTime, PI: CollectProperties>(
         &self,
         t: T,
         src: V,
