### 2023-07-12 c9bea1e
add vertex load from pandas and 2 ways of loading
