+++ raphtory/src/python/graph/graph.rs
@@ -5,7 +5,7 @@
 //! It is a wrapper around a set of shards, which are the actual graph data structures.
 //! In Python, this class wraps around the rust graph.
 use crate::{
-    core::{entities::vertices::input_vertex::InputVertex, utils::errors::GraphError},
+    core::utils::errors::GraphError,
     prelude::*,
     python::{
         graph::views::graph_view::PyGraphView,
@@ -20,15 +20,13 @@ use arrow2::{
 use pyo3::{
     create_exception, exceptions::PyException, ffi::Py_uintptr_t, prelude::*, types::PyDict,
 };
-// use pyo3_polars::PyDataFrame;
+
 use std::{
-    borrow::Borrow,
     collections::HashMap,
     fmt::{Debug, Formatter},
     path::{Path, PathBuf},
 };
 
-pub type ArrayRef = Box<dyn Array>;
 /// A temporal graph.
 #[derive(Clone)]
 #[pyclass(name="Graph", extends=PyGraphView)]
@@ -224,66 +222,157 @@ impl PyGraph {
     }
 
     #[staticmethod]
-    #[pyo3(signature = (df, src = "source", dst = "destination", time = "time", props = None))]
+    #[pyo3(signature = (edges_df, src = "source", dst = "destination", time = "time", props = None, vertex_df = None, vertex_col = None, vertex_time_col = None, vertex_props = None))]
     fn load_from_pandas(
-        df: &PyAny,
+        edges_df: &PyAny,
         src: &str,
         dst: &str,
         time: &str,
         props: Option<Vec<&str>>,
+        vertex_df: Option<&PyAny>,
+        vertex_col: Option<&str>,
+        vertex_time_col: Option<&str>,
+        vertex_props: Option<Vec<&str>>,
     ) -> Result<Graph, GraphError> {
-        let graph = Graph::new();
-        Python::with_gil(|py| {
-            // let pyarrow = py.import("pyarrow")?;
-            let globals = PyDict::new(py);
-
-            globals.set_item("df", df)?;
-            let locals = PyDict::new(py);
-            py.run(
-                r#"import pyarrow as pa; pa_table = pa.Table.from_pandas(df)"#,
-                Some(globals),
-                Some(locals),
+        let graph = PyGraph{graph: Graph::new()};
+        graph.load_edges_from_pandas(edges_df, src, dst, time, props)?;
+        if let (Some(vertex_df), Some(vertex_col), Some(vertex_time_col)) =
+            (vertex_df, vertex_col, vertex_time_col)
+        {
+            graph.load_vertices_from_pandas(
+                vertex_df,
+                vertex_col,
+                vertex_time_col,
+                vertex_props,
             )?;
+        }
+        Ok(graph.graph)
+    }
 
-            if let Some(table) = locals.get_item("pa_table") {
-                let rb = table.call_method0("to_batches")?.extract::<Vec<&PyAny>>()?;
-                let names = if let Some(batch0) = rb.get(0) {
-                    let schema = batch0.getattr("schema")?;
-                    schema.getattr("names")?.extract::<Vec<String>>()?
-                } else {
-                    vec![]
-                };
 
-                let arrays = rb
-                    .iter()
-                    .map(|rb| {
-                        (0..names.len())
-                            .map(|i| {
-                                let array = rb.call_method1("column", (i,))?;
-                                let arr = array_to_rust(array)?;
[…]
