+++ raphtory/src/algorithms/algorithm_result.rs
@@ -1,84 +1,23 @@
-use polars::prelude::*;
 use std::collections::HashMap;
+use ordered_float::OrderedFloat;
 use std::hash::Hash;
-use polars::export::num::ToPrimitive;
 
-pub struct AlgorithmResult<T> {
+struct AlgorithmResult<T> where T: Clone   {
     result: HashMap<String, T>,
-    is_categorical: bool,
 }
 
-pub trait AsValueSeries<T> {
-    fn as_series(value_vector: Vec<T>) -> Series;
-}
-
-impl AsValueSeries<u64> for u64 {
-    fn as_series(value_vector: Vec<u64>) -> Series {
-        Series::new("value", value_vector)
-    }
-}
-
-impl AsValueSeries<u32> for u32 {
-    fn as_series(value_vector: Vec<u32>) -> Series {
-        Series::new("value", value_vector)
-    }
-}
-
-impl AsValueSeries<i64> for i64 {
-    fn as_series(value_vector: Vec<i64>) -> Series {
-        Series::new("value", value_vector)
-    }
-}
-
-impl AsValueSeries<i32> for i32 {
-    fn as_series(value_vector: Vec<i32>) -> Series {
-        Series::new("value", value_vector)
-    }
-}
-
-impl AsValueSeries<f64> for f64 {
-    fn as_series(value_vector: Vec<f64>) -> Series {
-        Series::new("value", value_vector)
-    }
-}
-
-impl AsValueSeries<f32> for f32 {
-    fn as_series(value_vector: Vec<f32>) -> Series {
-        Series::new("value", value_vector)
-    }
-}
-
-impl AsValueSeries<bool> for bool {
-    fn as_series(value_vector: Vec<bool>) -> Series {
-        Series::new("value", value_vector)
-    }
-}
-
-impl<T> AlgorithmResult<T> where T: AsValueSeries<T> + Clone + PartialOrd + PartialEq + Hash + Clone {
-    pub fn new(result: HashMap<String, T>, is_categorical: bool) -> Self {
-        AlgorithmResult {
+impl<T> AlgorithmResult<T>
+    where
+        T: Clone + PartialOrd  {
+    pub fn new(result: HashMap<String, T>) -> Self {
+        Self {
             result,
-            is_categorical,
         }
     }
 
     pub fn get(&self, key: &str) -> Option<&T> {
         self.result.get(key)
     }
-
-    pub fn to_polars_df(&self) -> DataFrame {
-        let keys: Vec<&str> = self.result.keys().map(|s| s.as_str()).collect();
-        let y = self.result.values().cloned().collect();
-        let values: Series = T::as_series(y);
-
-        let series_keys = Series::new("key", keys);
-
-        let columns = vec![series_keys, values];
-        DataFrame::new(columns).unwrap()
-    }
-
-    // to pandas dataframe
-
 
     pub fn sort(&self, reverse: bool) -> Vec<(String, T)> {
         let mut sorted: Vec<(String, T)> = self.result.clone().into_iter().collect();
@@ -93,27 +32,36 @@ impl<T> AlgorithmResult<T> where T: AsValueSeries<T> + Clone + PartialOrd + Part
     }
 
     pub fn top_k(&self, k: usize, percentage: bool, reverse: bool) -> Option<Vec<(String, T)>> {
-        if !self.is_categorical {
-            if percentage {
-                let total_count = self.result.len();
-                let k = (total_count as f64 * (k as f64 / 100.0)) as usize;
-                let sorted_result = self.sort(reverse);
-                Some(sorted_result.iter().cloned().take(k).collect())
-            } else {
-                let sorted_result = self.sort(reverse);
-                Some(sorted_result.iter().cloned().take(k).collect())
-            }
+        if percentage {
[…]
