### 2023-07-12 23e2c17
at this point only god knows how it works. added tests for all types we currently support in our algorithms
