### 2023-07-13 15d866b
added tests for all loading functions
