### 2023-07-13 10a8d98
3nm docs (#1108)
* privatise some functions and initial docstring for three node motifs

* adding rust docs for temporal motifs

* added python docs
