### 2023-07-13 19aeffd
add Rust Benchmark (cargo) benchmark result for 592e22535b3e4dc8b51eb20b1235ece7ede33b09
