+++ raphtory/src/python/graph/mod.rs
@@ -2,4 +2,5 @@ pub mod edge;
 pub mod graph;
 pub mod graph_with_deletions;
 pub mod vertex;
+pub mod pandas;
 pub mod views;
