+++ raphtory/src/python/graph/pandas.rs
@@ -1,0 +1,323 @@
+use arrow2::{array::{Array, PrimitiveArray, Utf8Array, BooleanArray}, types::NativeType, ffi};
+use pyo3::{
+    create_exception, exceptions::PyException, ffi::Py_uintptr_t, prelude::*, types::PyDict,
+};
+
+use crate::{prelude::*, core::utils::errors::GraphError};
+
+fn i64_opt_into_u64_opt(x: Option<&i64>) -> Option<u64> {
+    x.map(|x| (*x).try_into().unwrap())
+}
+
+pub(crate) fn process_pandas_py_df<'a>(df: &PyAny, py: Python<'a>) -> PyResult<PretendDF> {
+    let globals = PyDict::new(py);
+
+    globals.set_item("df", df)?;
+    let locals = PyDict::new(py);
+    py.run(
+        r#"import pyarrow as pa; pa_table = pa.Table.from_pandas(df)"#,
+        Some(globals),
+        Some(locals),
+    )?;
+
+    if let Some(table) = locals.get_item("pa_table") {
+        let rb = table.call_method0("to_batches")?.extract::<Vec<&PyAny>>()?;
+        let names = if let Some(batch0) = rb.get(0) {
+            let schema = batch0.getattr("schema")?;
+            schema.getattr("names")?.extract::<Vec<String>>()?
+        } else {
+            vec![]
+        };
+
+        let arrays = rb
+            .iter()
+            .map(|rb| {
+                (0..names.len())
+                    .map(|i| {
+                        let array = rb.call_method1("column", (i,))?;
+                        let arr = array_to_rust(array)?;
+                        Ok::<Box<dyn Array>, PyErr>(arr)
+                    })
+                    .collect::<Result<Vec<_>, PyErr>>()
+            })
+            .collect::<Result<Vec<_>, PyErr>>()?;
+
+        let df = PretendDF { names, arrays };
+        Ok(df)
+    } else {
+        return Err(GraphLoadException::new_err(
+            "Failed to load graph, could not convert pandas dataframe to arrow table".to_string(),
+        ));
+    }
+}
+
+pub(crate) fn load_vertices_from_df<'a>(
+    df: &'a PretendDF,
+    vertex_id: &str,
+    time: &str,
+    props: Option<Vec<&str>>,
+    graph: &Graph,
+) -> Result<(), GraphError> {
+    let prop_iter = props
+        .unwrap_or_default()
+        .into_iter()
+        .map(|name| lift_property(name, &df))
+        .reduce(combine_prop_iters)
+        .unwrap_or_else(|| Box::new(std::iter::repeat(vec![])));
+
+    if let (Some(vertex_id), Some(time)) = (df.iter_col::<u64>(vertex_id), df.iter_col::<i64>(time))
+    {
+        let iter = vertex_id.map(|i| i.copied()).zip(time);
+        load_vertices_from_num_iter(graph, iter, prop_iter)?;
+    } else if let (Some(vertex_id), Some(time)) =
+        (df.iter_col::<i64>(vertex_id), df.iter_col::<i64>(time))
+    {
+        let iter = vertex_id.map(i64_opt_into_u64_opt).zip(time);
+        load_vertices_from_num_iter(graph, iter, prop_iter)?;
+    } else if let (Some(vertex_id), Some(time)) = (df.utf8(vertex_id), df.iter_col::<i64>(time)) {
+        let iter = vertex_id.into_iter().zip(time);
+        for ((vertex_id, time), props) in iter.zip(prop_iter) {
+            if let (Some(vertex_id), Some(time)) = (vertex_id, time) {
+                graph.add_vertex(*time, vertex_id, props)?;
+            }
+        }
[…]
