### 2023-07-13 da65bb9
add rust test for loading from arrow arrays
