+++ raphtory/src/python/graph/pandas.rs
@@ -1,9 +1,13 @@
-use arrow2::{array::{Array, PrimitiveArray, Utf8Array, BooleanArray}, types::NativeType, ffi};
+use arrow2::{
+    array::{Array, BooleanArray, PrimitiveArray, Utf8Array},
+    ffi,
+    types::{NativeType, Offset},
+};
 use pyo3::{
     create_exception, exceptions::PyException, ffi::Py_uintptr_t, prelude::*, types::PyDict,
 };
 
-use crate::{prelude::*, core::utils::errors::GraphError};
+use crate::{core::utils::errors::GraphError, prelude::*};
 
 fn i64_opt_into_u64_opt(x: Option<&i64>) -> Option<u64> {
     x.map(|x| (*x).try_into().unwrap())
@@ -74,7 +78,18 @@ pub(crate) fn load_vertices_from_df<'a>(
     {
         let iter = vertex_id.map(i64_opt_into_u64_opt).zip(time);
         load_vertices_from_num_iter(graph, iter, prop_iter)?;
-    } else if let (Some(vertex_id), Some(time)) = (df.utf8(vertex_id), df.iter_col::<i64>(time)) {
+    } else if let (Some(vertex_id), Some(time)) =
+        (df.utf8::<i32>(vertex_id), df.iter_col::<i64>(time))
+    {
+        let iter = vertex_id.into_iter().zip(time);
+        for ((vertex_id, time), props) in iter.zip(prop_iter) {
+            if let (Some(vertex_id), Some(time)) = (vertex_id, time) {
+                graph.add_vertex(*time, vertex_id, props)?;
+            }
+        }
+    } else if let (Some(vertex_id), Some(time)) =
+        (df.utf8::<i64>(vertex_id), df.iter_col::<i64>(time))
+    {
         let iter = vertex_id.into_iter().zip(time);
         for ((vertex_id, time), props) in iter.zip(prop_iter) {
             if let (Some(vertex_id), Some(time)) = (vertex_id, time) {
@@ -125,9 +140,22 @@ pub(crate) fn load_edges_from_df<'a>(
             .zip(dst.map(i64_opt_into_u64_opt))
             .zip(time);
         load_edges_from_num_iter(&graph, triplets, prop_iter)?;
-    } else if let (Some(src), Some(dst), Some(time)) =
-        (df.utf8(src), df.utf8(dst), df.iter_col::<i64>(time))
-    {
+    } else if let (Some(src), Some(dst), Some(time)) = (
+        df.utf8::<i32>(src),
+        df.utf8::<i32>(dst),
+        df.iter_col::<i64>(time),
+    ) {
+        let triplets = src.into_iter().zip(dst.into_iter()).zip(time.into_iter());
+        for (((src, dst), time), props) in triplets.zip(prop_iter) {
+            if let (Some(src), Some(dst), Some(time)) = (src, dst, time) {
+                graph.add_edge(*time, src, dst, props, None)?;
+            }
+        }
+    } else if let (Some(src), Some(dst), Some(time)) = (
+        df.utf8::<i64>(src),
+        df.utf8::<i64>(dst),
+        df.iter_col::<i64>(time),
+    ) {
         let triplets = src.into_iter().zip(dst.into_iter()).zip(time.into_iter());
         for (((src, dst), time), props) in triplets.zip(prop_iter) {
             if let (Some(src), Some(dst), Some(time)) = (src, dst, time) {
@@ -148,24 +176,30 @@ fn lift_property<'a: 'b, 'b>(
     df: &'b PretendDF,
 ) -> Box<dyn Iterator<Item = Vec<(&'b str, Prop)>> + 'b> {
     if let Some(col) = df.iter_col::<f64>(name) {
[…]
