+++ raphtory/src/python/graph/pandas.rs
@@ -374,44 +374,33 @@ create_exception!(exceptions, GraphLoadException, PyException);
 
 #[cfg(test)]
 mod test {
-    use crate::prelude::*;
+    use crate::{prelude::*, python::graph::pandas::load_vertices_from_df};
 
     use super::{load_edges_from_df, PretendDF};
     use arrow2::array::{PrimitiveArray, Utf8Array};
 
     #[test]
-    fn load_from_pretend_df() {
+    fn load_edges_from_pretend_df() {
         let df = PretendDF {
             names: vec!["src", "dst", "time", "prop1", "prop2"]
                 .iter()
                 .map(|s| s.to_string())
                 .collect(),
             arrays: vec![
-                vec![Box::new(PrimitiveArray::<u64>::from(vec![
-                    Some(1),
-                    Some(2),
-                    Some(3),
-                ]))],
-                vec![Box::new(PrimitiveArray::<u64>::from(vec![
-                    Some(2),
-                    Some(3),
-                    Some(4),
-                ]))],
-                vec![Box::new(PrimitiveArray::<i64>::from(vec![
-                    Some(1),
-                    Some(2),
-                    Some(3),
-                ]))],
-                vec![Box::new(PrimitiveArray::<f64>::from(vec![
-                    Some(1.0),
-                    Some(2.0),
-                    Some(3.0),
-                ]))],
-                vec![Box::new(Utf8Array::<i32>::from(vec![
-                    Some("a"),
-                    Some("b"),
-                    Some("c"),
-                ]))],
+                vec![
+                    Box::new(PrimitiveArray::<u64>::from(vec![Some(1)])),
+                    Box::new(PrimitiveArray::<u64>::from(vec![Some(2)])),
+                    Box::new(PrimitiveArray::<i64>::from(vec![Some(1)])),
+                    Box::new(PrimitiveArray::<f64>::from(vec![Some(1.0)])),
+                    Box::new(Utf8Array::<i32>::from(vec![Some("a")])),
+                ],
+                vec![
+                    Box::new(PrimitiveArray::<u64>::from(vec![Some(2), Some(3)])),
+                    Box::new(PrimitiveArray::<u64>::from(vec![Some(3), Some(4)])),
+                    Box::new(PrimitiveArray::<i64>::from(vec![Some(2), Some(3)])),
+                    Box::new(PrimitiveArray::<f64>::from(vec![Some(2.0), Some(3.0)])),
+                    Box::new(Utf8Array::<i32>::from(vec![Some("b"), Some("c")])),
+                ],
             ],
         };
         let graph = Graph::new();
@@ -426,15 +415,18 @@ mod test {
         )
         .expect("failed to load edges from pretend df");
 
-        let actual = graph.edges().map(|e| {
-            (
-                e.src().id(),
-                e.dst().id(),
-                e.latest_time(),
-                e.property("prop1", false),
-                e.property("prop2", false),
-            )
-        }).collect::<Vec<_>>();
+        let actual = graph
+            .edges()
+            .map(|e| {
+                (
+                    e.src().id(),
+                    e.dst().id(),
+                    e.latest_time(),
+                    e.property("prop1", false),
+                    e.property("prop2", false),
+                )
+            })
+            .collect::<Vec<_>>();
 
         assert_eq!(
             actual,
@@ -442,6 +434,52 @@ mod test {
                 (1, 2, Some(1), Some(Prop::F64(1.0)), Some(Prop::str("a"))),
                 (2, 3, Some(2), Some(Prop::F64(2.0)), Some(Prop::str("b"))),
                 (3, 4, Some(3), Some(Prop::F64(3.0)), Some(Prop::str("c"))),
-            ]);
+            ]
+        );
[…]
