### 2023-07-13 be2159c
core compiles again after refactoring of static/temporal properties, still need to update python/tests
