### 2023-07-13 0227ea6
Property returns value (#1106)
change graphql property function to return values
