### 2023-07-13 79a9914
add Rust Benchmark (cargo) benchmark result for 0227ea6b7a465df2731bca6ffb1f9ab2c7ab07b9
