+++ raphtory/src/db/graph/graph.rs
@@ -491,44 +491,38 @@ mod db_tests {
         g.add_vertex(0, 11, vec![("temp".to_string(), Prop::Bool(true))])
             .unwrap();
         let v11 = g.vertex(11).unwrap();
-        let v22 = g.vertex_ref(22).unwrap();
-        let v33 = g.vertex_ref(33).unwrap();
-        let edge1111 = g.edge_ref(v11.into(), v11.into(), 0).unwrap();
-        let edge2233 = g.edge_ref(v22.into(), v33.into(), 0).unwrap();
-        let edge3311 = g.edge_ref(v33.into(), v11.into(), 0).unwrap();
+        let v22 = g.vertex(22).unwrap();
+        let v33 = g.vertex(33).unwrap();
+        let edge1111 = g.edge(&v11, &v11, None).unwrap();
+        let edge2233 = g.edge(&v22, &v33, None).unwrap();
+        let edge3311 = g.edge(&v33, &v11, None).unwrap();
 
-        g.add_vertex_properties(
-            11,
-            vec![
-                ("a".to_string(), Prop::U64(11)),
-                ("b".to_string(), Prop::I64(11)),
-            ],
-        )
-        .unwrap();
-        g.add_vertex_properties(11, vec![("c".to_string(), Prop::U32(11))])
+        g.add_vertex_properties(11, vec![("a", Prop::U64(11)), ("b", Prop::I64(11))])
             .unwrap();
-        g.add_vertex_properties(22, vec![("b".to_string(), Prop::U64(22))])
+        g.add_vertex_properties(11, vec![("c", Prop::U32(11))])
             .unwrap();
-        g.add_edge_properties(11, 11, vec![("d".to_string(), Prop::U64(1111))], None)
+        g.add_vertex_properties(22, vec![("b", Prop::U64(22))])
             .unwrap();
-        g.add_edge_properties(33, 11, vec![("a".to_string(), Prop::U64(3311))], None)
+        g.add_edge_properties(11, 11, vec![("d", Prop::U64(1111))], None)
+            .unwrap();
+        g.add_edge_properties(33, 11, vec![("a", Prop::U64(3311))], None)
             .unwrap();
 
         assert_eq!(v11.properties().meta().keys(), vec!["a", "b", "c"]);
-        assert_eq!(g.static_vertex_prop_names(v22), vec!["b"]);
-        assert!(g.static_vertex_prop_names(v33).is_empty());
-        assert_eq!(g.static_edge_prop_names(edge1111), vec!["d"]);
-        assert_eq!(g.static_edge_prop_names(edge3311), vec!["a"]);
-        assert!(g.static_edge_prop_names(edge2233).is_empty());
+        assert_eq!(v22.properties().meta().keys(), vec!["b"]);
+        assert!(v33.properties().meta().keys().is_empty());
+        assert_eq!(edge1111.properties().meta().keys(), vec!["d"]);
+        assert_eq!(edge3311.properties().meta().keys(), vec!["a"]);
+        assert!(edge2233.properties().meta().keys().is_empty());
 
-        assert_eq!(g.static_vertex_prop(v11, "a"), Some(Prop::U64(11)));
-        assert_eq!(g.static_vertex_prop(v11, "b"), Some(Prop::I64(11)));
-        assert_eq!(g.static_vertex_prop(v11, "c"), Some(Prop::U32(11)));
-        assert_eq!(g.static_vertex_prop(v22, "b"), Some(Prop::U64(22)));
-        assert_eq!(g.static_vertex_prop(v22, "a"), None);
-        assert_eq!(g.static_edge_prop(edge1111, "d"), Some(Prop::U64(1111)));
[…]
