+++ raphtory/src/python/graph/properties.rs
@@ -4,7 +4,7 @@ use crate::db::api::properties::internal::{
     InheritPropertiesOps, InheritStaticPropertiesOps, InheritTemporalPropertiesOps,
     InheritTempralPropertyViewOps, PropertiesOps, StaticPropertiesOps, TemporalPropertyViewOps,
 };
-use crate::db::api::properties::StaticProperties;
+use crate::db::api::properties::{Properties, StaticProperties};
 use crate::db::api::properties::{TemporalProperties, TemporalPropertyView};
 use crate::db::api::view::internal::{DynamicGraph, Static};
 use crate::python::types::repr::{iterator_dict_repr, iterator_repr, Repr};
@@ -16,6 +16,7 @@ use std::sync::Arc;
 
 pub type DynTemporalProperties = TemporalProperties<Arc<dyn PropertiesOps + Send + Sync>>;
 pub type DynTemporalProperty = TemporalPropertyView<Arc<dyn PropertiesOps + Send + Sync>>;
+pub type DynProperties = Properties<Arc<dyn PropertiesOps + Send + Sync>>;
 
 impl InheritPropertiesOps for Arc<dyn PropertiesOps + Send + Sync> {}
 
@@ -42,7 +43,12 @@ impl<P: Into<DynTemporalProperties>> From<P> for PyTemporalProperties {
     }
 }
 
-#[pyclass(name = "Properties")]
+#[pyclass(name=Properties)]
+pub struct PyProperties {
+    props: DynProperties,
+}
+
+#[pyclass(name = "TemporalProperties")]
 pub struct PyTemporalProperties {
     props: DynTemporalProperties,
 }
