### 2023-07-13 cec0e21
all algos done except three node motifs
