+++ raphtory/src/algorithms/pagerank.rs
@@ -14,6 +14,8 @@ use crate::{
 };
 use num_traits::abs;
 use std::collections::HashMap;
+use ordered_float::OrderedFloat;
+use crate::algorithms::algorithm_result::AlgorithmResult;
 
 #[derive(Clone, Debug)]
 struct PageRankState {
@@ -41,7 +43,7 @@ pub fn unweighted_page_rank<G: GraphViewOps>(
     threads: Option<usize>,
     tol: Option<f64>,
     use_l2_norm: bool,
-) -> HashMap<String, f64> {
+) -> AlgorithmResult<OrderedFloat<f64>> {
     let n = g.num_vertices();
     let total_edges = g.num_edges();
 
@@ -158,9 +160,11 @@ pub fn unweighted_page_rank<G: GraphViewOps>(
         None,
     );
 
-    out.into_iter()
+    let res = out.into_iter()
         .map(|(k, v)| (g.vertex_name(k), v))
-        .collect()
+        .collect();
+
+    AlgorithmResult::new_with_float(res)
 }
 
 #[cfg(test)]
@@ -189,14 +193,12 @@ mod page_rank_tests {
     fn test_page_rank() {
         let graph = load_graph();
 
-        let results: HashMap<String, f64> = unweighted_page_rank(&graph, 1000, Some(1), None, true)
-            .into_iter()
-            .collect();
+        let results: AlgorithmResult<OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(1), None, true);
 
-        assert_eq_f64(results.get("1"), Some(&0.38694), 5);
-        assert_eq_f64(results.get("2"), Some(&0.20195), 5);
-        assert_eq_f64(results.get("4"), Some(&0.20195), 5);
-        assert_eq_f64(results.get("3"), Some(&0.20916), 5);
+        assert_eq_f64(Some(&results.get("1").unwrap().0), Some(&0.38694), 5);
+        assert_eq_f64(Some(&results.get("2").unwrap().0), Some(&0.20195), 5);
+        assert_eq_f64(Some(&results.get("4").unwrap().0), Some(&0.20195), 5);
+        assert_eq_f64(Some(&results.get("3").unwrap().0), Some(&0.20916), 5);
     }
 
     #[test]
@@ -233,21 +235,19 @@ mod page_rank_tests {
             graph.add_edge(t, src, dst, [], None).unwrap();
         }
 
-        let results: HashMap<String, f64> = unweighted_page_rank(&graph, 1000, Some(4), None, true)
-            .into_iter()
-            .collect();
+        let results: AlgorithmResult<OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(4), None, true);
 
-        assert_eq_f64(results.get("10"), Some(&0.072082), 5);
-        assert_eq_f64(results.get("8"), Some(&0.136473), 5);
-        assert_eq_f64(results.get("3"), Some(&0.15484), 5);
-        assert_eq_f64(results.get("6"), Some(&0.07208), 5);
-        assert_eq_f64(results.get("11"), Some(&0.06186), 5);
-        assert_eq_f64(results.get("2"), Some(&0.03557), 5);
-        assert_eq_f64(results.get("1"), Some(&0.11284), 5);
-        assert_eq_f64(results.get("4"), Some(&0.07944), 5);
-        assert_eq_f64(results.get("7"), Some(&0.01638), 5);
-        assert_eq_f64(results.get("9"), Some(&0.06186), 5);
-        assert_eq_f64(results.get("5"), Some(&0.19658), 5);
+        assert_eq_f64(Some(&results.get("10").unwrap().0), Some(&0.072082), 5);
[…]
