### 2023-07-13 11fa639
update the python loader to include polars dataframe (#1104)
* update the python loader to include polars dataframe

* add test to load from polars

* load property works

* add props for string vertex type too

* remove target from cache

* temporary remove comparison benchmark

* remove polars

* remove polars and read edge list from pandas

* fix CI build

* add vertex load from pandas and 2 ways of loading

* added tests for all loading functions

* pulled the pandas loaders int pandas.rs

* add rust test for loading from arrow arrays

* improve testing
