+++ raphtory/src/python/graph/edge.rs
@@ -404,7 +404,7 @@ impl Repr for PyEdge {
         let properties = &self
             .properties(Some(true))
             .iter()
-            .map(|(k, v)| k.to_string() + " : " + &v.to_string())
+            .map(|(k, v)|  format!("{k} : {:?}", v.to_string()))
             .join(", ");
 
         let source = self.edge.src().name();
@@ -420,7 +420,7 @@ impl Repr for PyEdge {
                 latest_time.unwrap_or(0),
             )
         } else {
-            let property_string: String = "{".to_string() + properties + "}";
+            let property_string: String = format!("{{{properties}}}");
             format!(
                 "Edge(source={}, target={}, earliest_time={}, latest_time={}, properties={})",
                 source.trim_matches('"'),
