+++ raphtory/src/graph_loader/source/mod.rs
@@ -1,4 +1,3 @@
 pub mod csv_loader;
 pub mod json_loader;
 pub mod neo4j_loader;
-pub mod polars_loader;
