+++ raphtory/src/graph_loader/example/company_house.rs
@@ -61,7 +61,7 @@ pub fn company_house_graph(path: Option<String>) -> Graph {
             .set_delimiter(",")
             .load_into_graph(&g, |company_house: CompanyHouse, g: &Graph| {
                 let pincode = &company_house.pincode;
-                let address = company_house.address + ", " + pincode;
+                let address = format!("{}, {pincode}", company_house.address);
                 let company = company_house.company;
                 let owner = company_house.owner;
                 // let illegal_flag : Option<String> = match company_house.illegal_hmo {
