+++ python/pyproject.toml
@@ -14,6 +14,8 @@ dependencies = [
     "pandas >= 1.3.3",
     "pyvis >= 0.3.2",
     "networkx >= 2.6.3",
+    "pandas >= 2.0.3",
+    "pyarrow >= 12.0.1"
 ]
 
 

[…]
