### 2023-07-13 3f3900f
add Rust Benchmark (cargo) benchmark result for 45623dd628975cecbc62be537e133876a6ab31e1
