### 2023-07-13 6ed0cc6
implement debug formatter for algorithm result object to pass last failed test
