### 2023-07-13 8fae8f0
😈Generic type the input as well, why not😈
