+++ raphtory/src/algorithms/algorithm_result.rs
@@ -4,25 +4,31 @@ use std::fmt::{Debug, Display};
 use ordered_float::OrderedFloat;
 use std::hash::Hash;
 
-pub struct AlgorithmResult<T> where T: Clone  {
-    result: HashMap<String, T>,
+pub struct AlgorithmResult<H, Y>
+    where
+        H: Clone + Hash + Eq + Ord,
+        Y: Clone
+{
+    result: HashMap<H, Y>,
 }
 
-impl<T> AlgorithmResult<T>
+impl<H, Y> AlgorithmResult<H, Y>
     where
-        T: Clone {
-    pub fn new(result: HashMap<String, T>) -> Self {
+        H: Clone + Hash + Eq + Ord,
+        Y: Clone,
+{
+    pub fn new(result: HashMap<H, Y>) -> Self {
         Self {
             result,
         }
     }
 
-    pub fn get_all(&self) -> &HashMap<String, T> {
+    pub fn get_all(&self) -> &HashMap<H, Y> {
         &self.result
     }
 
-    pub fn get(&self, key: &str) -> Option<&T> {
-        self.result.get(key)
+    pub fn get(&self, key: &H) -> Option<&Y> {
+        self.result.get(&key)
     }
 
     // TODO implement to pandas dataframe
@@ -31,11 +37,13 @@ impl<T> AlgorithmResult<T>
 }
 
 
-impl<T> AlgorithmResult<T>
+impl<H, Y> AlgorithmResult<H, Y>
     where
-        T: Clone + PartialOrd {
-    pub fn sort_by_value(&self, reverse: bool) -> Vec<(String, T)> {
-        let mut sorted: Vec<(String, T)> = self.result.clone().into_iter().collect();
+        H: Clone + Hash + Eq + Ord,
+        Y: Clone + PartialOrd
+{
+    pub fn sort_by_value(&self, reverse: bool) -> Vec<(H, Y)> {
+        let mut sorted: Vec<(H, Y)> = self.result.clone().into_iter().collect();
         sorted.sort_by(|(_, a), (_, b)| {
             if reverse {
                 b.partial_cmp(a).unwrap()
@@ -46,20 +54,20 @@ impl<T> AlgorithmResult<T>
         sorted
     }
 
-    pub fn sort_by_key(&self, reverse: bool) -> Vec<(String, T)> {
-        let mut sorted: Vec<(String, T)> = self.result.clone().into_iter().collect();
+    pub fn sort_by_key(&self, reverse: bool) -> Vec<(H, Y)> {
+        let mut sorted: Vec<(H, Y)> = self.result.clone().into_iter().collect();
         sorted.sort_by(|(a, _), (b, _)| {
             if reverse {
-                b.partial_cmp(a).unwrap()
+                b.cmp(a)
             } else {
-                a.partial_cmp(b).unwrap()
+                a.cmp(b)
             }
         });
         sorted
     }
 
 
-    pub fn top_k(&self, k: usize, percentage: bool, reverse: bool) -> Option<Vec<(String, T)>> {
+    pub fn top_k(&self, k: usize, percentage: bool, reverse: bool) -> Option<Vec<(H, Y)>> {
         if percentage {
             let total_count = self.result.len();
             let k = (total_count as f64 * (k as f64 / 100.0)) as usize;
@@ -73,9 +81,9 @@ impl<T> AlgorithmResult<T>
 
 }
 
-impl AlgorithmResult<f64> {
-    pub fn new_with_float(hashmap: HashMap<String, f64>) -> AlgorithmResult<OrderedFloat<f64>> {
-        let converted_hashmap: HashMap<String, OrderedFloat<f64>> = hashmap
+impl<H: Clone + Hash + Eq + Ord> AlgorithmResult<H, f64> {
[…]
