+++ raphtory/src/algorithms/hits.rs
@@ -37,7 +37,7 @@ pub fn hits<G: GraphViewOps>(
     window: Range<i64>,
     iter_count: usize,
     threads: Option<usize>,
-) -> AlgorithmResult<(f32, f32)> {
+) -> AlgorithmResult<String, (f32, f32)> {
     let mut ctx: Context<G, ComputeStateVec> = g.into();
 
     let recv_hub_score = sum::<f32>(2);
@@ -195,7 +195,7 @@ mod hits_tests {
 
         let window = 0..10;
 
-        let results: AlgorithmResult<(f32, f32)> =
+        let results: AlgorithmResult<String, (f32, f32)> =
             hits(&graph, window, 20, None);
 
         // NetworkX results
