+++ raphtory/src/algorithms/pagerank.rs
@@ -43,7 +43,7 @@ pub fn unweighted_page_rank<G: GraphViewOps>(
     threads: Option<usize>,
     tol: Option<f64>,
     use_l2_norm: bool,
-) -> AlgorithmResult<OrderedFloat<f64>> {
+) -> AlgorithmResult<String, OrderedFloat<f64>> {
     let n = g.num_vertices();
     let total_edges = g.num_edges();
 
@@ -193,12 +193,12 @@ mod page_rank_tests {
     fn test_page_rank() {
         let graph = load_graph();
 
-        let results: AlgorithmResult<OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(1), None, true);
+        let results: AlgorithmResult<String, OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(1), None, true);
 
-        assert_eq_f64(Some(&results.get("1").unwrap().0), Some(&0.38694), 5);
-        assert_eq_f64(Some(&results.get("2").unwrap().0), Some(&0.20195), 5);
-        assert_eq_f64(Some(&results.get("4").unwrap().0), Some(&0.20195), 5);
-        assert_eq_f64(Some(&results.get("3").unwrap().0), Some(&0.20916), 5);
+        assert_eq_f64(Some(&results.get(&"1".to_string()).unwrap().0), Some(&0.38694), 5);
+        assert_eq_f64(Some(&results.get(&"2".to_string()).unwrap().0), Some(&0.20195), 5);
+        assert_eq_f64(Some(&results.get(&"4".to_string()).unwrap().0), Some(&0.20195), 5);
+        assert_eq_f64(Some(&results.get(&"3".to_string()).unwrap().0), Some(&0.20916), 5);
     }
 
     #[test]
@@ -235,19 +235,19 @@ mod page_rank_tests {
             graph.add_edge(t, src, dst, [], None).unwrap();
         }
 
-        let results: AlgorithmResult<OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(4), None, true);
+        let results: AlgorithmResult<String, OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(4), None, true);
 
-        assert_eq_f64(Some(&results.get("10").unwrap().0), Some(&0.072082), 5);
-        assert_eq_f64(Some(&results.get("8").unwrap().0), Some(&0.136473), 5);
-        assert_eq_f64(Some(&results.get("3").unwrap().0), Some(&0.15484), 5);
-        assert_eq_f64(Some(&results.get("6").unwrap().0), Some(&0.07208), 5);
-        assert_eq_f64(Some(&results.get("11").unwrap().0), Some(&0.06186), 5);
-        assert_eq_f64(Some(&results.get("2").unwrap().0), Some(&0.03557), 5);
-        assert_eq_f64(Some(&results.get("1").unwrap().0), Some(&0.11284), 5);
-        assert_eq_f64(Some(&results.get("4").unwrap().0), Some(&0.07944), 5);
-        assert_eq_f64(Some(&results.get("7").unwrap().0), Some(&0.01638), 5);
-        assert_eq_f64(Some(&results.get("9").unwrap().0), Some(&0.06186), 5);
-        assert_eq_f64(Some(&results.get("5").unwrap().0), Some(&0.19658), 5);
+        assert_eq_f64(Some(&results.get(&"10".to_string()).unwrap().0), Some(&0.072082), 5);
+        assert_eq_f64(Some(&results.get(&"8".to_string()).unwrap().0), Some(&0.136473), 5);
[…]
