### 2023-07-13 f53061b
convert the three node motifs
