+++ raphtory-graphql/src/data.rs
@@ -1,4 +1,8 @@
-use raphtory::{prelude::{Graph, GraphViewOps}, search::IndexedGraph, db::api::view::internal::{IntoDynamic, DynamicGraph}};
+use raphtory::{
+    db::api::view::internal::{DynamicGraph, IntoDynamic},
+    prelude::{Graph, GraphViewOps},
+    search::IndexedGraph,
+};
 use std::{
     collections::{HashMap, HashSet},
     path::Path,
@@ -41,7 +45,7 @@ impl Data {
             .map(|path| {
                 println!("loading graph from {path}");
                 let graph = Graph::load_from_file(&path).expect("Unable to load from graph");
-                let maybe_graph_name = graph.static_properties().get("name");
+                let maybe_graph_name = graph.properties().get("name");
 
                 return match maybe_graph_name {
                     None => {
@@ -54,7 +58,13 @@ impl Data {
                         (graph_name.to_string(), graph)
                     }
                 };
-            }).map(|(name, g)| (name, IndexedGraph::from_graph(&g.into_dynamic()).expect("Unable to index graph")))
+            })
+            .map(|(name, g)| {
+                (
+                    name,
+                    IndexedGraph::from_graph(&g.into_dynamic()).expect("Unable to index graph"),
+                )
+            })
             .collect();
 
         Self { graphs }
