### 2023-07-13 2b41910
fix windowed property keys
