+++ raphtory/src/python/graph/views/graph_view.rs
@@ -1,6 +1,6 @@
 //! The API for querying a view of the graph in a read-only state
-use crate::db::api::properties::TemporalProperties;
-use crate::db::api::properties::{Properties, StaticProperties};
+
+use crate::db::api::properties::Properties;
 use crate::{
     core::{
         entities::vertices::vertex_ref::VertexRef,
