+++ raphtory/src/db/api/view/edge.rs
@@ -1,8 +1,7 @@
 use crate::db::api::properties::internal::{
     StaticPropertiesOps, TemporalPropertiesOps, TemporalPropertyViewOps,
 };
-use crate::db::api::properties::TemporalProperties;
-use crate::db::api::properties::{Properties, StaticProperties};
+use crate::db::api::properties::Properties;
 use crate::{
     core::entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef},
     db::api::view::{internal::*, *},
