+++ raphtory/src/db/api/view/vertex.rs
@@ -1,5 +1,4 @@
-use crate::db::api::properties::TemporalProperties;
-use crate::db::api::properties::{Properties, StaticProperties};
+use crate::db::api::properties::Properties;
 use crate::db::api::view::{edge::EdgeListOps, GraphViewOps, TimeOps};
 use crate::db::graph::vertex::VertexView;
 
