### 2023-07-14 dcca71e
python properties api
