+++ raphtory/src/core/entities/edges/edge_ref.rs
@@ -1,121 +1,106 @@
+use bitvec::prelude::BitArray;
+
 use crate::core::entities::{vertices::vertex_ref::VertexRef, EID, VID};
 
+const LAYER_SIZE:usize = 4; // sizeof(usize) * 4
+
 #[derive(Debug, Copy, Clone, PartialEq)]
-pub enum EdgeRef {
-    LocalInto {
-        e_pid: EID,
-        layer_id: usize,
-        src_pid: VID,
-        dst_pid: VID,
-        time: Option<i64>,
-    },
-    LocalOut {
-        e_pid: EID,
-        layer_id: usize,
-        src_pid: VID,
-        dst_pid: VID,
-        time: Option<i64>,
-    },
+pub struct EdgeRef{
+    e_pid: EID,
+    src_pid: VID,
+    dst_pid: VID,
+    e_type: Dir,
+    time: Option<i64>,
+    // layer_id: Option<BitArray<[usize; LAYER_SIZE]>>,
+}
+
+#[derive(Debug, Copy, Clone, PartialEq)]
+enum Dir {
+    Into,
+    Out,
 }
 
 impl EdgeRef {
-    #[inline(always)]
-    pub fn layer(&self) -> usize {
-        match &self {
-            EdgeRef::LocalInto { layer_id, .. } => *layer_id,
-            EdgeRef::LocalOut { layer_id, .. } => *layer_id,
+
+    pub fn new_outgoing(e_pid: EID, src_pid: VID, dst_pid: VID) -> Self {
+        EdgeRef {
+            e_pid,
+            src_pid,
+            dst_pid,
+            e_type: Dir::Out,
+            time: None,
+            // layer_id: None,
         }
     }
+
+    pub fn new_incoming(e_pid: EID, src_pid: VID, dst_pid: VID) -> Self {
+        EdgeRef {
+            e_pid,
+            src_pid,
+            dst_pid,
+            e_type: Dir::Into,
+            time: None,
+            // layer_id: None,
+        }
+    }
+
+    // #[inline(always)]
+    // pub fn layers(&self) -> Option<Vec<usize>> {
+    //     self.layer_id.clone()
+    // }
 
     #[inline(always)]
     pub fn time(&self) -> Option<i64> {
-        match self {
-            EdgeRef::LocalInto { time, .. } => *time,
-            EdgeRef::LocalOut { time, .. } => *time,
-        }
+        self.time
     }
 
     pub fn src(&self) -> VertexRef {
-        match self {
-            EdgeRef::LocalInto { src_pid, .. } => (*src_pid).into(),
-            EdgeRef::LocalOut { src_pid, .. } => (*src_pid).into(),
-        }
+        self.src_pid.into()
     }
 
     pub fn dst(&self) -> VertexRef {
-        match self {
-            EdgeRef::LocalInto { dst_pid, .. } => (*dst_pid).into(),
-            EdgeRef::LocalOut { dst_pid, .. } => (*dst_pid).into(),
-        }
+        self.dst_pid.into()
     }
 
     pub fn remote(&self) -> VertexRef {
-        match self {
-            EdgeRef::LocalInto { .. } => self.src(),
-            EdgeRef::LocalOut { .. } => self.dst(),
+        match self.e_type {
+            Dir::Into => self.src(),
+            Dir::Out => self.dst(),
         }
     }
 
     pub fn local(&self) -> VertexRef {
-        match self {
-            EdgeRef::LocalInto { .. } => self.dst(),
-            EdgeRef::LocalOut { .. } => self.src(),
+        match self.e_type {
+            Dir::Into => self.dst(),
+            Dir::Out => self.src(),
         }
     }
 
[…]
