+++ raphtory/Cargo.toml
@@ -16,6 +16,7 @@ homepage.workspace = true
 
 [dependencies]
 bincode = "1"
+bitvec = "1"
 chrono = { version = "0.4", features = ["serde"] }
 flume = "0.10"
 futures = {version = "0.3", features = ["thread-pool"] }
