### 2023-07-14 b3a9ce6
add Rust Benchmark (cargo) benchmark result for 77cfb5c604c3d3eb23e44b29d17e98b60a4369b4
