+++ .github/workflows/test_python_workflow.yml
@@ -20,7 +20,7 @@ jobs:
       - id: set-matrix
         run: |
           if [ ${{ inputs.test_python_lower }} == true ]; then
-            echo "python-versions=[\"3.7\",\"3.11\"]" >> $GITHUB_OUTPUT
+            echo "python-versions=[\"3.8\",\"3.11\"]" >> $GITHUB_OUTPUT
           else
             echo "python-versions=[\"3.11\"]" >> $GITHUB_OUTPUT
           fi
