+++ .github/workflows/_release_python.yml
@@ -43,7 +43,6 @@ jobs:
       - uses: actions/setup-python@v4
         with:
           python-version: |
-            3.7
             3.8
             3.9
             3.10
@@ -116,7 +115,6 @@ jobs:
       - uses: actions/setup-python@v4
         with:
           python-version: |
-            3.7
             3.8
             3.9
             3.10
@@ -163,7 +161,6 @@ jobs:
       - uses: actions/setup-python@v4
         with:
           python-version: |
-            3.7
             3.8
             3.9
             3.10
