### 2023-07-14 7feb4fa
add Rust Benchmark (cargo) benchmark result for c44c2dfcfeac048f5491a972efe395bf6c646a32
