### 2023-07-14 1a926fb
support for vectorised property access
