### 2023-07-14 55435c6
impl expanded edges
