### 2023-07-14 653a084
actually add some methods for vectorised props
