+++ .github/workflows/test_python_workflow.yml
@@ -22,7 +22,7 @@ jobs:
           if [ ${{ inputs.test_python_lower }} == true ]; then
             echo "python-versions=[\"3.8\",\"3.11\"]" >> $GITHUB_OUTPUT
           else
-            echo "python-versions=[\"3.11\"]" >> $GITHUB_OUTPUT
+            echo "python-versions=[\"3.8\"]" >> $GITHUB_OUTPUT
           fi
   python-test:
     if: ${{ !inputs.skip_tests }}
