### 2023-07-14 052a6d6
add Rust Benchmark (cargo) benchmark result for 6e7d1c7d43293226a7b4de9052187f5997b901e6
