### 2023-07-17 a96b79b
sort of works but python api could be cleaner
