### 2023-07-17 e0dd3a4
add k core algorithm ahead of triangle optimisations
