+++ raphtory/src/core/entities/properties/props.rs
@@ -169,9 +169,17 @@ impl Meta {
 
     pub fn get_all_property_names(&self, is_static: bool) -> Vec<String> {
         if is_static {
-            self.meta_prop_static.map.iter().map(|entry| entry.key().clone()).collect()
+            self.meta_prop_static
+                .map
+                .iter()
+                .map(|entry| entry.key().clone())
+                .collect()
         } else {
-            self.meta_prop_temporal.map.iter().map(|entry| entry.key().clone()).collect()
+            self.meta_prop_temporal
+                .map
+                .iter()
+                .map(|entry| entry.key().clone())
+                .collect()
         }
     }
 
@@ -182,7 +190,6 @@ impl Meta {
             self.meta_prop_temporal.reverse_lookup(&prop_id)
         }
     }
-
 }
 
 #[derive(Serialize, Deserialize, Default, Debug)]
