+++ python/src/lib.rs
@@ -5,6 +5,7 @@ use raphtory_core::python::{
         edge::{PyEdge, PyEdges},
         graph::PyGraph,
         graph_with_deletions::PyGraphWithDeletions,
+        index::GraphIndex,
         vertex::{PyVertex, PyVertices},
     },
     packages::{algorithms::*, graph_gen::*, graph_loader::*},
@@ -15,6 +16,7 @@ use raphtory_core::python::{
 fn raphtory(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_class::<PyGraph>()?;
     m.add_class::<PyGraphWithDeletions>()?;
+    m.add_class::<GraphIndex>()?;
 
     let algorithm_module = PyModule::new(py, "algorithms")?;
     algorithm_module.add_function(wrap_pyfunction!(global_reciprocity, algorithm_module)?)?;
