+++ raphtory/src/python/graph/views/graph_view.rs
@@ -451,9 +451,7 @@ impl PyGraphView {
         self.graph.subgraph(vertices)
     }
 
-    /// Returns a graph clone
-    ///
-    /// Arguments:
+    /// Returns a 'materialized' clone of the graph view - i.e. a new graph with a copy of the data seen within the view instead of just a mask over the original graph
     ///
     /// Returns:
     ///    GraphView - Returns a graph clone
@@ -461,6 +459,12 @@ impl PyGraphView {
         self.graph.materialize()
     }
 
+    /// Indexes all vertex and edge properties.
+    /// Returns a GraphIndex which allows the user to search the edges and vertices of the graph via tantivity fuzzy matching queries.
+    /// Note this is currently immutable and will not update if the graph changes. This is to be improved in a future release.
+    ///
+    /// Returns:
+    ///    GraphIndex - Returns a GraphIndex
     fn index(&self) -> GraphIndex {
         GraphIndex::new(self.graph.clone())
     }
