### 2023-07-17 45574b9
more python vectorised apis
