+++ raphtory/src/db/graph/views/vertex_subgraph.rs
@@ -1,7 +1,7 @@
 use crate::db::api::properties::internal::InheritPropertiesOps;
 use crate::{
     core::{
-        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, VID, EID},
+        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, EID, VID},
         Direction,
     },
     db::api::view::internal::{
@@ -43,7 +43,6 @@ impl<G: GraphViewOps> VertexSubgraph<G> {
 }
 
 impl<G: GraphViewOps> GraphOps for VertexSubgraph<G> {
-
     fn find_edge_id(&self, e_id: EID) -> Option<EdgeRef> {
         let edge_ref = self.graph.find_edge_id(e_id)?;
         let vid_src = self.local_vertex_ref(edge_ref.src())?;
@@ -156,6 +155,7 @@ mod subgraph_tests {
         g.add_vertex(1, 1, NO_PROPS).unwrap();
         g.add_vertex(2, 2, NO_PROPS).unwrap();
         let sg = g.subgraph([1, 2]);
+
         let actual = sg.materialize().unwrap().into_events().unwrap();
         assert_eq!(actual, sg);
     }
