### 2023-07-17 42983d7
optimisations to triangle_count
