### 2023-07-17 062061e
optimisations to triangle_count
