+++ raphtory/src/algorithms/triangle_count.rs
@@ -6,10 +6,11 @@ use crate::{
             context::Context,
             task::{ATask, Job, Step},
             task_runner::TaskRunner,
-        },
+        }, graph::views::vertex_subgraph::{self, VertexSubgraph},
     },
 };
 use rustc_hash::FxHashSet;
+use crate::algorithms::k_core::k_core_set;
 
 /// Computes the number of triangles in a graph using a fast algorithm
 ///
@@ -55,9 +56,12 @@ use rustc_hash::FxHashSet;
 /// let actual_tri_count = triangle_count(&graph, None);
 /// ```
 ///
-pub fn triangle_count<G: GraphViewOps>(g: &G, threads: Option<usize>) -> usize {
-    let mut ctx: Context<G, ComputeStateVec> = g.into();
+pub fn triangle_count<G: GraphViewOps>(graph: &G, threads: Option<usize>) -> usize {
+    // let vertex_set = k_core_set(graph,2,usize::MAX, None);
+    // let g = graph.subgraph(vertex_set);
+    // let mut ctx: Context<VertexSubgraph<G>, ComputeStateVec> = Context::from(&g);
 
+    let mut ctx: Context<G, ComputeStateVec> = graph.into();
     let neighbours_set = accumulators::hash_set::<u64>(0);
     let count = accumulators::sum::<usize>(1);
 
