### 2023-07-17 b975ba4
add Rust Benchmark (cargo) benchmark result for c3c7f16d6d62e356f1c6055617485d98cb3aae89
