+++ raphtory/src/db/api/view/internal/core_ops.rs
@@ -126,7 +126,7 @@ pub trait CoreGraphOps {
     /// # Returns
     ///
     /// A vector of strings representing the names of the temporal properties
-    fn all_edge_prop_names(&self, is_static:bool)-> Vec<String>;
+    fn all_edge_prop_names(&self, is_static: bool) -> Vec<String>;
     /// Returns the static edge property with the given name for the
     /// given edge reference.
     ///
