+++ raphtory/src/db/task/task_runner.rs
@@ -222,9 +222,8 @@ impl<G: GraphViewOps, CS: ComputeState> TaskRunner<G, CS> {
         let morcel_size = num_vertices.min(16_000);
         let num_chunks = (num_vertices + morcel_size - 1) / morcel_size;
 
-        let mut shard_state = shard_initial_state.unwrap_or_else(|| {
-            Shard::new(num_vertices, num_chunks, morcel_size)
-        });
+        let mut shard_state = shard_initial_state
+            .unwrap_or_else(|| Shard::new(num_vertices, num_chunks, morcel_size));
 
         let mut global_state = global_initial_state.unwrap_or_else(|| Global::new());
 
