+++ raphtory/src/python/graph/mod.rs
@@ -1,6 +1,6 @@
 pub mod edge;
 pub mod graph;
 pub mod graph_with_deletions;
-pub mod vertex;
 pub mod pandas;
+pub mod vertex;
 pub mod views;
