### 2023-07-17 de8e694
added python support for algorithm result object
