+++ raphtory/src/python/graph/mod.rs
@@ -3,3 +3,4 @@ pub mod graph;
 pub mod graph_with_deletions;
 pub mod vertex;
 pub mod views;
+mod algorithm_result;
