+++ python/src/lib.rs
@@ -19,7 +19,6 @@ fn raphtory(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     let algorithm_module = PyModule::new(py, "algorithms")?;
     algorithm_module.add_function(wrap_pyfunction!(global_reciprocity, algorithm_module)?)?;
     algorithm_module.add_function(wrap_pyfunction!(all_local_reciprocity, algorithm_module)?)?;
-    algorithm_module.add_function(wrap_pyfunction!(all_local_reciprocity_df, algorithm_module)?)?;
 
     algorithm_module.add_function(wrap_pyfunction!(triplet_count, algorithm_module)?)?;
     algorithm_module.add_function(wrap_pyfunction!(

[…]
