### 2023-07-17 53253f5
add tests for new pyalgo fns
