### 2023-07-17 d5e9ae9
fix the issue with compute and subgraphs
