### 2023-07-17 a51d8ce
change multiple layers edge_tuples to avoid duplicate edges
