### 2023-07-17 3153169
Bug/compute state (#1122)
* add test for triangle count on subgraph

* fix the issue with compute and subgraphs

---------

Co-authored-by: Lucas Jeub <lucas.jeub@pometry.com>
