### 2023-07-17 e369e02
add Rust Benchmark (cargo) benchmark result for 496a7fd4908205b2655feb71cb8401ce10ed905a
