+++ raphtory/src/algorithms/local_clustering_coefficient.rs
@@ -79,10 +79,13 @@ pub fn local_clustering_coefficient<G: GraphViewOps, V: Into<VertexRef>>(
 #[cfg(test)]
 mod clustering_coefficient_tests {
     use super::local_clustering_coefficient;
-    use crate::{db::{
-        api::{mutation::AdditionOps, view::*},
-        graph::graph::Graph,
-    }, prelude::NO_PROPS};
+    use crate::{
+        db::{
+            api::{mutation::AdditionOps, view::*},
+            graph::graph::Graph,
+        },
+        prelude::NO_PROPS,
+    };
 
     #[test]
     fn clusters_of_triangles() {
