+++ raphtory-graphql/src/model/graph/graph.rs
@@ -7,10 +7,13 @@ use crate::model::{
 };
 use dynamic_graphql::{ResolvedObject, ResolvedObjectFields};
 use itertools::Itertools;
-use raphtory::{db::api::view::{
-    internal::{DynamicGraph, IntoDynamic},
-    GraphViewOps, TimeOps, VertexViewOps,
-}, search::IndexedGraph};
+use raphtory::{
+    db::api::view::{
+        internal::{DynamicGraph, IntoDynamic},
+        GraphViewOps, TimeOps, VertexViewOps,
+    },
+    search::IndexedGraph,
+};
 
 #[derive(ResolvedObject)]
 pub(crate) struct GraphMeta {
@@ -61,7 +64,6 @@ impl<G: GraphViewOps + IntoDynamic> From<G> for GqlGraph {
 }
 
 impl GqlGraph {
-
     pub(crate) fn new(graph: IndexedGraph<DynamicGraph>) -> Self {
         Self { graph }
     }
@@ -69,7 +71,6 @@ impl GqlGraph {
 
 #[ResolvedObjectFields]
 impl GqlGraph {
-
     async fn window(&self, t_start: i64, t_end: i64) -> GqlGraph {
         let w = self.graph.window(t_start, t_end);
         w.into()
@@ -104,7 +105,6 @@ impl GqlGraph {
             .map(|vv| vv.into())
             .collect()
     }
-
 
     async fn search_edges(&self, query: String, limit: usize, offset: usize) -> Vec<Edge> {
         self.graph
