### 2023-07-18 b888a9d
Fixed formatting (#1118)
* Fixed formatting

* Formatting master
