+++ raphtory/src/db/graph/vertex.rs
@@ -494,7 +494,13 @@ mod vertex_test {
 
         let v1 = g.vertex(1).unwrap();
         let v1_w = g.window(0, 1).vertex(1).unwrap();
-        assert_eq!(v1.properties(false), props.into_iter().map(|(k, v)| (k.to_string(), v.as_prop())).collect());
+        assert_eq!(
+            v1.properties(false),
+            props
+                .into_iter()
+                .map(|(k, v)| (k.to_string(), v.as_prop()))
+                .collect()
+        );
         assert_eq!(v1_w.properties(false), HashMap::default())
     }
 }
