### 2023-07-18 4c84d70
triangle motifs first run
