+++ raphtory/src/algorithms/triangle_count.rs
@@ -57,11 +57,11 @@ use crate::algorithms::k_core::k_core_set;
 /// ```
 ///
 pub fn triangle_count<G: GraphViewOps>(graph: &G, threads: Option<usize>) -> usize {
-    // let vertex_set = k_core_set(graph,2,usize::MAX, None);
-    // let g = graph.subgraph(vertex_set);
-    // let mut ctx: Context<VertexSubgraph<G>, ComputeStateVec> = Context::from(&g);
+    let vertex_set = k_core_set(graph,2,usize::MAX, None);
+    let g = graph.subgraph(vertex_set);
+    let mut ctx: Context<VertexSubgraph<G>, ComputeStateVec> = Context::from(&g);
 
-    let mut ctx: Context<G, ComputeStateVec> = graph.into();
+    // let mut ctx: Context<G, ComputeStateVec> = graph.into();
     let neighbours_set = accumulators::hash_set::<u64>(0);
     let count = accumulators::sum::<usize>(1);
 
@@ -107,7 +107,7 @@ pub fn triangle_count<G: GraphViewOps>(graph: &G, threads: Option<usize>) -> usi
     let init_tasks = vec![Job::new(step1)];
     let tasks = vec![Job::new(step2)];
 
-    let mut runner: TaskRunner<G, _> = TaskRunner::new(ctx);
+    let mut runner: TaskRunner<VertexSubgraph<G>, _> = TaskRunner::new(ctx);
 
     runner.run(
         init_tasks,
