+++ raphtory/src/algorithms/motifs/three_node_local_redone.rs
@@ -1,0 +1,325 @@
+// Imports ///////////////////////////////////////////
+use crate::{
+    algorithms::{motifs::three_node_motifs::*, k_core::k_core_set},
+    core::state::{accumulator_id::{accumulators::{self, val}, AccId}, compute_state::ComputeStateVec, agg::ValDef},
+    db::{
+        api::view::{GraphViewOps, VertexViewOps, *},
+        task::{
+            context::Context,
+            task::{ATask, Job, Step},
+            task_runner::TaskRunner,
+            vertex::eval_vertex::EvalVertexView,
+        }, graph::views::vertex_subgraph::{self, VertexSubgraph},
+    },
+};
+
+use num_traits::Zero;
+use rustc_hash::FxHashSet;
+use std::{collections::HashMap, ops::Add, slice::Iter};
+///////////////////////////////////////////////////////
+
+// State objects for three node motifs
+#[derive(Eq, PartialEq, Clone, Debug, Default)]
+pub struct MotifCounter {
+    pub two_nodes: [usize; 8],
+    pub star_nodes: [usize; 24],
+    pub triangle: [usize; 8],
+}
+
+impl MotifCounter {
+    fn new(two_nodes: [usize; 8], star_nodes: [usize; 24], triangle: [usize; 8]) -> Self {
+        Self {
+            two_nodes,
+            star_nodes,
+            triangle,
+        }
+    }
+
+    pub(crate) fn from_triangle_counter(triangle: [usize; 8]) -> Self {
+        Self {
+            two_nodes: [0; 8],
+            star_nodes: [0; 24],
+            triangle,
+        }
+    }
+}
+
+impl Add for MotifCounter {
+    type Output = MotifCounter;
+
+    fn add(self, rhs: Self) -> Self::Output {
+        rhs
+    }
+}
+
+impl Zero for MotifCounter {
+    fn zero() -> Self {
+        MotifCounter {
+            two_nodes: [0; 8],
+            star_nodes: [0; 24],
+            triangle: [0; 8],
+        }
+    }
+
+    fn set_zero(&mut self) {
+        *self = Zero::zero();
+    }
+
+    fn is_zero(&self) -> bool {
+        self.two_nodes == [0; 8] && self.star_nodes == [0; 24] && self.triangle == [0; 8]
+    }
+}
+
+///////////////////////////////////////////////////////
+
+fn update_counter<G: GraphViewOps>(
+    vs: Vec<&EvalVertexView<G, ComputeStateVec, MotifCounter>>,
+    motif_counter: AccId<MotifCounter, MotifCounter, MotifCounter, ValDef<MotifCounter>>,
+    tmp_counts: Iter<usize>,
+) {
+    for v in vs {
+        let mc = v.read(&motif_counter);
+        let triangle: [usize; 8] = mc
+            .triangle
+            .iter()
+            .zip(tmp_counts.clone())
+            .map(|(&i1, &i2)| i1 + i2)
+            .collect::<Vec<usize>>()
+            .try_into()
+            .unwrap();
+        v.update(
+            &motif_counter,
+            MotifCounter::from_triangle_counter(triangle),
+        );
+    }
+}
+
+///////////////////////////////////////////////////////
+
+pub fn star_motif_count<G: GraphViewOps>(
+    evv: &EvalVertexView<G, ComputeStateVec, MotifCounter>,
+    delta: i64,
+) -> [usize; 24] {
+    let neigh_map: HashMap<u64, usize> = evv
+        .neighbours()
+        .into_iter()
+        .enumerate()
+        .map(|(num, nb)| (nb.id(), num))
+        .into_iter()
+        .collect();
+    let mut exploded_edges = evv
+        .edges()
+        .explode()
[…]
