### 2023-07-18 ff7c730
add Rust Benchmark (cargo) benchmark result for b888a9d286cd499a8375d1f2ebc2722a676822d9
