### 2023-07-18 d1370bc
even more vectorised operations
