### 2023-07-18 8384c6d
add tests, add to_df fn
