+++ raphtory/src/python/graph/algorithm_result.rs
@@ -1,4 +1,3 @@
-use crate::algorithms::algorithm_result::AlgorithmResult;
 use ordered_float::OrderedFloat;
 use pyo3::prelude::*;
 
@@ -26,6 +25,28 @@ macro_rules! py_algorithm_result_base {
 
             fn get(&self, key: $rustKey) -> Option<$rustValue> {
                 self.0.get(&key).cloned()
+            }
+
+            #[pyo3(signature = ())]
+            pub fn to_df(
+                &self,
+            ) -> PyResult<PyObject>
+            {
+                let hashmap = &self.0.result;
+                let mut keys = Vec::new();
+                let mut values = Vec::new();
+                Python::with_gil(|py| {
+                    for (key, value) in hashmap.iter() {
+                        keys.push(key.to_object(py));
+                        values.push(value.to_object(py));
+                    }
+                    let dict = pyo3::types::PyDict::new(py);
+                    dict.set_item("Key", pyo3::types::PyList::new(py, keys.as_slice()))?;
+                    dict.set_item("Value", pyo3::types::PyList::new(py, values.as_slice()))?;
+                    let pandas = pyo3::types::PyModule::import(py, "pandas")?;
+                    let df: &PyAny = pandas.getattr("DataFrame")?.call1((dict,))?;
+                    Ok(df.to_object(py))
+                })
             }
         }
     }
@@ -82,11 +103,13 @@ py_algorithm_result!(AlgorithmResultStrF64, String, OrderedFloat<f64>);
 
 #[pymethods]
 impl AlgorithmResultStrF64 {
+    #[pyo3(signature = ())]
     fn get_all(&self) -> std::collections::HashMap<String, f64> {
         self.0.get_all().into_iter()
             .map(|(key, of)| (key.clone(), of.into_inner()))
             .collect()
     }
+
 
     fn get(&self, key: String) -> Option<f64> {
         Some(self.0.get(&key).unwrap().0)
@@ -117,6 +140,28 @@ impl AlgorithmResultStrF64 {
             })
     }
 
+    #[pyo3(signature = ())]
+    fn to_df(&self) -> PyResult<PyObject> {
+        let hashmap: std::collections::HashMap<String, f64> = self.0.get_all().clone().into_iter()
+            .map(|(k, v)| (k, v.into()))
+            .collect();
+        let mut keys = Vec::new();
+        let mut values = Vec::new();
+        Python::with_gil(|py| {
+            for (key, value) in hashmap.iter() {
+                keys.push(key.to_object(py));
+                values.push(value.to_object(py));
+            }
+            let dict = pyo3::types::PyDict::new(py);
+            dict.set_item("Key", pyo3::types::PyList::new(py, keys.as_slice()))?;
+            dict.set_item("Value", pyo3::types::PyList::new(py, values.as_slice()))?;
+            let pandas = PyModule::import(py, "pandas")?;
+            let df: &PyAny = pandas.getattr("DataFrame")?.call1((dict,))?;
+            Ok(df.to_object(py))
+        })
+    }
+
+    #[pyo3(signature = ())]
     fn group_by(&self) -> std::collections::HashMap<String, Vec<String>> {
         let ordered_map = self.0.group_by();
         let mut f64_map: std::collections::HashMap<String, Vec<String>> = std::collections::HashMap::new();
