+++ raphtory/src/algorithms/algorithm_result.rs
@@ -3,14 +3,21 @@ use std::fmt;
 use std::fmt::Debug;
 use ordered_float::OrderedFloat;
 use std::hash::Hash;
-// use pyo3::prelude::*;
-// use pyo3::types::{IntoPyDict, PyDict, PyList};
 
+/// A generic `AlgorithmResult` struct that represents the result of an algorithm computation.
+///
+/// The `AlgorithmResult` contains a hashmap, where keys (`H`) are cloneable, hashable, and comparable,
+/// and values (`Y`) are cloneable. The keys and values can be of any type that satisfies the specified
+/// trait bounds.
+///
+/// This `AlgorithmResult` is returned for all algorithms that return a HashMap
+///
 pub struct AlgorithmResult<H, Y>
     where
         H: Clone + Hash + Eq + Ord,
         Y: Clone
 {
+    /// The result hashmap that stores keys of type `H` and values of type `Y`.
     pub result: HashMap<H, Y>,
 }
 
@@ -19,42 +26,30 @@ impl<H, Y> AlgorithmResult<H, Y>
         H: Clone + Hash + Eq + Ord,
         Y: Clone,
 {
+    /// Creates a new instance of `AlgorithmResult` with the provided hashmap.
+    ///
+    /// # Arguments
+    ///
+    /// * `result`: A `HashMap` with keys of type `H` and values of type `Y`.
     pub fn new(result: HashMap<H, Y>) -> Self {
         Self {
             result,
         }
     }
 
+    /// Returns a reference to the entire `result` hashmap.
     pub fn get_all(&self) -> &HashMap<H, Y> {
         &self.result
     }
 
+    /// Returns the value corresponding to the provided key in the `result` hashmap.
+    ///
+    /// # Arguments
+    ///
+    /// * `key`: The key of type `H` for which the value is to be retrieved.
     pub fn get(&self, key: &H) -> Option<&Y> {
         self.result.get(&key)
     }
-
-    // pub fn to_df<'a>(
-    //     &self,
-    //     py: Python<'a>
-    // ) -> PyResult<PyObject>
-    // where
-    //     H: Clone + ToPyObject,
-    //     Y: Clone + ToPyObject,
-    // {
-    //     let hashmap = &self.result;
-    //     let mut keys = Vec::new();
-    //     let mut values = Vec::new();
-    //     for (key, value) in hashmap.iter() {
-    //         keys.push(key.to_object(py));
-    //         values.push(value.to_object(py));
-    //     }
-    //     let dict = PyDict::new(py);
-    //     dict.set_item("Key", PyList::new(py, keys.as_slice()))?;
-    //     dict.set_item("Value", PyList::new(py, values.as_slice()))?;
-    //     let pandas = pyo3::types::PyModule::import(py, "pandas")?;
-    //     let df: &PyAny = pandas.getattr("DataFrame")?.call1((dict,))?;
-    //     Ok(df.to_object(py))
-    // }
 }
 
 impl<H, Y> AlgorithmResult<H, Y>
@@ -62,6 +57,15 @@ impl<H, Y> AlgorithmResult<H, Y>
         H: Clone + Hash + Eq + Ord,
         Y: Clone + PartialOrd
 {
+    /// Sorts the `AlgorithmResult` by its values in ascending or descending order.
+    ///
+    /// # Arguments
+    ///
+    /// * `reverse`: If `true`, sorts the result in descending order; otherwise, sorts in ascending order.
+    ///
+    /// # Returns
+    ///
+    /// A sorted vector of tuples containing keys of type `H` and values of type `Y`.
     pub fn sort_by_value(&self, reverse: bool) -> Vec<(H, Y)> {
[…]
