### 2023-07-18 cb2f245
Adding layers to load_edges_from_num_iter
