+++ raphtory-graphql/src/data.rs
@@ -1,8 +1,8 @@
 use raphtory::{
+    db::api::view::internal::{DynamicGraph, IntoDynamic},
     prelude::{Graph, GraphViewOps},
     search::IndexedGraph,
 };
-use raphtory::{db::api::view::internal::{IntoDynamic, DynamicGraph}};
 use std::{
     collections::{HashMap, HashSet},
     path::Path,
@@ -24,14 +24,21 @@ impl Data {
         Self { graphs }
     }
 
-    pub fn from_map_and_directory(graphs: HashMap<String, DynamicGraph>, directory_path: &str) -> Self {
+    pub fn from_map_and_directory(
+        graphs: HashMap<String, DynamicGraph>,
+        directory_path: &str,
+    ) -> Self {
         let graphs = Self::convert_graphs(graphs);
         let mut graphs_from_files = Self::load_from_file(directory_path);
         graphs_from_files.extend(graphs);
-        Self { graphs: graphs_from_files }
+        Self {
+            graphs: graphs_from_files,
+        }
     }
 
-    fn convert_graphs(graphs: HashMap<String, DynamicGraph>) -> HashMap<String, IndexedGraph<DynamicGraph>> {
+    fn convert_graphs(
+        graphs: HashMap<String, DynamicGraph>,
+    ) -> HashMap<String, IndexedGraph<DynamicGraph>> {
         graphs
             .into_iter()
             .map(|(name, g)| {
@@ -46,10 +53,7 @@ impl Data {
     fn load_from_file(path: &str) -> HashMap<String, IndexedGraph<DynamicGraph>> {
         let mut valid_paths = HashSet::<String>::new();
 
-        for entry in WalkDir::new(path)
-            .into_iter()
-            .filter_map(|e| e.ok())
-        {
+        for entry in WalkDir::new(path).into_iter().filter_map(|e| e.ok()) {
             let path = entry.path();
             let path_string = path.display().to_string();
             let filename = path.file_name().and_then(|name| name.to_str());
