+++ raphtory/src/python/graph/graph.rs
@@ -219,7 +219,7 @@ impl PyGraph {
     }
 
     #[staticmethod]
-    #[pyo3(signature = (edges_df, src = "source", dst = "destination", time = "time", props = None, layer = None, vertex_df = None, vertex_col = None, vertex_time_col = None, vertex_props = None))]
+    #[pyo3(signature = (edges_df, src = "source", dst = "destination", time = "time", props = None, layer = None, layer_in_df = None, vertex_df = None, vertex_col = None, vertex_time_col = None, vertex_props = None))]
     fn load_from_pandas(
         edges_df: &PyAny,
         src: &str,
@@ -227,6 +227,7 @@ impl PyGraph {
         time: &str,
         props: Option<Vec<&str>>,
         layer: Option<&str>,
+        layer_in_df: Option<&str>,
         vertex_df: Option<&PyAny>,
         vertex_col: Option<&str>,
         vertex_time_col: Option<&str>,
@@ -235,7 +236,7 @@ impl PyGraph {
         let graph = PyGraph {
             graph: Graph::new(),
         };
-        graph.load_edges_from_pandas(edges_df, src, dst, time, props,layer)?;
+        graph.load_edges_from_pandas(edges_df, src, dst, time, props,layer,layer_in_df)?;
         if let (Some(vertex_df), Some(vertex_col), Some(vertex_time_col)) =
             (vertex_df, vertex_col, vertex_time_col)
         {
@@ -269,7 +270,7 @@ impl PyGraph {
         Ok(())
     }
 
-    #[pyo3(signature = (edge_df, src_col = "source", dst_col = "destination", time_col = "time", props = None, layer=None))]
+    #[pyo3(signature = (edge_df, src_col = "source", dst_col = "destination", time_col = "time", props = None, layer=None,layer_in_df=None))]
     fn load_edges_from_pandas(
         &self,
         edge_df: &PyAny,
@@ -278,11 +279,12 @@ impl PyGraph {
         time_col: &str,
         props: Option<Vec<&str>>,
         layer: Option<&str>,
+        layer_in_df: Option<&str>,
     ) -> Result<(), GraphError> {
         let graph = &self.graph;
         Python::with_gil(|py| {
             let df = process_pandas_py_df(edge_df, py)?;
-            load_edges_from_df(&df, src_col, dst_col, time_col, props, layer, graph)
+            load_edges_from_df(&df, src_col, dst_col, time_col, props, layer, layer_in_df, graph)
                 .map_err(|e| GraphLoadException::new_err(format!("{:?}", e)))?;
 
             Ok::<(), PyErr>(())
