### 2023-07-18 e54be47
surface the multi layer API changes to the graph traits
