### 2023-07-18 ef8d5b1
Feature/layering in pandas (#1124)
* Adding layers to load_edges_from_num_iter

* Added explit layer setting

* Fixed test

* Fix test

* fix tests
