### 2023-07-18 69d55be
add Rust Benchmark (cargo) benchmark result for ef8d5b107feeae0b0b332de2e6fa7d67edcfa01b
