### 2023-07-18 41882c7
fix graph ql issues during compiling
