+++ raphtory-graphql/src/model/algorithm.rs
@@ -12,6 +12,7 @@ use raphtory::{
     db::api::view::{internal::DynamicGraph, GraphViewOps},
 };
 use std::{borrow::Cow, collections::HashMap, sync::Mutex};
+use ordered_float::OrderedFloat;
 
 type RegisterFunction = fn(&str, Registry, Object) -> (Registry, Object);
 
@@ -95,6 +96,12 @@ impl From<(String, f64)> for Pagerank {
     }
 }
 
+impl From<(&String, &OrderedFloat<f64>)> for Pagerank {
+    fn from((name, rank): (&String, &OrderedFloat<f64>)) -> Self {
+        Self { name: name.to_string(), rank: rank.into_inner() }
+    }
+}
+
 impl Algorithm for Pagerank {
     fn output_type() -> TypeRef {
         // first _nn means that the list is never null, second _nn means no element is null
@@ -115,7 +122,8 @@ impl Algorithm for Pagerank {
         let threads = ctx.args.get("threads").map(|v| v.u64()).transpose()?;
         let threads = threads.map(|v| v as usize);
         let tol = ctx.args.get("tol").map(|v| v.f64()).transpose()?;
-        let result = unweighted_page_rank(graph, iter_count, threads, tol, true)
+        let binding = unweighted_page_rank(graph, iter_count, threads, tol, true);
+        let result = binding
             .into_iter()
             .map(|pair| FieldValue::owned_any(Pagerank::from(pair)));
         Ok(Some(FieldValue::list(result)))
