+++ raphtory/src/python/graph/algorithm_result.rs
@@ -19,14 +19,25 @@ macro_rules! py_algorithm_result_base {
     ($name:ident, $rustKey:ty, $rustValue:ty) => {
         #[pymethods]
         impl $name {
+            /// Returns a reference to the entire `result` hashmap.
             fn get_all(&self) -> std::collections::HashMap<$rustKey, $rustValue> {
                 self.0.get_all().clone()
             }
 
+            /// Returns the value corresponding to the provided key in the `result` hashmap.
+            ///
+            /// # Arguments
+            ///
+            /// * `key`: The key of type `H` for which the value is to be retrieved.
             fn get(&self, key: $rustKey) -> Option<$rustValue> {
                 self.0.get(&key).cloned()
             }
 
+            /// Creates a dataframe from the result
+            ///
+            /// # Returns
+            ///
+            /// A `pandas.DataFrame` containing the result
             #[pyo3(signature = ())]
             pub fn to_df(
                 &self,
@@ -56,16 +67,49 @@ macro_rules! py_algorithm_result_partial_ord {
     ($name:ident, $rustKey:ty, $rustValue:ty) => {
         #[pymethods]
         impl $name {
+
+            /// Sorts the `AlgorithmResult` by its values in ascending or descending order.
+            ///
+            /// # Arguments
+            ///
+            /// * `reverse`: If `true`, sorts the result in descending order; otherwise, sorts in ascending order.
+            ///
+            /// # Returns
+            ///
+            /// A sorted vector of tuples containing keys of type `H` and values of type `Y`.
             #[pyo3(signature = (reverse=true))]
             fn sort_by_value(&self, reverse: bool) -> Vec<($rustKey, $rustValue)> {
                 self.0.sort_by_value(reverse)
             }
 
+            /// Sorts the `AlgorithmResult` by its keys in ascending or descending order.
+            ///
+            /// # Arguments
+            ///
+            /// * `reverse`: If `true`, sorts the result in descending order; otherwise, sorts in ascending order.
+            ///
+            /// # Returns
+            ///
+            /// A sorted vector of tuples containing keys of type `H` and values of type `Y`.
             #[pyo3(signature = (reverse=true))]
             fn sort_by_key(&self, reverse: bool) -> Vec<($rustKey, $rustValue)> {
                 self.0.sort_by_key(reverse)
             }
 
+            /// Retrieves the top-k elements from the `AlgorithmResult` based on its values.
+            ///
+            /// # Arguments
+            ///
+            /// * `k`: The number of elements to retrieve.
+            /// * `percentage`: If `true`, the `k` parameter is treated as a percentage of total elements.
+            /// * `reverse`: If `true`, retrieves the elements in descending order; otherwise, in ascending order.
+            ///
+            /// # Returns
+            ///
+            /// An `Option` containing a vector of tuples with keys of type `H` and values of type `Y`.
+            /// If `percentage` is `true`, the returned vector contains the top `k` percentage of elements.
+            /// If `percentage` is `false`, the returned vector contains the top `k` elements.
+            /// Returns `None` if the result is empty or if `k` is 0.
             #[pyo3(signature = (k, percentage=false, reverse=true))]
             fn top_k(&self, k: usize, percentage: bool, reverse: bool) -> Option<Vec<($rustKey, $rustValue)>> {
                 self.0.top_k(k, percentage, reverse)
[…]
