### 2023-07-19 419b9bd
vectorised NestedTemporalProperty
