### 2023-07-19 e96207c
:boom: add needs check to prevent premature startup
