+++ .github/workflows/test_during_pr.yml
@@ -17,16 +17,20 @@ jobs:
     name: Run Rust tests
     uses: ./.github/workflows/test_rust_workflow.yml
     secrets: inherit
+    needs: rust-format-check
   call-test-python-workflow-in-local-repo:
     name: Run Python tests
     uses: ./.github/workflows/test_python_workflow.yml
     with:
       test_python_lower: false
     secrets: inherit
+    needs: rust-format-check
   call-benchmark-workflow-in-local-repo:
     name: Run benchmarks
     uses: ./.github/workflows/benchmark.yml
     secrets: inherit
+    needs: rust-format-check
   call-code-coverage:
     name: Code Coverage
     uses: ./.github/workflows/code_coverage.yml
+    needs: rust-format-check
