+++ raphtory/src/db/task/edge/window_eval_edge.rs
@@ -175,7 +175,7 @@ impl<'a, G: GraphViewOps, CS: ComputeState, S: 'static> EdgeViewOps
                         self.eref().dst(),
                         t,
                         t.saturating_add(1),
-                        0, // FIXME should be able to pass in an array of layers
+                        0.into(), // FIXME should be able to pass in an array of layers
                     )
             }
         }
