### 2023-07-19 3fa1b52
implement vectorised static properties and start experimenting with __eq__ support
