+++ raphtory/src/python/graph/properties/static_props.rs
@@ -4,9 +4,14 @@ use crate::db::api::properties::StaticProperties;
 use crate::db::api::view::internal::Static;
 use crate::python::graph::properties::DynProps;
 use crate::python::types::repr::Repr;
-use crate::python::utils::PyGenericIterator;
-use pyo3::exceptions::PyKeyError;
-use pyo3::{pyclass, pymethods, PyResult};
+use crate::python::types::wrappers::iterators::{NestedOptionPropIterable, OptionPropIterable};
+use crate::python::utils::{PyGenericIterator, PyNestedGenericIterator};
+use itertools::Itertools;
+use pyo3::exceptions::{PyKeyError, PyNotImplementedError, PyTypeError};
+use pyo3::prelude::*;
+use pyo3::pyclass::CompareOp;
+use std::collections::HashMap;
+use std::ops::Deref;
 use std::sync::Arc;
 
 pub type DynStaticProperties = StaticProperties<DynProps>;
@@ -18,6 +23,33 @@ impl<P: PropertiesOps + Send + Sync + Static + 'static> From<StaticProperties<P>
         StaticProperties {
             props: Arc::new(value.props),
         }
+    }
+}
+
+#[derive(PartialEq)]
+pub struct StaticPropsComparable(HashMap<String, Prop>);
+
+impl<'source> FromPyObject<'source> for StaticPropsComparable {
+    fn extract(ob: &'source PyAny) -> PyResult<Self> {
+        if let Ok(sp) = ob.extract::<PyRef<PyStaticProperties>>() {
+            Ok(sp.deref().into())
+        } else if let Ok(m) = ob.extract::<HashMap<String, Prop>>() {
+            Ok(StaticPropsComparable(m))
+        } else {
+            Err(PyTypeError::new_err("not comparable with properties"))
+        }
+    }
+}
+
+impl From<&PyStaticProperties> for StaticPropsComparable {
+    fn from(value: &PyStaticProperties) -> Self {
+        Self(value.props.as_map())
+    }
+}
+
+impl From<DynStaticProperties> for StaticPropsComparable {
+    fn from(value: DynStaticProperties) -> Self {
+        Self(value.as_map())
     }
 }
 
@@ -73,5 +105,143 @@ impl<P: PropertiesOps + Send + Sync + 'static> From<StaticProperties<P>> for PyS
 impl Repr for PyStaticProperties {
     fn repr(&self) -> String {
         self.props.repr()
+    }
+}
+
+py_iterable!(
+    StaticPropsIterable,
+    DynStaticProperties,
+    PyStaticProperties,
+    PyGenericIterator
+);
+
+pub enum StaticPropsIterComparable {
+    Vec(Vec<StaticPropsComparable>),
+    This(Py<StaticPropsIterable>),
+}
+
+impl<'source> FromPyObject<'source> for StaticPropsIterComparable {
+    fn extract(ob: &'source PyAny) -> PyResult<Self> {
+        if let Ok(s) = ob.extract::<Py<StaticPropsIterable>>() {
+            Ok(StaticPropsIterComparable::This(s))
+        } else if let Ok(v) = ob.extract::<Vec<StaticPropsComparable>>() {
+            Ok(StaticPropsIterComparable::Vec(v))
+        } else {
+            Err(PyTypeError::new_err("cannot compare"))
+        }
+    }
+}
+
+#[pymethods]
+impl StaticPropsIterable {
+    fn keys(&self) -> Vec<String> {
+        self.iter().map(|p| p.keys()).kmerge().dedup().collect()
+    }
+
+    fn values(&self) -> Vec<OptionPropIterable> {
+        self.keys()
+            .into_iter()
[…]

[…]
