### 2023-07-19 d427825
add bootstrap file, hooks and instructions
