### 2023-07-19 0a49a27
:art: fixed most issues
