+++ raphtory/src/algorithms/pagerank.rs
@@ -1,4 +1,5 @@
 use crate::{
+    algorithms::algorithm_result::AlgorithmResult,
     core::{
         entities::VID,
         state::{accumulator_id::accumulators, compute_state::ComputeStateVec},
@@ -13,9 +14,8 @@ use crate::{
     },
 };
 use num_traits::abs;
-use std::collections::HashMap;
 use ordered_float::OrderedFloat;
-use crate::algorithms::algorithm_result::AlgorithmResult;
+use std::collections::HashMap;
 
 #[derive(Clone, Debug)]
 struct PageRankState {
@@ -175,7 +175,8 @@ pub fn unweighted_page_rank<G: GraphViewOps>(
         None,
     );
 
-    let res = out.into_iter()
+    let res = out
+        .into_iter()
         .map(|(k, v)| (g.vertex_name(k), v))
         .collect();
 
@@ -211,12 +212,29 @@ mod page_rank_tests {
     fn test_page_rank() {
         let graph = load_graph();
 
-        let results: AlgorithmResult<String, OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(1), None, true);
+        let results: AlgorithmResult<String, OrderedFloat<f64>> =
+            unweighted_page_rank(&graph, 1000, Some(1), None, true);
 
-        assert_eq_f64(Some(&results.get(&"1".to_string()).unwrap().0), Some(&0.38694), 5);
-        assert_eq_f64(Some(&results.get(&"2".to_string()).unwrap().0), Some(&0.20195), 5);
-        assert_eq_f64(Some(&results.get(&"4".to_string()).unwrap().0), Some(&0.20195), 5);
-        assert_eq_f64(Some(&results.get(&"3".to_string()).unwrap().0), Some(&0.20916), 5);
+        assert_eq_f64(
+            Some(&results.get(&"1".to_string()).unwrap().0),
+            Some(&0.38694),
+            5,
+        );
+        assert_eq_f64(
+            Some(&results.get(&"2".to_string()).unwrap().0),
+            Some(&0.20195),
+            5,
+        );
+        assert_eq_f64(
+            Some(&results.get(&"4".to_string()).unwrap().0),
+            Some(&0.20195),
+            5,
+        );
+        assert_eq_f64(
+            Some(&results.get(&"3".to_string()).unwrap().0),
+            Some(&0.20916),
+            5,
+        );
     }
 
     #[test]
@@ -253,19 +271,64 @@ mod page_rank_tests {
             graph.add_edge(t, src, dst, NO_PROPS, None).unwrap();
         }
 
-        let results: AlgorithmResult<String, OrderedFloat<f64>> = unweighted_page_rank(&graph, 1000, Some(4), None, true);
+        let results: AlgorithmResult<String, OrderedFloat<f64>> =
+            unweighted_page_rank(&graph, 1000, Some(4), None, true);
 
-        assert_eq_f64(Some(&results.get(&"10".to_string()).unwrap().0), Some(&0.072082), 5);
-        assert_eq_f64(Some(&results.get(&"8".to_string()).unwrap().0), Some(&0.136473), 5);
-        assert_eq_f64(Some(&results.get(&"3".to_string()).unwrap().0), Some(&0.15484), 5);
-        assert_eq_f64(Some(&results.get(&"6".to_string()).unwrap().0), Some(&0.07208), 5);
-        assert_eq_f64(Some(&results.get(&"11".to_string()).unwrap().0), Some(&0.06186), 5);
-        assert_eq_f64(Some(&results.get(&"2".to_string()).unwrap().0), Some(&0.03557), 5);
-        assert_eq_f64(Some(&results.get(&"1".to_string()).unwrap().0), Some(&0.11284), 5);
[…]
