+++ raphtory/src/algorithms/algorithm_result.rs
@@ -1,8 +1,5 @@
-use std::collections::HashMap;
-use std::fmt;
-use std::fmt::Debug;
 use ordered_float::OrderedFloat;
-use std::hash::Hash;
+use std::{collections::HashMap, fmt, fmt::Debug, hash::Hash};
 
 /// A generic `AlgorithmResult` struct that represents the result of an algorithm computation.
 ///
@@ -12,33 +9,31 @@ use std::hash::Hash;
 ///
 /// This `AlgorithmResult` is returned for all algorithms that return a HashMap
 ///
-pub struct AlgorithmResult<H, Y>
-    where
-        H: Clone + Hash + Eq + Ord,
-        Y: Clone
+pub struct AlgorithmResult<K, V>
+where
+    K: Clone + Hash + Eq + Ord,
+    V: Clone,
 {
     /// The result hashmap that stores keys of type `H` and values of type `Y`.
-    pub result: HashMap<H, Y>,
+    pub result: HashMap<K, V>,
 }
 
-impl<H, Y> AlgorithmResult<H, Y>
-    where
-        H: Clone + Hash + Eq + Ord,
-        Y: Clone,
+impl<K, V> AlgorithmResult<K, V>
+where
+    K: Clone + Hash + Eq + Ord,
+    V: Clone,
 {
     /// Creates a new instance of `AlgorithmResult` with the provided hashmap.
     ///
     /// # Arguments
     ///
     /// * `result`: A `HashMap` with keys of type `H` and values of type `Y`.
-    pub fn new(result: HashMap<H, Y>) -> Self {
-        Self {
-            result,
-        }
+    pub fn new(result: HashMap<K, V>) -> Self {
+        Self { result }
     }
 
     /// Returns a reference to the entire `result` hashmap.
-    pub fn get_all(&self) -> &HashMap<H, Y> {
+    pub fn get_all(&self) -> &HashMap<K, V> {
         &self.result
     }
 
@@ -47,31 +42,30 @@ impl<H, Y> AlgorithmResult<H, Y>
     /// # Arguments
     ///
     /// * `key`: The key of type `H` for which the value is to be retrieved.
-    pub fn get(&self, key: &H) -> Option<&Y> {
+    pub fn get(&self, key: &K) -> Option<&V> {
         self.result.get(&key)
     }
 }
 
-pub struct AlgorithmResultIterator<'a, H, Y>
-{
-    iter: std::collections::hash_map::Iter<'a, H, Y>,
+pub struct AlgorithmResultIterator<'a, K, V> {
+    iter: std::collections::hash_map::Iter<'a, K, V>,
 }
 
-impl<'a, H, Y> Iterator for AlgorithmResultIterator<'a, H, Y> {
-    type Item = (&'a H, &'a Y);
+impl<'a, K, V> Iterator for AlgorithmResultIterator<'a, K, V> {
+    type Item = (&'a K, &'a V);
 
     fn next(&mut self) -> Option<Self::Item> {
         self.iter.next()
     }
 }
 
-impl<'a, H, Y> IntoIterator for &'a AlgorithmResult<H, Y>
+impl<'a, K, V> IntoIterator for &'a AlgorithmResult<K, V>
 where
-    H: Clone + Hash + Ord,
-    Y: Clone
+    K: Clone + Hash + Ord,
+    V: Clone,
 {
-    type Item = (&'a H, &'a Y);
-    type IntoIter = AlgorithmResultIterator<'a, H, Y>;
+    type Item = (&'a K, &'a V);
+    type IntoIter = AlgorithmResultIterator<'a, K, V>;
 
     fn into_iter(self) -> Self::IntoIter {
[…]

[…]
