### 2023-07-20 c43207a
Adds a format checker to the workflows, which fails if the code has not been formatted, adds a dev folder with instructions to setup a nightly env + point the git hooks to the local folder (#1126)
* fixs docs

* fixed spelling, be me, cant spell

* :boom: add needs check to prevent premature startup

* :zap: prevent failure

* testing hook

* add bootstrap file, hooks and instructions

* testing hook
