### 2023-07-20 8fb34c6
add Rust Benchmark (cargo) benchmark result for c43207a451a7c8e7b4d4d93f2e1f6c3e30428ac7
