### 2023-07-20 9d921d2
macro for implementing == for iterables
