+++ raphtory/src/python/types/macros/iterable.rs
@@ -180,40 +180,50 @@ macro_rules! py_float_iterable {
     };
 }
 
-// macro_rules! py_iterable_comp {
-//     ($name:ident, $item:ty, $cmp_item:ty, $cmp_internal:ident) => {
-//         enum $cmp_internal {
-//             Vec(Vec<$cmp_item>),
-//             This(Py<$name>)
-//         }
-//         
-//         impl FromPyObject
-//         
-//         #[pymethods]
-//         impl $name {
-//             pub fn __richcmp__(
-//                 &self,
-//                 other: $cmp_item,
-//                 op: CompareOp,
-//                 py: Python<'_>,
-//             ) -> PyResult<bool> {
-//                 match op {
-//                     CompareOp::Lt => Err(PyNotImplementedError::new_err("not ordered")),
-//                     CompareOp::Le => Err(PyNotImplementedError::new_err("cannot compare")),
-//                     CompareOp::Eq => {
-//                         if let Ok(s) = other.extract::<PyRef<Self>>(py) {
-//                             Ok(self.iter().zip(s.iter()).all(|(s, o)| $cmp_item::from(s) == $cmp_item::from(o)))
-//                         } else if let Ok(s) = other.extract::<Vec<HashMap<String, Prop>>>(py) {
-//                             Ok(self.iter().zip(s).all(|(p1, p2)| p1.as_map() == p2))
-//                         } else {
-//                             Ok(false)
-//                         }
-//                     }
-//                     CompareOp::Ne => Ok(!self.__richcmp__(other, CompareOp::Eq, py)?),
-//                     CompareOp::Gt => Err(PyNotImplementedError::new_err("cannot compare")),
-//                     CompareOp::Ge => Err(PyNotImplementedError::new_err("cannot compare")),
-//                 }
-//             }
-//         }
-//     };
-// }
+macro_rules! py_iterable_comp {
+    ($name:ident, $item:ty, $cmp_item:ty, $cmp_internal:ident) => {
+        enum $cmp_internal {
+            Vec(Vec<$cmp_item>),
+            This(Py<$name>),
+        }
+
+        impl<'source> FromPyObject<'source> for $cmp_internal {
+            fn extract(ob: &'source PyAny) -> PyResult<Self> {
+                if let Ok(s) = ob.extract::<Py<$name>>() {
+                    Ok($cmp_internal::This(s))
+                } else if let Ok(v) = ob.extract::<Vec<$cmp_item>>() {
+                    Ok($cmp_internal::Vec(v))
+                } else {
+                    Err(PyTypeError::new_err("cannot compare"))
+                }
+            }
+        }
+
+        #[pymethods]
+        impl $name {
+            fn __richcmp__(
+                &self,
+                other: $cmp_internal,
+                op: CompareOp,
+                py: Python<'_>,
+            ) -> PyResult<bool> {
+                match op {
+                    CompareOp::Lt => Err(PyNotImplementedError::new_err("not ordered")),
+                    CompareOp::Le => Err(PyNotImplementedError::new_err("cannot compare")),
+                    CompareOp::Eq => match other {
+                        $cmp_internal::Vec(v) => {
+                            Ok(self.iter().zip(v).all(|(t, o)| <$cmp_item>::from(t) == o))
+                        }
+                        $cmp_internal::This(o) => Ok(self
+                            .iter()
+                            .zip(o.borrow(py).iter())
+                            .all(|(t, o)| <$cmp_item>::from(t) == <$cmp_item>::from(o))),
+                    },
+                    CompareOp::Ne => Ok(!self.__richcmp__(other, CompareOp::Eq, py)?),
[…]
