### 2023-07-20 1677da7
Algorithms now have their own result object with many features (#1105)
* initial version of creating a polaris df from polars and fixing all the str& their dodgy prelude breaks

* hello algorithm result, how do you do you?

* fix tests for pr

* checkin

* hits fxhashmap you will not be missed 🫡

* checkin with my code :salute:

* at this point only god knows how it works. added tests for all types we currently support in our algorithms

* add cargo missing ordered float package

* fix tests

* all algos done except three node motifs

* implement debug formatter for algorithm result object to 
[…]
