+++ raphtory/src/python/graph/mod.rs
@@ -1,3 +1,4 @@
+pub mod algorithm_result;
 pub mod edge;
 pub mod graph;
 pub mod graph_with_deletions;
