+++ raphtory/src/python/graph/algorithm_result.rs
@@ -1,0 +1,283 @@
+use ordered_float::OrderedFloat;
+use pyo3::prelude::*;
+
+/// Create a macro for py_algorithm_result
+macro_rules! py_algorithm_result {
+    ($name:ident, $rustKey:ty, $rustValue:ty) => {
+        #[pyclass]
+        pub struct $name(
+            $crate::algorithms::algorithm_result::AlgorithmResult<$rustKey, $rustValue>,
+        );
+
+        impl pyo3::IntoPy<pyo3::PyObject>
+            for $crate::algorithms::algorithm_result::AlgorithmResult<$rustKey, $rustValue>
+        {
+            fn into_py(self, py: Python<'_>) -> pyo3::PyObject {
+                $name(self).into_py(py)
+            }
+        }
+    };
+}
+
+macro_rules! py_algorithm_result_base {
+    ($name:ident, $rustKey:ty, $rustValue:ty) => {
+        #[pymethods]
+        impl $name {
+            /// Returns a reference to the entire `result` hashmap.
+            fn get_all(&self) -> std::collections::HashMap<$rustKey, $rustValue> {
+                self.0.get_all().clone()
+            }
+
+            /// Returns the value corresponding to the provided key in the `result` hashmap.
+            ///
+            /// # Arguments
+            ///
+            /// * `key`: The key of type `H` for which the value is to be retrieved.
+            fn get(&self, key: $rustKey) -> Option<$rustValue> {
+                self.0.get(&key).cloned()
+            }
+
+            /// Creates a dataframe from the result
+            ///
+            /// # Returns
+            ///
+            /// A `pandas.DataFrame` containing the result
+            pub fn to_df(&self) -> PyResult<PyObject> {
+                let hashmap = &self.0.result;
+                let mut keys = Vec::new();
+                let mut values = Vec::new();
+                Python::with_gil(|py| {
+                    for (key, value) in hashmap.iter() {
+                        keys.push(key.to_object(py));
+                        values.push(value.to_object(py));
+                    }
+                    let dict = pyo3::types::PyDict::new(py);
+                    dict.set_item("Key", pyo3::types::PyList::new(py, keys.as_slice()))?;
+                    dict.set_item("Value", pyo3::types::PyList::new(py, values.as_slice()))?;
+                    let pandas = pyo3::types::PyModule::import(py, "pandas")?;
+                    let df: &PyAny = pandas.getattr("DataFrame")?.call1((dict,))?;
+                    Ok(df.to_object(py))
+                })
+            }
+        }
+    };
+}
+
+macro_rules! py_algorithm_result_partial_ord {
+    ($name:ident, $rustKey:ty, $rustValue:ty) => {
+        #[pymethods]
+        impl $name {
+            /// Sorts the `AlgorithmResult` by its values in ascending or descending order.
+            ///
+            /// # Arguments
+            ///
+            /// * `reverse`: If `true`, sorts the result in descending order; otherwise, sorts in ascending order.
+            ///
+            /// # Returns
+            ///
+            /// A sorted vector of tuples containing keys of type `H` and values of type `Y`.
+            #[pyo3(signature = (reverse=true))]
+            fn sort_by_value(&self, reverse: bool) -> Vec<($rustKey, $rustValue)> {
+                self.0.sort_by_value(reverse)
+            }
+
+            /// Sorts the `AlgorithmResult` by its keys in ascending or descending order.
+            ///
+            /// # Arguments
+            ///
+            /// * `reverse`: If `true`, sorts the result in descending order; otherwise, sorts in ascending order.
+            ///
[…]
