+++ python/tests/test_graphdb.py
@@ -4,6 +4,8 @@ import sys
 import time
 import datetime
 
+import pandas as pd
+import pandas.core.frame
 import pytest
 from raphtory import Graph, GraphWithDeletions
 from raphtory import algorithms
@@ -11,7 +13,6 @@ from raphtory import graph_loader
 import tempfile
 from math import isclose
 import datetime
-
 
 edges = [
     (1, 1, 2),
@@ -667,7 +668,7 @@ def test_graph_time_api():
     earliest_time = g.earliest_time()
     latest_time = g.latest_time()
     assert len(list(g.rolling(1))) == latest_time - earliest_time + 1
-    assert len(list(g.expanding(2))) == math.ceil((latest_time+1 - earliest_time) / 2)
+    assert len(list(g.expanding(2))) == math.ceil((latest_time + 1 - earliest_time) / 2)
 
     w = g.window(2, 6)
     assert len(list(w.rolling(window=10, step=3))) == 2
@@ -974,7 +975,7 @@ def test_lotr_edge_history():
     assert (g.edge('Frodo', 'Gandalf').window(100, 1000).history() == [329, 555, 861])
 
 
-def test_connected_components():
+def gen_graph():
     g = Graph()
     g.add_edge(10, 1, 3, {})
     g.add_edge(11, 1, 2, {})
@@ -988,28 +989,57 @@ def test_connected_components():
     g.add_edge(15, 4, 7, {})
     g.add_edge(10, 4, 7, {})
     g.add_edge(10, 5, 8, {})
+    return g
+
+
+def test_connected_components():
+    g = gen_graph()
+    actual = algorithms.weakly_connected_components(g, 20)
+    expected = {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1, '6': 1, '7': 1, '8': 1}
+    assert (actual.get_all() == expected)
+    assert (actual.get('1') == 1)
+
+
+def test_algo_result():
+    g = gen_graph()
 
     actual = algorithms.weakly_connected_components(g, 20)
     expected = {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1, '6': 1, '7': 1, '8': 1}
-    assert (actual == expected)
+    assert (actual.get_all() == expected)
+    assert (actual.get('1') == 1)
+    expected_array = [('1', 1), ('2', 1), ('3', 1), ('4', 1), ('5', 1), ('6', 1), ('7', 1), ('8', 1)]
+    assert (sorted(actual.sort_by_value()) == expected_array)
+    assert (actual.sort_by_key() == sorted(expected_array, reverse=True))
+    assert (actual.sort_by_key(reverse=False) == expected_array)
+    assert (sorted(actual.top_k(8)) == expected_array)
+    assert (len(actual.group_by()[1]) == 8)
+    assert type(actual.to_df()) == pandas.core.frame.DataFrame
+    df = actual.to_df()
+    expected_result = pd.DataFrame({
+        'Key': ['1'],
+        'Value': [1]
+    })
+    row_with_one = df[df['Key'] == '1']
+    row_with_one.reset_index(inplace=True, drop=True)
+    assert row_with_one.equals(expected_result)
+    # Algo Str u64
+    actual = algorithms.weakly_connected_components(g)
+    all_res = actual.get_all()
+    sorted_res = {k: all_res[k] for k in sorted(all_res)}
[…]
