+++ docs/source/api/raphtory.rst
@@ -25,6 +25,14 @@ Algorithms
    :private-members:
    :inherited-members:
 
+.. :: raphtory.AlgorithmResultStrU64
+   :autosummary:
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+
 
 Visualisation
 -------------------
