### 2023-07-20 069a4bd
add Rust Benchmark (cargo) benchmark result for ba8daba4e23ae19c2f1d97537a4c2b9f633c6100
