### 2023-07-20 2f3dee9
need to find a way to filter edges by time
