+++ raphtory/src/core/entities/graph/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod timer;
 mod test {
     use crate::{
         core::Direction,
-        prelude::{AsProp, NO_PROPS},
+        prelude::{AsProp, NO_PROPS, Prop},
     };
 
     use super::{tgraph::InnerTemporalGraph, *};
@@ -51,5 +51,23 @@ mod test {
         let edges = first.edge_tuples([0, 1, 2, 3, 4].into(), Direction::OUT).collect::<Vec<_>>();
 
         assert_eq!(edges.len(), 1, "should only have one edge {:?}", edges);
+    }
+
+    #[test]
+    fn simple_triangle() {
+
+        let g: InnerTemporalGraph<2> = InnerTemporalGraph::default();
+        let vs = vec![(1, 1, 2), (2, 1, 3), (3, 2, 1), (4, 3, 2)];
+
+        let empty:Vec<(String, Prop)> = vec![];
+        for (t, src, dst) in vs{
+            g.add_edge_internal(t, src, dst, empty.clone(), None);
+        }
+
+        let v = g.vertex(0.into());
+
+        let ns = v.neighbours(vec![], Direction::BOTH).map(|v| v.id().0).collect::<Vec<_>>();
+        assert_eq!(ns, vec![1, 2]);
+
     }
 }
