+++ raphtory/src/algorithms/local_triangle_count.rs
@@ -44,6 +44,8 @@ use itertools::Itertools;
 pub fn local_triangle_count<G: GraphViewOps, V: Into<VertexRef>>(graph: &G, v: V) -> Option<usize> {
     if let Some(vertex) = graph.vertex(v) {
         if vertex.degree() >= 2 {
+            let wtf = vertex.neighbours().id().collect_vec();
+            println!("wtf: {:?}", wtf);
             let x: Vec<usize> = vertex
                 .neighbours()
                 .id()
