### 2023-07-20 c2e445d
add hits to python and remove window
