+++ python/tests/test_graphdb.py
@@ -1,0 +1,1520 @@
+import math
+import re
+import sys
+import time
+import datetime
+
+import pandas as pd
+import pandas.core.frame
+import pytest
+from raphtory import Graph, GraphWithDeletions
+from raphtory import algorithms
+from raphtory import graph_loader
+import tempfile
+from math import isclose
+import datetime
+
+edges = [
+    (1, 1, 2),
+    (2, 1, 3),
+    (-1, 2, 1),
+    (0, 1, 1),
+    (7, 3, 2),
+    (1, 1, 1)
+]
+
+
+def create_graph():
+    g = Graph()
+
+    g.add_vertex(0, 1, {"type": "wallet", "cost": 99.5})
+    g.add_vertex(-1, 2, {"type": "wallet", "cost": 10.0})
+    g.add_vertex(6, 3, {"type": "wallet", "cost": 76})
+
+    for e in edges:
+        g.add_edge(e[0], e[1], e[2], {"prop1": 1,
+                                      "prop2": 9.8, "prop3": "test"})
+
+    return g
+
+
+def create_graph_with_deletions():
+    g = GraphWithDeletions()
+
+    g.add_vertex(0, 1, {"type": "wallet", "cost": 99.5})
+    g.add_vertex(-1, 2, {"type": "wallet", "cost": 10.0})
+    g.add_vertex(6, 3, {"type": "wallet", "cost": 76})
+
+    for e in edges:
+        g.add_edge(e[0], e[1], e[2], {"prop1": 1,
+                                      "prop2": 9.8, "prop3": "test"})
+
+    g.add_edge_properties(edges[0][1], edges[0][2], {"static": "test"})
+    g.delete_edge(10, edges[0][1], edges[0][2])
+    return g
+
+
+def test_graph_len_edge_len():
+    g = create_graph()
+
+    assert g.num_vertices() == 3
+    assert g.num_edges() == 5
+
+
+def test_id_iterable():
+    g = create_graph()
+
+    assert g.vertices.id().max() == 3
+    assert g.vertices.id().min() == 1
+    assert set(g.vertices.id().collect()) == {1, 2, 3}
+    out_neighbours = g.vertices.out_neighbours().id().collect()
+    out_neighbours = (set(n) for n in out_neighbours)
+    out_neighbours = dict(zip(g.vertices.id(), out_neighbours))
+
+    assert out_neighbours == {1: {1, 2, 3}, 2: {1}, 3: {2}}
+
+
+def test_degree_iterable():
+    g = create_graph()
+    assert g.vertices.degree().min() == 2
+    assert g.vertices.degree().max() == 3
+    assert g.vertices.in_degree().min() == 1
+    assert g.vertices.in_degree().max() == 2
+    assert g.vertices.out_degree().min() == 1
+    assert g.vertices.out_degree().max() == 3
+    assert isclose(g.vertices.degree().mean(), 7 / 3)
+    assert g.vertices.degree().sum() == 7
+    degrees = g.vertices.degree().collect()
+    degrees.sort()
+    assert degrees == [2, 2, 3]
+
+
+def test_vertices_time_iterable():
+    g = create_graph()
+
+    assert g.vertices.earliest_time().min() == -1
+    assert g.vertices.latest_time().max() == 7
+
+
+def test_graph_has_edge():
+    g = create_graph()
+
+    assert not g.window(-1, 1).has_edge(1, 3)
+    assert g.window(-1, 3).has_edge(1, 3)
+    assert not g.window(10, 11).has_edge(1, 3)
+
+
+def test_graph_has_vertex():
+    g = create_graph()
+
+    assert g.has_vertex(3)
+
+
[…]
