### 2023-07-20 e5c7624
Fixed tests for windows
