### 2023-07-20 dd0917a
Feature/python graphql (#1095)
* Commit for merging with search

* Messing with async

* temp

* Running from files working - running from dict next

* Fixed segfault with run_from_dict

* Merged and fixed issues

* Hiding async bits away

* Added port

* Running the server in the background

* Added a proper health check

* Basic documentation

* Added docs

* Added tests

* fmt

* revert file name change

* Added dependencies

* Fixed tests for windows
