+++ python/python/raphtory/__init__.py
@@ -3,7 +3,7 @@ from .raphtory import *
 sys.modules["raphtory.algorithms"] = algorithms
 sys.modules["raphtory.graph_gen"] = graph_gen
 sys.modules["raphtory.graph_loader"] = graph_loader
-
+#sys.modules["raphtory.graphql"] = graphql
 
 from .nullmodels import *
 
