+++ python/src/lib.rs
@@ -1,4 +1,7 @@
+mod graphql;
+
 extern crate core;
+use graphql::*;
 use pyo3::prelude::*;
 use raphtory_core::python::{
     graph::{
@@ -13,20 +16,41 @@ use raphtory_core::python::{
 /// Raphtory graph analytics library
 #[pymodule]
 fn raphtory(py: Python<'_>, m: &PyModule) -> PyResult<()> {
+    //Graph classes
     m.add_class::<PyGraph>()?;
     m.add_class::<PyGraphWithDeletions>()?;
+    m.add_class::<PyVertex>()?;
+    m.add_class::<PyVertices>()?;
+    m.add_class::<PyEdge>()?;
+    m.add_class::<PyEdges>()?;
 
+    //GRAPHQL
+    let graphql_module = PyModule::new(py, "internal_graphql")?;
+    graphql_module.add_function(wrap_pyfunction!(from_map, graphql_module)?)?;
+    graphql_module.add_function(wrap_pyfunction!(from_directory, graphql_module)?)?;
+    graphql_module.add_function(wrap_pyfunction!(from_map_and_directory, graphql_module)?)?;
+    m.add_submodule(graphql_module)?;
+
+    //ALGORITHMS
     let algorithm_module = PyModule::new(py, "algorithms")?;
     algorithm_module.add_function(wrap_pyfunction!(global_reciprocity, algorithm_module)?)?;
     algorithm_module.add_function(wrap_pyfunction!(all_local_reciprocity, algorithm_module)?)?;
     m.add_class::<AlgorithmResultStrU64>()?;
 
     algorithm_module.add_function(wrap_pyfunction!(triplet_count, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(local_triangle_count, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(average_degree, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(directed_graph_density, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(max_out_degree, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(max_in_degree, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(min_out_degree, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(min_in_degree, algorithm_module)?)?;
+    algorithm_module.add_function(wrap_pyfunction!(pagerank, algorithm_module)?)?;
     algorithm_module.add_function(wrap_pyfunction!(
         global_clustering_coefficient,
         algorithm_module
     )?)?;
-    algorithm_module.add_function(wrap_pyfunction!(local_triangle_count, algorithm_module)?)?;
+
     algorithm_module.add_function(wrap_pyfunction!(
         temporally_reachable_nodes,
         algorithm_module
@@ -35,13 +59,6 @@ fn raphtory(py: Python<'_>, m: &PyModule) -> PyResult<()> {
         local_clustering_coefficient,
         algorithm_module
     )?)?;
-    algorithm_module.add_function(wrap_pyfunction!(average_degree, algorithm_module)?)?;
-    algorithm_module.add_function(wrap_pyfunction!(directed_graph_density, algorithm_module)?)?;
-    algorithm_module.add_function(wrap_pyfunction!(max_out_degree, algorithm_module)?)?;
-    algorithm_module.add_function(wrap_pyfunction!(max_in_degree, algorithm_module)?)?;
[…]
