### 2023-07-20 cbd08c5
add Rust Benchmark (cargo) benchmark result for dd0917ae4a69c73d9f5448224b932b38bfc5d358
