+++ raphtory/src/python/graph/properties/temporal_props.rs
@@ -5,9 +5,8 @@ use crate::db::api::properties::StaticProperties;
 use crate::db::api::properties::{TemporalProperties, TemporalPropertyView};
 use crate::db::api::view::internal::{DynamicGraph, Static};
 use crate::python::graph::properties::static_props::PyStaticProperties;
-use crate::python::graph::properties::DynProps;
+use crate::python::graph::properties::{DynProps, NestedOptionPropIterable, OptionPropIterable};
 use crate::python::types::repr::{iterator_dict_repr, iterator_repr, Repr};
-use crate::python::types::wrappers::iterators::{NestedOptionPropIterable, OptionPropIterable};
 use crate::python::utils::{PyGenericIterator, PyNestedGenericIterator, PyTime};
 use itertools::Itertools;
 use pyo3::exceptions::PyKeyError;
