### 2023-07-21 37a1888
completely remove macro-based python iterators as they have no advantage over the generic ones
