+++ raphtory/src/python/types/macros/mod.rs
@@ -1,6 +1,4 @@
 #[macro_use]
-pub mod iter;
-#[macro_use]
 pub mod iterable;
 #[macro_use]
 pub mod nested_iterable;
