+++ raphtory/src/python/types/wrappers/iterators.rs
@@ -1,14 +1,7 @@
-use crate::python::graph::properties::{
-    DynStaticProperties, DynTemporalProperties, PyStaticProperties, PyTemporalProperties,
-};
-use crate::python::utils::{PyGenericIterator, PyNestedGenericIterator};
 use crate::{
     core as db_c,
     db::api::view::BoxedIter,
-    python::types::{
-        repr::Repr,
-        wrappers::prop::{PropHistory, PropValue},
-    },
+    python::types::{repr::Repr, wrappers::prop::PropHistory},
 };
 use num::cast::AsPrimitive;
 use pyo3::prelude::*;
@@ -35,106 +28,53 @@ where
 
 impl<I: ?Sized + Iterator<Item = V>, V: AsPrimitive<f64> + Sum<V>> MeanExt<V> for I {}
 
-py_iterator!(Float64Iter, f64);
-py_float_iterable!(Float64Iterable, f64, Float64Iter);
+py_float_iterable!(Float64Iterable, f64);
 
-py_iterator!(U64Iter, u64);
-py_numeric_iterable!(U64Iterable, u64, U64Iter);
-py_iterator!(NestedU64Iter, BoxedIter<u64>, U64Iter);
-py_nested_numeric_iterable!(
-    NestedU64Iterable,
-    u64,
-    NestedU64Iter,
-    U64Iterable,
-    OptionU64Iterable
-);
+py_numeric_iterable!(U64Iterable, u64);
+py_nested_numeric_iterable!(NestedU64Iterable, u64, U64Iterable, OptionU64Iterable);
 
-py_iterable!(OptionU64U64Iterable, Option<(u64, u64)>, PyGenericIterator);
-py_ordered_iterable!(U64U64Iterable, (u64, u64), PyGenericIterator);
-py_nested_ordered_iterable!(
-    NestedU64U64Iterable,
-    (u64, u64),
-    PyNestedGenericIterator,
-    OptionU64U64Iterable
-);
+py_iterable!(OptionU64U64Iterable, Option<(u64, u64)>);
+py_ordered_iterable!(U64U64Iterable, (u64, u64));
+py_nested_ordered_iterable!(NestedU64U64Iterable, (u64, u64), OptionU64U64Iterable);
 
-py_iterator!(OptionU64Iter, Option<u64>);
-py_iterable!(OptionU64Iterable, Option<u64>, Option<u64>, OptionU64Iter);
+py_iterable!(OptionU64Iterable, Option<u64>, Option<u64>);
 _py_ord_max_min_methods!(OptionU64Iterable, Option<u64>);
 
-py_iterator!(I64Iter, i64);
-py_numeric_iterable!(I64Iterable, i64, I64Iter);
-py_iterator!(NestedI64Iter, BoxedIter<i64>, I64Iter);
-py_nested_numeric_iterable!(
-    NestedI64Iterable,
-    i64,
-    NestedI64Iter,
-    I64Iterable,
-    OptionI64Iterable
-);
+py_numeric_iterable!(I64Iterable, i64);
+py_nested_numeric_iterable!(NestedI64Iterable, i64, I64Iterable, OptionI64Iterable);
 
-py_iterator!(OptionI64Iter, Option<i64>);
-py_iterable!(OptionI64Iterable, Option<i64>, OptionI64Iter);
+py_iterable!(OptionI64Iterable, Option<i64>);
 _py_ord_max_min_methods!(OptionI64Iterable, Option<i64>);
-py_iterator!(OptionOptionI64Iter, Option<Option<i64>>);
-py_iterable!(
-    OptionOptionI64Iterable,
-    Option<Option<i64>>,
-    OptionOptionI64Iter
-);
[…]
