+++ raphtory/src/db/graph/views/layer_graph.rs
@@ -1,12 +1,15 @@
+use std::ops::Range;
+
 use crate::{
     core::{
         entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, LayerIds, EID, VID},
         Direction,
     },
-    db::api::view::{internal::{
-        Base, GraphOps, InheritCoreOps, InheritMaterialize, InheritTimeSemantics,
-    }, Layer},
-    prelude::GraphViewOps,
+    db::api::view::{
+        internal::{Base, GraphOps, InheritCoreOps, InheritMaterialize, TimeSemantics},
+        BoxedIter, Layer,
+    },
+    prelude::{GraphViewOps, Prop},
 };
 use itertools::Itertools;
 
@@ -26,7 +29,7 @@ impl<G: GraphViewOps> Base for LayeredGraph<G> {
     }
 }
 
-impl<G: GraphViewOps> InheritTimeSemantics for LayeredGraph<G> {}
+// impl<G: GraphViewOps> InheritTimeSemantics for LayeredGraph<G> {}
 
 impl<G: GraphViewOps> InheritCoreOps for LayeredGraph<G> {}
 
@@ -34,10 +37,7 @@ impl<G: GraphViewOps> InheritMaterialize for LayeredGraph<G> {}
 
 impl<G: GraphViewOps> LayeredGraph<G> {
     pub fn new(graph: G, layers: LayerIds) -> Self {
-        Self {
-            graph,
-            layers,
-        }
+        Self { graph, layers }
     }
 
     /// Return None if the intersection between the previously requested layers and the layer of
@@ -62,9 +62,9 @@ impl<G: GraphViewOps> LayeredGraph<G> {
 impl<G: GraphViewOps> GraphOps for LayeredGraph<G> {
     fn find_edge_id(&self, e_id: EID) -> Option<EdgeRef> {
         let edge_ref = self.graph.find_edge_id(e_id)?;
-        let edge_ref_in_layer = self
-            .graph
-            .has_edge_ref(edge_ref.src(), edge_ref.dst(), self.layers.clone());
+        let edge_ref_in_layer =
+            self.graph
+                .has_edge_ref(edge_ref.src(), edge_ref.dst(), self.layers.clone());
 
         if edge_ref_in_layer {
             Some(edge_ref)
@@ -99,7 +99,7 @@ impl<G: GraphViewOps> GraphOps for LayeredGraph<G> {
             .unwrap_or(0)
     }
 
-    fn has_edge_ref(&self, src: VertexRef, dst: VertexRef, layer: LayerIds ) -> bool {
+    fn has_edge_ref(&self, src: VertexRef, dst: VertexRef, layer: LayerIds) -> bool {
         // FIXME: there is something wrong here, the layer should be able to be None, which would mean, whatever layer this is
         self.constrain(layer)
             .map(|layer| self.graph.has_edge_ref(src, dst, layer))
@@ -142,7 +142,7 @@ impl<G: GraphViewOps> GraphOps for LayeredGraph<G> {
         d: Direction,
         layer: LayerIds,
     ) -> Box<dyn Iterator<Item = EdgeRef> + Send> {
-self.constrain(layer)
+        self.constrain(layer)
             .map(|layer| self.graph.vertex_edges(v, d, layer))
             .unwrap_or_else(|| Box::new(std::iter::empty()))
     }
@@ -161,5 +161,122 @@ self.constrain(layer)
     fn get_layer_ids(&self, e_id: EID) -> Option<LayerIds> {
         let layer_ids = self.graph.get_layer_ids(e_id)?;
[…]
