### 2023-07-21 c338404
fix semantics for Properties iterables
