### 2023-07-21 6c45519
implemented TimeSemantics for LayeredGraph
