+++ raphtory/src/db/graph/views/window_graph.rs
@@ -39,7 +39,7 @@
 
 use crate::{
     core::{
-        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, EID, VID, LayerIds},
+        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, LayerIds, EID, VID},
         utils::time::IntoTime,
         Direction, Prop,
     },
@@ -135,8 +135,11 @@ impl<G: GraphViewOps> TimeSemantics for WindowedGraph<G> {
     }
 
     fn include_edge_window(&self, e: EdgeRef, w: Range<i64>, layer_ids: LayerIds) -> bool {
-        self.graph
-            .include_edge_window(e, self.actual_start(w.start)..self.actual_end(w.end), layer_ids)
+        self.graph.include_edge_window(
+            e,
+            self.actual_start(w.start)..self.actual_end(w.end),
+            layer_ids,
+        )
     }
 
     fn vertex_history(&self, v: VID) -> Vec<i64> {
@@ -150,12 +153,16 @@ impl<G: GraphViewOps> TimeSemantics for WindowedGraph<G> {
     }
 
     fn edge_t(&self, e: EdgeRef, layer_ids: LayerIds) -> BoxedIter<EdgeRef> {
-        self.graph.edge_window_t(e, self.t_start..self.t_end, layer_ids)
+        self.graph
+            .edge_window_t(e, self.t_start..self.t_end, layer_ids)
     }
 
     fn edge_window_t(&self, e: EdgeRef, w: Range<i64>, layer_ids: LayerIds) -> BoxedIter<EdgeRef> {
-        self.graph
-            .edge_window_t(e, self.actual_start(w.start)..self.actual_end(w.end), layer_ids)
+        self.graph.edge_window_t(
+            e,
+            self.actual_start(w.start)..self.actual_end(w.end),
+            layer_ids,
+        )
     }
 
     fn edge_earliest_time(&self, e: EdgeRef, layer_ids: LayerIds) -> Option<i64> {
@@ -163,9 +170,17 @@ impl<G: GraphViewOps> TimeSemantics for WindowedGraph<G> {
             .edge_earliest_time_window(e, self.t_start..self.t_end, layer_ids)
     }
 
-    fn edge_earliest_time_window(&self, e: EdgeRef, w: Range<i64>, layer_ids: LayerIds) -> Option<i64> {
-        self.graph
-            .edge_earliest_time_window(e, self.actual_start(w.start)..self.actual_end(w.end), layer_ids)
+    fn edge_earliest_time_window(
+        &self,
+        e: EdgeRef,
+        w: Range<i64>,
+        layer_ids: LayerIds,
+    ) -> Option<i64> {
+        self.graph.edge_earliest_time_window(
+            e,
+            self.actual_start(w.start)..self.actual_end(w.end),
+            layer_ids,
+        )
     }
 
     fn edge_latest_time(&self, e: EdgeRef, layer_ids: LayerIds) -> Option<i64> {
@@ -173,9 +188,17 @@ impl<G: GraphViewOps> TimeSemantics for WindowedGraph<G> {
             .edge_latest_time_window(e, self.t_start..self.t_end, layer_ids)
     }
 
-    fn edge_latest_time_window(&self, e: EdgeRef, w: Range<i64>, layer_ids: LayerIds) -> Option<i64> {
-        self.graph
-            .edge_latest_time_window(e, self.actual_start(w.start)..self.actual_end(w.end), layer_ids)
+    fn edge_latest_time_window(
+        &self,
+        e: EdgeRef,
+        w: Range<i64>,
+        layer_ids: LayerIds,
+    ) -> Option<i64> {
[…]
