### 2023-07-21 37fe10d
more cleanup but PropertiesIterable behaviour is inconsistent
