+++ python/tests/test_graphdb.py
@@ -474,112 +474,106 @@ def test_vertex_properties():
     time_property_test(1, "prop 2", None)
 
     # testing properties
-    assert g.vertex(1).properties() == {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
+    assert g.vertex(1).properties == {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
                                         'static prop': 123}
-    assert g.vertices.properties().collect() == [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
+    assert g.vertices.properties == [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
                                                   'static prop': 123}]
-    assert g.vertices.out_neighbours().properties().collect() == [[
+    assert g.vertices.out_neighbours().properties == [[
         {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
          'static prop': 123}]]
 
-    assert g.vertex(1).properties(include_static=False) == {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
+    assert g.vertex(1).properties.temporal.latest() == {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
                                                             'prop 4': True}
-    assert g.vertices.properties(include_static=False).collect() == [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
+    assert g.vertices.properties.temporal.latest() == [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
                                                                       'prop 4': True}]
-    assert g.vertices.out_neighbours().properties(include_static=False).collect() == [
+    assert g.vertices.out_neighbours().properties.temporal.latest() == [
         [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
           'prop 4': True}]]
 
-    assert g.at(2).vertex(1).properties() == {'prop 1': 2, 'prop 4': False, 'prop 2': 0.6, 'static prop': 123,
+    assert g.at(2).vertex(1).properties == {'prop 1': 2, 'prop 4': False, 'prop 2': 0.6, 'static prop': 123,
                                               'prop 3': 'hi'}
-    assert g.at(2).vertices.properties().collect() == [{'prop 1': 2, 'prop 4': False, 'prop 2': 0.6, 'static prop': 123,
+    assert g.at(2).vertices.properties == [{'prop 1': 2, 'prop 4': False, 'prop 2': 0.6, 'static prop': 123,
                                                         'prop 3': 'hi'}]
-    assert g.at(2).vertices.out_neighbours().properties().collect() == [
+    assert g.at(2).vertices.out_neighbours().properties == [
         [{'prop 1': 2, 'prop 4': False, 'prop 2': 0.6, 'static prop': 123,
           'prop 3': 'hi'}]]
 
     # testing property histories
-    assert g.vertex(1).property_histories() == {'prop 3': [(1, 'hi'), (3, 'hello')], 'prop 1': [(1, 1), (2, 2)],
+    assert g.vertex(1).properties.temporal == {'prop 3': [(1, 'hi'), (3, 'hello')], 'prop 1': [(1, 1), (2, 2)],
                                                 'prop 4': [(1, True), (2, False), (3, True)],
                                                 'prop 2': [(2, 0.6), (3, 0.9)]}
-    assert g.vertices.property_histories().collect() == [
+    assert g.vertices.properties.temporal == [
[…]
