### 2023-07-21 790c663
:art: connected components now uses a vertex view but the copying has gone through the roof
