### 2023-07-21 e516f14
start fixing inconsistencies for PropsIterable
