### 2023-07-21 52e6a76
update vertex test to reflect expected semantics
