+++ python/tests/test_graphdb.py
@@ -12,7 +12,6 @@ import tempfile
 from math import isclose
 import datetime
 
-
 edges = [
     (1, 1, 2),
     (2, 1, 3),
@@ -334,19 +333,20 @@ def test_graph_properties():
 
     # testing properties
     assert g.properties.as_dict() == {"prop 1": 1, "prop 2": "hi", "prop 3": True, "prop 4": 11, "prop 5": "world",
-                              "prop 6": True}
+                                      "prop 6": True}
 
     assert g.properties.temporal.latest() == {"prop 4": 11, "prop 5": "world",
-                                                  "prop 6": True}
-    assert g.at(2).properties.as_dict() == {"prop 1": 1, "prop 2": "hi", "prop 3": True, "prop 4": 11, "prop 5": "world",
-                                    "prop 6": True}
+                                              "prop 6": True}
+    assert g.at(2).properties.as_dict() == {"prop 1": 1, "prop 2": "hi", "prop 3": True, "prop 4": 11,
+                                            "prop 5": "world",
+                                            "prop 6": True}
 
     # testing property histories
     assert g.properties.temporal.histories() == {"prop 4": [(1, 11)], "prop 5": [(1, "world")],
-                                      "prop 6": [(1, False), (2, True)]}
+                                                 "prop 6": [(1, False), (2, True)]}
 
     assert g.at(2).properties.temporal.histories() == {"prop 4": [(1, 11)], "prop 5": [(1, "world")],
-                                            "prop 6": [(1, False), (2, True)]}
+                                                       "prop 6": [(1, False), (2, True)]}
 
     # testing property names
     expected_names = sorted(['prop 1', 'prop 2', 'prop 3', 'prop 4', 'prop 5', 'prop 6'])
@@ -475,49 +475,46 @@ def test_vertex_properties():
 
     # testing properties
     assert g.vertex(1).properties == {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
-                                        'static prop': 123}
-    assert g.vertices.properties == [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
-                                                  'static prop': 123}]
-    assert g.vertices.out_neighbours().properties == [[
-        {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2, 'prop 4': True,
-         'static prop': 123}]]
+                                      'static prop': 123}
+    assert g.vertices.properties == {'prop 2': [0.9], 'prop 3': ['hello'], 'prop 1': [2], 'prop 4': [True],
+                                     'static prop': [123]}
+    assert g.vertices.out_neighbours().properties == {'prop 2': [[0.9]], 'prop 3': [['hello']], 'prop 1': [[2]],
+                                                      'prop 4': [[True]], 'static prop': [[123]]}
 
     assert g.vertex(1).properties.temporal.latest() == {'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
-                                                            'prop 4': True}
-    assert g.vertices.properties.temporal.latest() == [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
-                                                                      'prop 4': True}]
-    assert g.vertices.out_neighbours().properties.temporal.latest() == [
-        [{'prop 2': 0.9, 'prop 3': 'hello', 'prop 1': 2,
-          'prop 4': True}]]
+                                                        'prop 4': True}
[…]
