+++ raphtory/src/lib.rs
@@ -107,8 +107,8 @@ pub mod prelude {
             api::{
                 mutation::{AdditionOps, DeletionOps, PropertyAdditionOps},
                 view::{
-                    EdgeListOps, EdgeViewOps, GraphViewOps, Layer, TimeOps, VertexListOps, LayerOps,
-                    VertexViewOps,
+                    EdgeListOps, EdgeViewOps, GraphViewOps, Layer, LayerOps, TimeOps,
+                    VertexListOps, VertexViewOps,
                 },
             },
             graph::graph::Graph,
