### 2023-07-21 6129959
improved __eq__ for properties
