+++ raphtory/src/python/graph/properties/props.rs
@@ -6,7 +6,7 @@ use crate::python::graph::properties::{
     DynProps, DynStaticProperties, DynTemporalProperties, NestedStaticPropsIterable,
     NestedTemporalPropsIterable, PyStaticProperties, StaticPropsIterable, TemporalPropsIterable,
 };
-use crate::python::types::iterable::Iterable;
+use crate::python::types::iterable::{Iterable, NestedIterable};
 use crate::python::types::repr::{iterator_dict_repr, Repr};
 use crate::python::types::wrappers::prop::PropValue;
 use crate::python::utils::PyGenericIterator;
@@ -196,19 +196,43 @@ impl Repr for PyProperties {
 }
 
 #[derive(PartialEq, Clone)]
-pub struct PyPropsIterableComparable(HashMap<String, Vec<Option<Prop>>>);
+pub struct PyPropsIterableComparable(HashMap<String, OptionPropIterCmp>);
 
 impl<'source> FromPyObject<'source> for PyPropsIterableComparable {
     fn extract(ob: &'source PyAny) -> PyResult<Self> {
         if let Ok(sp) = ob.extract::<PyRef<StaticPropsIterable>>() {
-            Ok(Self(sp.deref().as_dict()))
+            Ok(sp.deref().into())
         } else if let Ok(p) = ob.extract::<PyRef<PyPropsIterable>>() {
-            Ok(Self(p.deref().as_dict()))
-        } else if let Ok(m) = ob.extract::<HashMap<String, Vec<Option<Prop>>>>() {
+            Ok(p.deref().into())
+        } else if let Ok(m) = ob.extract::<HashMap<String, OptionPropIterCmp>>() {
             Ok(Self(m))
         } else {
             Err(PyTypeError::new_err("not comparable with properties"))
         }
+    }
+}
+
+impl From<&StaticPropsIterable> for PyPropsIterableComparable {
+    fn from(value: &StaticPropsIterable) -> Self {
+        Self(
+            value
+                .items()
+                .into_iter()
+                .map(|(k, v)| (k, v.into()))
+                .collect(),
+        )
+    }
+}
+
+impl From<&PyPropsIterable> for PyPropsIterableComparable {
+    fn from(value: &PyPropsIterable) -> Self {
+        Self(
+            value
+                .items()
+                .into_iter()
+                .map(|(k, v)| (k, v.into()))
+                .collect(),
+        )
     }
 }
 
@@ -232,9 +256,6 @@ where
         Self(Iterable::new("PropsIterable", value))
     }
 }
-
-// py_iterable!(PropsIterable, DynProperties, PyProperties);
-// py_iterable_comp!(PropsIterable, PropsComparable, PropsIterableComparable);
 
 #[pymethods]
 impl PyPropsIterable {
@@ -333,8 +354,8 @@ impl PyPropsIterable {
         match op {
             CompareOp::Lt => Err(PyTypeError::new_err("not ordered")),
             CompareOp::Le => Err(PyTypeError::new_err("not ordered")),
-            CompareOp::Eq => Ok(PyPropsIterableComparable(self.as_dict()) == other),
-            CompareOp::Ne => Ok(PyPropsIterableComparable(self.as_dict()) != other),
+            CompareOp::Eq => Ok(PyPropsIterableComparable::from(self) == other),
+            CompareOp::Ne => Ok(PyPropsIterableComparable::from(self) != other),
             CompareOp::Gt => Err(PyTypeError::new_err("not ordered")),
             CompareOp::Ge => Err(PyTypeError::new_err("not ordered")),
         }
[…]
