### 2023-07-21 52f81df
fix static properties
