+++ raphtory/src/python/graph/properties/static_props.rs
@@ -3,7 +3,11 @@ use crate::db::api::properties::internal::PropertiesOps;
 use crate::db::api::properties::StaticProperties;
 use crate::db::api::view::internal::Static;
 use crate::python::graph::properties::props::PropsComparable;
-use crate::python::graph::properties::{DynProps, NestedOptionPropIterable, OptionPropIterable};
+use crate::python::graph::properties::{
+    DynProps, NestedOptionPropIterable, OptionPropIterable, PyNestedPropsIterableComparable,
+    PyPropsIterableComparable,
+};
+use crate::python::types::iterable::{Iterable, NestedIterable};
 use crate::python::types::repr::Repr;
 use crate::python::utils::PyGenericIterator;
 use itertools::Itertools;
@@ -11,6 +15,7 @@ use pyo3::exceptions::{PyKeyError, PyTypeError};
 use pyo3::prelude::*;
 use pyo3::pyclass::CompareOp;
 use std::collections::HashMap;
+use std::ops::Deref;
 use std::sync::Arc;
 
 pub type DynStaticProperties = StaticProperties<DynProps>;
@@ -99,16 +104,37 @@ impl Repr for PyStaticProperties {
     }
 }
 
-py_iterable!(StaticPropsIterable, DynStaticProperties, PyStaticProperties);
+#[pyclass(name = "StaticPropertiesIterable")]
+pub struct PyStaticPropsIterable(Iterable<DynStaticProperties, PyStaticProperties>);
 
-py_iterable_comp!(
-    StaticPropsIterable,
-    PropsComparable,
-    StaticPropsIterComparable
-);
+impl Deref for PyStaticPropsIterable {
+    type Target = Iterable<DynStaticProperties, PyStaticProperties>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<F: Fn() -> It + Send + Sync + 'static, It: Iterator + Send + 'static> From<F>
+    for PyStaticPropsIterable
+where
+    It::Item: Into<DynStaticProperties>,
+{
+    fn from(value: F) -> Self {
+        Self(Iterable::new("StaticPropsIterable", value))
+    }
+}
+
+// py_iterable!(StaticPropsIterable, DynStaticProperties, PyStaticProperties);
+
+// py_iterable_comp!(
+//     StaticPropsIterable,
+//     PropsComparable,
+//     StaticPropsIterComparable
+// );
 
 #[pymethods]
-impl StaticPropsIterable {
+impl PyStaticPropsIterable {
     pub fn keys(&self) -> Vec<String> {
         self.iter().map(|p| p.keys()).kmerge().dedup().collect()
     }
@@ -143,6 +169,21 @@ impl StaticPropsIterable {
         self.iter().any(|p| p.contains(key))
     }
 
+    pub fn __iter__(&self) -> PyGenericIterator {
+        self.keys().into_iter().into()
+    }
+
+    pub fn __richcmp__(&self, other: PyPropsIterableComparable, op: CompareOp) -> PyResult<bool> {
+        match op {
+            CompareOp::Lt => Err(PyTypeError::new_err("not ordered")),
+            CompareOp::Le => Err(PyTypeError::new_err("not ordered")),
+            CompareOp::Eq => Ok(PyPropsIterableComparable::from(self) == other),
+            CompareOp::Ne => Ok(PyPropsIterableComparable::from(self) != other),
+            CompareOp::Gt => Err(PyTypeError::new_err("not ordered")),
+            CompareOp::Ge => Err(PyTypeError::new_err("not ordered")),
+        }
+    }
+
     pub fn as_dict(&self) -> HashMap<String, Vec<Option<Prop>>> {
         self.items()
             .into_iter()
[…]
