### 2023-07-21 26982a3
start implementing comparisons for temporal properties
