+++ raphtory/src/python/types/macros/mod.rs
@@ -2,3 +2,5 @@
 pub mod iterable;
 #[macro_use]
 pub mod nested_iterable;
+#[macro_use]
+pub mod cmp;
