+++ python/src/graphql.rs
@@ -3,39 +3,40 @@ use raphtory_core::{
     db::api::view::internal::DynamicGraph,
     python::{graph::graph::PyGraph, utils::errors::adapt_err_value},
 };
-use raphtory_graphql::RaphtoryServer;
-use std::collections::HashMap;
+use raphtory_graphql::{RaphtoryServer, server::TokioRuntime};
+use std::{
+    collections::HashMap,
+    sync::mpsc::Sender,
+};
 
 #[pyfunction]
 pub fn from_map(
     py: Python,
     graphs: HashMap<String, PyGraph>,
     port: Option<u16>,
-) -> PyResult<&PyAny> {
+) -> PyResult<ShutdownHandler> {
     let graphs: HashMap<String, DynamicGraph> = graphs
         .into_iter()
         .map(|(key, value)| (key, value.into()))
         .collect();
-    let server = RaphtoryServer::from_map(graphs);
+    let (server, receiver) = RaphtoryServer::from_map(graphs);
+    let mut handler = ShutdownHandler::new(server.sender());
     let port = port.unwrap_or(1736);
-    pyo3_asyncio::tokio::future_into_py(py, async move {
-        server
-            .run_with_port(port)
-            .await
-            .map_err(|e| adapt_err_value(&e))
-    })
+
+    let runtime = server.run_and_forget(port, Some(receiver));
+    handler.set_runtime(runtime);
+    Ok(handler)
 }
 
 #[pyfunction]
-pub fn from_directory(py: Python, path: String, port: Option<u16>) -> PyResult<&PyAny> {
-    let server = RaphtoryServer::from_directory(path.as_str());
+pub fn from_directory(py: Python, path: String, port: Option<u16>) -> PyResult<ShutdownHandler> {
+    let (server, receiver) = RaphtoryServer::from_directory(path.as_str());
     let port = port.unwrap_or(1736);
-    pyo3_asyncio::tokio::future_into_py(py, async move {
-        server
-            .run_with_port(port)
-            .await
-            .map_err(|e| adapt_err_value(&e))
-    })
+    let mut handler = ShutdownHandler::new(server.sender());
+
+    let runtime = server.run_and_forget(port, Some(receiver));
+    handler.set_runtime(runtime);
+    Ok(handler)
 }
 
 #[pyfunction]
@@ -44,17 +45,45 @@ pub fn from_map_and_directory(
     graphs: HashMap<String, PyGraph>,
     path: String,
     port: Option<u16>,
-) -> PyResult<&PyAny> {
+) -> PyResult<ShutdownHandler> {
     let graphs: HashMap<String, DynamicGraph> = graphs
         .into_iter()
         .map(|(key, value)| (key, value.into()))
         .collect();
     let port = port.unwrap_or(1736);
-    let server = RaphtoryServer::from_map_and_directory(graphs, path.as_str());
-    pyo3_asyncio::tokio::future_into_py(py, async move {
-        server
-            .run_with_port(port)
-            .await
-            .map_err(|e| adapt_err_value(&e))
-    })
+    let (server, receiver) = RaphtoryServer::from_map_and_directory(graphs, path.as_str());
+    let mut handler = ShutdownHandler::new(server.sender());
+    let runtime = server.run_and_forget(port, Some(receiver));
+    handler.set_runtime(runtime);
+    Ok(handler)
+}
+
+#[pyclass]
+pub struct ShutdownHandler {
+    sender: Sender<()>,
+    runtime: Option<TokioRuntime>,
+}
+
+#[pymethods]
+impl ShutdownHandler {
+    fn shutdown(&self) -> PyResult<()> {
+        println!("Shutting down server");
[…]
