### 2023-07-21 9de2be1
Added some tests and found some broken things that need to be made into issues
