+++ python/tests/test_graphdb.py
@@ -1516,3 +1516,150 @@ def load_from_pandas_into_existing_graph():
         vertices.append((v.id(), name))
 
     assert vertices == [(1, "Alice"), (2, "Bob"), (3, "Carol"), (4, "Dave"), (5, "Eve"), (6, "Frank")]
+
+
+def test_search_in_python():
+    g = Graph()
+    g.add_vertex(1,"hamza",properties={"value":60,"value_f":31.3,"value_str":"abc123"})
+    g.add_vertex(2,"ben",properties={"value":59,"value_f":11.4,"value_str":"test test test"})
+    g.add_vertex(3,"haaroon",properties={"value":199,"value_f":52.6,"value_str":"I wanna rock right now"})
+
+    g.add_edge(2,"haaroon","hamza",properties={"value":60,"value_f":31.3,"value_str":"abc123"})
+    g.add_edge(1,"ben","hamza",properties={"value":59,"value_f":11.4,"value_str":"test test test"})
+    g.add_edge(3,"ben","haaroon",properties={"value":199,"value_f":52.6,"value_str":"I wanna rock right now"})
+
+    index = g.index()
+
+    #Name tests
+    assert len(index.search_vertices("name:ben")) == 1
+    assert len(index.search_vertices("name:ben OR name:hamza")) == 2
+    assert len(index.search_vertices("name:ben AND name:hamza")) == 0
+    assert len(index.search_vertices("name: IN [ben, hamza]")) == 2
+
+    #Property tests
+    assert len(index.search_vertices("value:<120 OR value_f:>30")) == 3
+    assert len(index.search_vertices("value: [0 TO 60]")) == 2
+    assert len(index.search_vertices("value: [0 TO 60}")) == 1 # } == exclusive
+    assert len(index.search_vertices("value:>59 AND value_str:abc123")) == 1
+
+    #edge tests
+    assert len(index.search_edges("from:ben")) == 2
+    assert len(index.search_edges("from:ben OR from:haaroon")) == 3
+    assert len(index.search_edges("to:haaroon AND from:ben")) == 1
+    assert len(index.search_edges("to: IN [ben, hamza]")) == 2
+
+    #edge prop tests
+    assert len(index.search_edges("value:<120 OR value_f:>30")) == 3
+    assert len(index.search_edges("value: [0 TO 60]")) == 2
+    assert len(index.search_edges("value: [0 TO 60}")) == 1 # } == exclusive
+    assert len(index.search_edges("value:>59 AND value_str:abc123")) == 1
+
+    #Multiple history points test
+    g = Graph()
+    g.add_vertex(1,"hamza",properties={"value":60,"value_f":31.3,"value_str":"abc123"})
+    g.add_vertex(2,"hamza",properties={"value":70,"value_f":21.3,"value_str":"avc125"})
+    g.add_vertex(3,"hamza",properties={"value":80,"value_f":11.3,"value_str":"dsc2312"})
+
+    index = g.index()
+
+    #The semantics here are that the expressions independently need to evaluate at ANY point in the lifetime of the vertex - hence hamza is returned even though at no point does he have both these values at the same time
+    assert len(index.search_vertices("value:<70 AND value_f:<19.2")) == 1
+
+    g.add_vertex(4,"hamza",properties={"value":100,"value_f":11.3,"value_str":"dsc2312"})
+    # the graph isn't currently reindexed so this will not return hamza even though he now has a value which fits the bill
+    assert len(index.search_vertices("value:>99")) == 0
+
+
+def test_search_with_windows():
+
[…]
