### 2023-07-24 018c518
impl graph expanded edges
