+++ raphtory-graphql/src/model/graph/node.rs
@@ -1,3 +1,4 @@
+use crate::model::graph::get_expanded_edges;
 use crate::model::{
     filters::edgefilter::EdgeFilter,
     graph::{edge::Edge, property::Property, property_update::PropertyUpdate},
@@ -5,7 +6,6 @@ use crate::model::{
 use async_graphql::Context;
 use dynamic_graphql::{ResolvedObject, ResolvedObjectFields};
 use itertools::Itertools;
-use raphtory::db::graph::edge::EdgeView;
 use raphtory::{
     core::Prop,
     db::{
@@ -196,87 +196,18 @@ impl Node {
         graph_nodes: Vec<String>,
         filter: Option<EdgeFilter>,
     ) -> Vec<Edge> {
-        let get_expanded_edges = || -> Vec<Edge> {
-            let node_found_in_graph_nodes =
-                |node_name: String| -> bool { graph_nodes.iter().contains(&node_name) };
-
-            let mut fetched_edges: Vec<EdgeView<DynamicGraph>> = self
-                .vv
-                .edges()
-                .into_iter()
-                .map(|ee| ee.clone())
-                .collect_vec();
-
-            let first_hop_edges = fetched_edges
-                .clone()
-                .into_iter()
-                .filter(|e| {
-                    !node_found_in_graph_nodes((*e).src().name())
-                        || !node_found_in_graph_nodes((*e).dst().name())
-                })
-                .collect_vec();
-
-            let mut first_hop_nodes: HashSet<String> = HashSet::new();
-            first_hop_edges.clone().into_iter().for_each(|e| {
-                first_hop_nodes.insert(e.src().name());
-                first_hop_nodes.insert(e.dst().name());
-            });
-
-            let first_hop_nodes = first_hop_nodes
-                .into_iter()
-                .filter(|e| (*e).to_string() != *self.vv.name())
-                .collect_vec();
-
-            let node_found_in_first_hop_nodes =
-                |node_name: String| -> bool { first_hop_nodes.contains(&node_name) };
-
-            let mut first_hop_node_edges: Vec<EdgeView<DynamicGraph>> = vec![];
-
-            first_hop_edges.into_iter().for_each(|e| {
-                if node_found_in_graph_nodes(e.src().name()) {
-                    // Return only those edges whose either src or dst already exist
-                    let mut r = e
-                        .dst()
-                        .edges()
-                        .filter(|e| {
-                            (node_found_in_first_hop_nodes(e.src().name())
-                                && node_found_in_first_hop_nodes(e.dst().name()))
-                                || node_found_in_graph_nodes(e.src().name())
-                                || node_found_in_graph_nodes(e.dst().name())
-                        })
-                        .collect_vec();
-
-                    first_hop_node_edges.append(&mut r);
-                } else {
-                    let mut r = e
-                        .src()
-                        .edges()
-                        .filter(|e| {
-                            (node_found_in_first_hop_nodes(e.src().name())
-                                && node_found_in_first_hop_nodes(e.dst().name()))
-                                || node_found_in_graph_nodes(e.src().name())
-                                || node_found_in_graph_nodes(e.dst().name())
-                        })
-                        .collect_vec();
-
-                    first_hop_node_edges.append(&mut r);
-                }
-            });
-
-            fetched_edges.append(&mut first_hop_node_edges);
-
-            fetched_edges
-                .iter()
-                .map(|ee| ee.clone().into())
-                .collect_vec()
-        };
+        let all_graph_nodes: HashSet<String> = graph_nodes.into_iter().collect();
+        let fetched_edges = get_expanded_edges(all_graph_nodes, self.vv.clone())
+            .iter()
+            .map(|ee| ee.clone().into())
[…]

[…]
