### 2023-07-24 b1d6e08
almost working just got a borrow issue
