+++ raphtory/src/algorithms/motifs/three_node_local_redone.rs
@@ -3,7 +3,7 @@ use crate::{
     algorithms::{motifs::three_node_motifs::*, k_core::k_core_set},
     core::state::{accumulator_id::{accumulators::{self, val}, AccId}, compute_state::ComputeStateVec, agg::ValDef},
     db::{
-        api::view::{GraphViewOps, VertexViewOps, *},
+        api::view::{GraphViewOps, VertexViewOps, *, internal::CoreGraphOps},
         task::{
             context::Context,
             task::{ATask, Job, Step},
@@ -13,9 +13,10 @@ use crate::{
     },
 };
 
+use itertools::enumerate;
 use num_traits::Zero;
 use rustc_hash::FxHashSet;
-use std::{collections::HashMap, ops::Add, slice::Iter};
+use std::{collections::{HashMap, HashSet}, ops::Add, slice::Iter};
 ///////////////////////////////////////////////////////
 
 // State objects for three node motifs
@@ -159,19 +160,18 @@ pub fn twonode_motif_count<G: GraphViewOps>(
 
 ///////////////////////////////////////////////////////
 
-pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, threads: Option<usize>) {
+pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, threads: Option<usize>) -> HashMap<String, [usize;8]>{
     let vertex_set = k_core_set(graph,2,usize::MAX, None);
     let g = graph.subgraph(vertex_set);
     let mut ctx: Context<VertexSubgraph<G>, ComputeStateVec> = Context::from(&g);
-
     let motifs_counter = val::<MotifCounter>(0);
     ctx.agg(motifs_counter);
 
-    let neighbours_set = accumulators::hash_set::<u64>(0);
+    let neighbours_set = accumulators::hash_set::<u64>(1);
 
     ctx.agg(neighbours_set);
 
-    let step1 = ATask::new(move |u| {
+    let step1 = ATask::new(move |u: &mut EvalVertexView<VertexSubgraph<G>, ComputeStateVec, MotifCounter>| {
         for v in u.neighbours() {
             if u.id() > v.id() {
                 v.update(&neighbours_set, u.id());
@@ -180,15 +180,12 @@ pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, threads: Option<us
         Step::Continue
     });
 
-    let step2 = ATask::new(move |u| {
+    let step2 = ATask::new(move |u : &mut EvalVertexView<VertexSubgraph<G>, ComputeStateVec, MotifCounter>| {
         for v in u.neighbours() {
 
             // Find triangles on the UV edge
             if u.id() > v.id() {
                 let intersection_nbs = {
-                    // when using entry() we need to make sure the reference is released before we can update the state, otherwise we break the Rc<RefCell<_>> invariant
-                    // where there can either be one mutable or many immutable references
-
                     match (
                         u.entry(&neighbours_set)
                             .read_ref()
@@ -198,19 +195,20 @@ pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, threads: Option<us
                             .unwrap_or(&FxHashSet::default()),
                     ) {
                         (u_set, v_set) => {
-                            let intersection = u_set.intersection(v_set);
+                            let intersection = u_set.intersection(v_set).cloned().collect::<Vec<_>>();
+                            // println!("{:?}",intersection.len());
                             intersection
[…]
