### 2023-07-24 ff6efc2
finish implementation for temporal properties and update all python tests
