+++ python/tests/test_graphdb.py
@@ -378,7 +378,7 @@ def test_vertex_properties():
 
     g.add_vertex_properties(1, {"static prop": 123})
 
-    # testing property_history
+    # testing property history
     def history_test(key, value):
         if value is None:
             assert g.vertex(1).properties.temporal.get(key) is None
@@ -527,9 +527,10 @@ def test_vertex_properties():
     assert sorted(g.vertices.properties.temporal.keys()) == expected_names_no_static
     assert sorted(g.vertices.out_neighbours().properties.temporal.keys()) == expected_names_no_static
 
-    assert sorted(g.at(1).vertex(1).properties.temporal.keys()) == expected_names_no_static
-    assert sorted(g.at(1).vertices.properties.temporal.keys()) == expected_names_no_static
-    assert sorted(g.at(1).vertices.out_neighbours().properties.temporal.keys()) == expected_names_no_static
+    expected_names_no_static_at_1 = sorted(['prop 4', 'prop 1', 'prop 3'])
+    assert sorted(g.at(1).vertex(1).properties.temporal.keys()) == expected_names_no_static_at_1
+    assert sorted(g.at(1).vertices.properties.temporal.keys()) == expected_names_no_static_at_1
+    assert sorted(g.at(1).vertices.out_neighbours().properties.temporal.keys()) == expected_names_no_static_at_1
 
     # testing has_property
     assert "prop 4" in g.vertex(1).properties
@@ -584,66 +585,66 @@ def test_edge_properties():
 
     g.add_edge_properties(1, 2, {"static prop": 123})
 
-    # testing property_history
-    assert g.edge(1, 2).property_history("prop 1") == [(1, 1), (2, 2)]
-    assert g.edge(1, 2).property_history("prop 2") == [(2, 0.6), (3, 0.9)]
-    assert g.edge(1, 2).property_history("prop 3") == [(1, "hi"), (3, 'hello')]
-    assert g.edge(1, 2).property_history("prop 4") == [(1, True), (2, False), (3, True)]
-    assert g.edge(1, 2).property_history("undefined") == []
-    assert g.at(1).edge(1, 2).property_history("prop 4") == [(1, True)]
-    assert g.at(1).edge(1, 2).property_history("static prop") == []
+    # testing property history
+    assert g.edge(1, 2).properties.temporal.get("prop 1") == [(1, 1), (2, 2)]
+    assert g.edge(1, 2).properties.temporal.get("prop 2") == [(2, 0.6), (3, 0.9)]
+    assert g.edge(1, 2).properties.temporal.get("prop 3") == [(1, "hi"), (3, 'hello')]
+    assert g.edge(1, 2).properties.temporal.get("prop 4") == [(1, True), (2, False), (3, True)]
+    assert g.edge(1, 2).properties.temporal.get("undefined") is None
+    assert g.at(1).edge(1, 2).properties.temporal.get("prop 4") == [(1, True)]
+    assert g.at(1).edge(1, 2).properties.temporal.get("static prop") is None
 
-    assert g.at(1).edge(1, 2).static_property("static prop") == 123
-    assert g.at(100).edge(1, 2).static_property("static prop") == 123
-    assert g.edge(1, 2).static_property("static prop") == 123
-    assert g.edge(1, 2).static_property("prop 4") is None
+    assert g.at(1).edge(1, 2).properties.meta.get("static prop") == 123
+    assert g.at(100).edge(1, 2).properties.meta.get("static prop") == 123
+    assert g.edge(1, 2).properties.meta.get("static prop") == 123
[…]
