+++ raphtory/src/python/graph/properties/temporal_props.rs
@@ -42,9 +42,15 @@ impl<P: Into<DynTemporalProperties>> From<P> for PyTemporalProperties {
 }
 
 #[derive(PartialEq)]
-pub struct TemporalPropsCmp(HashMap<String, PyTemporalPropertyViewCmp>);
+pub struct PyTemporalPropsCmp(HashMap<String, PyTemporalPropertyViewCmp>);
 
-impl From<&PyTemporalProperties> for TemporalPropsCmp {
+impl From<HashMap<String, PyTemporalPropertyViewCmp>> for PyTemporalPropsCmp {
+    fn from(value: HashMap<String, PyTemporalPropertyViewCmp>) -> Self {
+        Self(value)
+    }
+}
+
+impl From<&PyTemporalProperties> for PyTemporalPropsCmp {
     fn from(value: &PyTemporalProperties) -> Self {
         Self(
             value
@@ -56,9 +62,15 @@ impl From<&PyTemporalProperties> for TemporalPropsCmp {
     }
 }
 
-impl FromPyObject for TemporalPropsCmp {
+impl<'source> FromPyObject<'source> for PyTemporalPropsCmp {
     fn extract(ob: &'source PyAny) -> PyResult<Self> {
-        todo!()
+        if let Ok(v) = ob.extract::<PyRef<PyTemporalProperties>>() {
+            Ok(PyTemporalPropsCmp::from(v.deref()))
+        } else if let Ok(v) = ob.extract::<HashMap<String, PyTemporalPropertyViewCmp>>() {
+            Ok(PyTemporalPropsCmp::from(v))
+        } else {
+            Err(PyTypeError::new_err("cannot compare"))
+        }
     }
 }
 
@@ -66,6 +78,8 @@ impl FromPyObject for TemporalPropsCmp {
 pub struct PyTemporalProperties {
     props: DynTemporalProperties,
 }
+
+py_eq!(PyTemporalProperties, PyTemporalPropsCmp);
 
 #[pymethods]
 impl PyTemporalProperties {
@@ -147,6 +161,12 @@ impl From<&PyTemporalPropertyView> for PyTemporalPropertyViewCmp {
 impl From<Vec<(i64, Prop)>> for PyTemporalPropertyViewCmp {
     fn from(value: Vec<(i64, Prop)>) -> Self {
         Self(value)
+    }
+}
+
+impl From<DynTemporalProperty> for PyTemporalPropertyViewCmp {
+    fn from(value: DynTemporalProperty) -> Self {
+        PyTemporalPropertyViewCmp(value.iter().collect())
     }
 }
 
@@ -238,11 +258,46 @@ impl<P: PropertiesOps + Send + Sync + 'static> IntoPy<PyObject> for TemporalProp
     }
 }
 
-py_iterable!(
+py_iterable_base!(
     TemporalPropsIterable,
     DynTemporalProperties,
     PyTemporalProperties
 );
+
+#[derive(PartialEq)]
+pub struct TemporalPropsIterableCmp(HashMap<String, TemporalPropertyIterableCmp>);
+
+impl From<&TemporalPropsIterable> for TemporalPropsIterableCmp {
+    fn from(value: &TemporalPropsIterable) -> Self {
+        Self(
+            value
+                .items()
+                .into_iter()
+                .map(|(k, v)| (k, v.into()))
+                .collect(),
+        )
+    }
+}
+
+impl From<HashMap<String, TemporalPropertyIterableCmp>> for TemporalPropsIterableCmp {
+    fn from(value: HashMap<String, TemporalPropertyIterableCmp>) -> Self {
+        Self(value)
+    }
+}
+
+impl<'source> FromPyObject<'source> for TemporalPropsIterableCmp {
[…]
