### 2023-07-24 a2a09cd
fix quickcheck test (properties at t0 still have a latest value at t1)
