### 2023-07-24 c3a85fb
Fix merge issues and warnings
