### 2023-07-24 8133541
use lazy iterables everywhere in temporal properties and start cleaning up internal names
