+++ raphtory/src/python/graph/properties/props.rs
@@ -6,9 +6,8 @@ use crate::{
     },
     python::{
         graph::properties::{
-            DynProps, DynStaticProperties, DynTemporalProperties, NestedStaticPropsIterable,
-            NestedTemporalPropsIterable, PyStaticProperties, PyStaticPropsIterable,
-            TemporalPropsIterable,
+            DynProps, DynStaticProperties, DynTemporalProperties, PyMetaProps, PyMetaPropsList,
+            PyMetaPropsListList, PyTemporalPropsList, PyTemporalPropsListList,
         },
         types::{
             repr::{iterator_dict_repr, Repr},
@@ -27,41 +26,41 @@ use std::{collections::HashMap, ops::Deref, sync::Arc};
 pub type DynProperties = Properties<Arc<dyn PropertiesOps + Send + Sync>>;
 
 #[derive(PartialEq, Clone)]
-pub struct PropsComparable(HashMap<String, Prop>);
+pub struct PropsComp(HashMap<String, Prop>);
 
-impl<'source> FromPyObject<'source> for PropsComparable {
+impl<'source> FromPyObject<'source> for PropsComp {
     fn extract(ob: &'source PyAny) -> PyResult<Self> {
-        if let Ok(sp) = ob.extract::<PyRef<PyStaticProperties>>() {
+        if let Ok(sp) = ob.extract::<PyRef<PyMetaProps>>() {
             Ok(sp.deref().into())
         } else if let Ok(p) = ob.extract::<PyRef<PyProperties>>() {
             Ok(p.deref().into())
         } else if let Ok(m) = ob.extract::<HashMap<String, Prop>>() {
-            Ok(PropsComparable(m))
+            Ok(PropsComp(m))
         } else {
             Err(PyTypeError::new_err("not comparable with properties"))
         }
     }
 }
 
-impl From<&PyStaticProperties> for PropsComparable {
-    fn from(value: &PyStaticProperties) -> Self {
+impl From<&PyMetaProps> for PropsComp {
+    fn from(value: &PyMetaProps) -> Self {
         Self(value.as_dict())
     }
 }
 
-impl From<&PyProperties> for PropsComparable {
+impl From<&PyProperties> for PropsComp {
     fn from(value: &PyProperties) -> Self {
         Self(value.as_dict())
     }
 }
 
-impl From<DynStaticProperties> for PropsComparable {
+impl From<DynStaticProperties> for PropsComp {
     fn from(value: DynStaticProperties) -> Self {
         Self(value.as_map())
     }
 }
 
-impl From<DynProperties> for PropsComparable {
+impl From<DynProperties> for PropsComp {
     fn from(value: DynProperties) -> Self {
         Self(value.as_map())
     }
@@ -72,7 +71,7 @@ pub struct PyProperties {
     props: DynProperties,
 }
 
-py_eq!(PyProperties, PropsComparable);
+py_eq!(PyProperties, PropsComp);
 
 #[pymethods]
 impl PyProperties {
@@ -193,15 +192,15 @@ impl Repr for PyProperties {
 }
 
 #[derive(PartialEq, Clone)]
-pub struct PyPropsIterableComparable(HashMap<String, OptionPropIterCmp>);
+pub struct PyPropsListCmp(HashMap<String, PyPropValueListCmp>);
 
-impl<'source> FromPyObject<'source> for PyPropsIterableComparable {
+impl<'source> FromPyObject<'source> for PyPropsListCmp {
     fn extract(ob: &'source PyAny) -> PyResult<Self> {
-        if let Ok(sp) = ob.extract::<PyRef<PyStaticPropsIterable>>() {
+        if let Ok(sp) = ob.extract::<PyRef<PyMetaPropsList>>() {
             Ok(sp.deref().into())
[…]
