+++ raphtory/src/python/graph/properties/temporal_props.rs
@@ -5,8 +5,11 @@ use crate::{
         view::internal::{DynamicGraph, Static},
     },
     python::{
-        graph::properties::{DynProps, NestedOptionPropIterable, OptionPropIterable},
-        types::repr::{iterator_dict_repr, iterator_repr, Repr},
+        graph::properties::{DynProps, PyPropValueList, PyPropValueListList},
+        types::{
+            repr::{iterator_dict_repr, iterator_repr, Repr},
+            wrappers::prop::PropHistItems,
+        },
         utils::{PyGenericIterator, PyTime},
     },
 };
@@ -44,10 +47,10 @@ impl<P: Into<DynTemporalProperties>> From<P> for PyTemporalProperties {
 }
 
 #[derive(PartialEq)]
-pub struct PyTemporalPropsCmp(HashMap<String, PyTemporalPropertyViewCmp>);
+pub struct PyTemporalPropsCmp(HashMap<String, PyTemporalPropCmp>);
 
-impl From<HashMap<String, PyTemporalPropertyViewCmp>> for PyTemporalPropsCmp {
-    fn from(value: HashMap<String, PyTemporalPropertyViewCmp>) -> Self {
+impl From<HashMap<String, PyTemporalPropCmp>> for PyTemporalPropsCmp {
+    fn from(value: HashMap<String, PyTemporalPropCmp>) -> Self {
         Self(value)
     }
 }
@@ -68,7 +71,7 @@ impl<'source> FromPyObject<'source> for PyTemporalPropsCmp {
     fn extract(ob: &'source PyAny) -> PyResult<Self> {
         if let Ok(v) = ob.extract::<PyRef<PyTemporalProperties>>() {
             Ok(PyTemporalPropsCmp::from(v.deref()))
-        } else if let Ok(v) = ob.extract::<HashMap<String, PyTemporalPropertyViewCmp>>() {
+        } else if let Ok(v) = ob.extract::<HashMap<String, PyTemporalPropCmp>>() {
             Ok(PyTemporalPropsCmp::from(v))
         } else {
             Err(PyTypeError::new_err("cannot compare"))
@@ -135,16 +138,16 @@ impl PyTemporalProperties {
 }
 
 #[pyclass(name = "Property")]
-pub struct PyTemporalPropertyView {
+pub struct PyTemporalProp {
     prop: DynTemporalProperty,
 }
 
 #[derive(PartialEq, Clone)]
-pub struct PyTemporalPropertyViewCmp(Vec<(i64, Prop)>);
+pub struct PyTemporalPropCmp(Vec<(i64, Prop)>);
 
-impl<'source> FromPyObject<'source> for PyTemporalPropertyViewCmp {
+impl<'source> FromPyObject<'source> for PyTemporalPropCmp {
     fn extract(ob: &'source PyAny) -> PyResult<Self> {
-        if let Ok(sp) = ob.extract::<PyRef<PyTemporalPropertyView>>() {
+        if let Ok(sp) = ob.extract::<PyRef<PyTemporalProp>>() {
             Ok(sp.deref().into())
         } else if let Ok(m) = ob.extract::<Vec<(i64, Prop)>>() {
             Ok(Self(m))
@@ -154,28 +157,28 @@ impl<'source> FromPyObject<'source> for PyTemporalPropertyViewCmp {
     }
 }
 
-impl From<&PyTemporalPropertyView> for PyTemporalPropertyViewCmp {
-    fn from(value: &PyTemporalPropertyView) -> Self {
+impl From<&PyTemporalProp> for PyTemporalPropCmp {
+    fn from(value: &PyTemporalProp) -> Self {
         Self(value.items())
     }
[…]
