### 2023-07-24 708fbcf
rename static/meta to constant
