### 2023-07-24 e7788c7
fix issue with edges filtering
