+++ raphtory/src/algorithms/motifs/three_node_local_redone.rs
@@ -73,27 +73,28 @@ impl Zero for MotifCounter {
 
 ///////////////////////////////////////////////////////
 
-fn update_counter<G: GraphViewOps>(
-    vs: Vec<&EvalVertexView<G, ComputeStateVec, MotifCounter>>,
-    motif_counter: AccId<MotifCounter, MotifCounter, MotifCounter, ValDef<MotifCounter>>,
-    tmp_counts: Iter<usize>,
-) {
-    for v in vs {
-        let mc = v.read(&motif_counter);
-        let triangle: [usize; 8] = mc
-            .triangle
-            .iter()
-            .zip(tmp_counts.clone())
-            .map(|(&i1, &i2)| i1 + i2)
-            .collect::<Vec<usize>>()
-            .try_into()
-            .unwrap();
-        v.update(
-            &motif_counter,
-            MotifCounter::from_triangle_counter(triangle),
-        );
-    }
-}
+// fn update_counter<G: GraphViewOps>(
+//     vs: &mut Vec<&EvalVertexView<G, ComputeStateVec, MotifCounter>>,
+//     motif_counter: AccId<MotifCounter, MotifCounter, MotifCounter, ValDef<MotifCounter>>,
+//     tmp_counts: Iter<usize>,
+// ) {
+//     for mut v in vs {
+//         let mc = v.get_mut();
+//         let triangle: [usize; 8] = mc
+//             .triangle
+//             .iter()
+//             .zip(tmp_counts.clone())
+//             .map(|(&i1, &i2)| i1 + i2)
+//             .collect::<Vec<usize>>()
+//             .try_into()
+//             .unwrap();
+//         v.update(
+//             &motif_counter,
+//             MotifCounter::from_triangle_counter(triangle),
+//         );
+//         println!("{:?}",v.read(&motif_counter))
+//     }
+// }
 
 ///////////////////////////////////////////////////////
 
@@ -160,15 +161,15 @@ pub fn twonode_motif_count<G: GraphViewOps>(
 
 ///////////////////////////////////////////////////////
 
-pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, threads: Option<usize>) -> HashMap<String, [usize;8]>{
+pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, motifs_count_id:AccId<MotifCounter, MotifCounter,MotifCounter,ValDef<MotifCounter>>, threads: Option<usize>) -> HashMap<String, [usize;8]>{
     let vertex_set = k_core_set(graph,2,usize::MAX, None);
     // vertex_set.sort();
     // let vertex_set
     let g: VertexSubgraph<G> = graph.subgraph(vertex_set);
     let vertex_set_sorted = g.vertices().id().sorted().collect_vec();
     let mut ctx: Context<VertexSubgraph<G>, ComputeStateVec> = Context::from(&g);
-    let motifs_counter = val::<MotifCounter>(0);
-    ctx.agg(motifs_counter);
+
+    // ctx.agg(motifs_count_id);
 
     let neighbours_set = accumulators::hash_set::<u64>(1);
 
@@ -184,7 +185,7 @@ pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, threads: Option<us
     });
 
     let step2 = ATask::new(move |u : &mut EvalVertexView<VertexSubgraph<G>, ComputeStateVec, MotifCounter>| {
-        for v in u.neighbours() {
+        for mut v in u.neighbours() {
 
             // Find triangles on the UV edge
             if u.id() > v.id() {
[…]
