### 2023-07-24 3705890
add Rust Benchmark (cargo) benchmark result for 543b733cfffc581c9ddc0c84cfe2fab360fc52a7
