+++ raphtory/src/algorithms/motifs/three_node_local.rs
@@ -359,6 +359,8 @@ mod local_motif_test {
         // FIXME: Should test this
         let _global_counts = global_temporal_three_node_motifs(&graph, 10);
 
+        println!("{:?}",_global_counts);
+
         let expected: HashMap<u64, Vec<usize>> = HashMap::from([
             (
                 1,
