### 2023-07-24 be11be8
minor fixes and renames
