### 2023-07-24 6fbde1c
add hits to python and remove window (#1130)
* add hits to python and remove window

* :rocket: clean final raphtory warning

* Update lib.rs
