### 2023-07-24 880b0d6
add hits to python and remove window (#1130)
* add hits to python and remove window

* :rocket: clean final raphtory warning

* Update lib.rs
