### 2023-07-24 20ee935
add Rust Benchmark (cargo) benchmark result for 880b0d691d1abbdf42147b999be2137e9679796a
