### 2023-07-25 2eab238
Feature/expanded edges (#1132)
* impl expanded edges

* impl graph expanded edges

* cargo fmt

* fix issue with edges filtering

* fix typo
