+++ raphtory-graphql/src/model/graph/graph.rs
@@ -1,17 +1,24 @@
-use std::ops::Deref;
+use std::{
+    collections::{HashMap, HashSet},
+    ops::Deref,
+};
 
 use crate::model::{
     algorithm::Algorithms,
     filters::{edgefilter::EdgeFilter, nodefilter::NodeFilter},
-    graph::{edge::Edge, node::Node, property::Property},
+    graph::{edge::Edge, get_expanded_edges, node::Node, property::Property},
 };
 use dynamic_graphql::{ResolvedObject, ResolvedObjectFields};
 use itertools::Itertools;
 use raphtory::{
-    db::api::view::{
-        internal::{DynamicGraph, IntoDynamic},
-        GraphViewOps, TimeOps, VertexViewOps,
+    db::{
+        api::view::{
+            internal::{DynamicGraph, IntoDynamic},
+            GraphViewOps, TimeOps, VertexViewOps,
+        },
+        graph::edge::EdgeView,
     },
+    prelude::EdgeViewOps,
     search::IndexedGraph,
 };
 
@@ -125,6 +132,58 @@ impl GqlGraph {
                 .filter(|ev| filter.matches(ev))
                 .collect(),
             None => self.graph.edges().into_iter().map(|ev| ev.into()).collect(),
+        }
+    }
+
+    async fn expanded_edges(
+        &self,
+        nodes_to_expand: Vec<String>,
+        graph_nodes: Vec<String>,
+        filter: Option<EdgeFilter>,
+    ) -> Vec<Edge> {
+        if nodes_to_expand.is_empty() {
+            return vec![];
+        }
+
+        let nodes: Vec<Node> = self
+            .graph
+            .vertices()
+            .iter()
+            .map(|vv| vv.into())
+            .filter(|n| NodeFilter::new(nodes_to_expand.clone()).matches(n))
+            .collect();
+
+        let mut all_graph_nodes: HashSet<String> = graph_nodes.into_iter().collect();
+        let mut all_expanded_edges: HashMap<String, EdgeView<DynamicGraph>> = HashMap::new();
+
+        let mut maybe_layers: Option<Vec<String>> = None;
+        if filter.is_some() {
+            maybe_layers = filter.clone().unwrap().layer_names.map(|l| l.contains);
+        }
+
+        for node in nodes {
+            let expanded_edges =
+                get_expanded_edges(all_graph_nodes.clone(), node.vv, maybe_layers.clone());
+            expanded_edges.clone().into_iter().for_each(|e| {
+                let src = e.src().name();
+                let dst = e.dst().name();
+                all_expanded_edges.insert(src.to_owned() + &dst, e);
+                all_graph_nodes.insert(src);
+                all_graph_nodes.insert(dst);
+            });
+        }
+
+        let fetched_edges = all_expanded_edges
+            .values()
+            .map(|ee| ee.clone().into())
+            .collect_vec();
+
+        match filter {
+            Some(filter) => fetched_edges
+                .into_iter()
+                .filter(|ev| filter.matches(ev))
+                .collect(),
+            None => fetched_edges,
         }
     }
 
