### 2023-07-25 084b1ab
add Rust Benchmark (cargo) benchmark result for 2eab2389b2cd9ca8f37ec7bf45730e553b8277cf
