### 2023-07-25 bf4e49d
add test to confirm materialize is broken
