### 2023-07-25 e325a2b
more improvements to api
