### 2023-07-25 22ecb29
add test to confirm static properties are broken
