+++ raphtory/src/db/api/view/layer.rs
@@ -18,30 +18,30 @@ pub enum Layer {
     Multiple(Arc<[String]>),
 }
 
-impl<'a> From<Option<&'a str>> for Layer {
-    fn from(name: Option<&'a str>) -> Self {
+impl<'a, T: ToOwned<Owned = String> + ?Sized> From<Option<&'a T>> for Layer {
+    fn from(name: Option<&'a T>) -> Self {
         match name {
-            Some(name) => Layer::One(name.to_string()),
+            Some(name) => Layer::One(name.to_owned()),
             None => Layer::All,
         }
     }
 }
 
-impl<'a> From<&'a str> for Layer {
-    fn from(name: &'a str) -> Self {
-        Layer::One(name.to_string())
+impl<'a, T: ToOwned<Owned = String> + ?Sized> From<&'a T> for Layer {
+    fn from(name: &'a T) -> Self {
+        Layer::One(name.to_owned())
     }
 }
 
-impl<'a> From<Vec<&'a str>> for Layer {
-    fn from(names: Vec<&'a str>) -> Self {
+impl<'a, T: ToOwned<Owned = String> + ?Sized> From<Vec<&'a T>> for Layer {
+    fn from(names: Vec<&'a T>) -> Self {
         match names.len() {
             0 => Layer::All,
-            1 => Layer::One(names[0].to_string()),
+            1 => Layer::One(names[0].to_owned()),
             _ => Layer::Multiple(
                 names
                     .into_iter()
-                    .map(|s| s.to_string())
+                    .map(|s| s.to_owned())
                     .collect::<Vec<_>>()
                     .into(),
             ),
