### 2023-07-25 caccae8
start adding some python docs
