+++ raphtory-graphql/src/model/graph/node.rs
@@ -197,34 +197,26 @@ impl Node {
         graph_nodes: Vec<String>,
         filter: Option<EdgeFilter>,
     ) -> Vec<Edge> {
-        let mut res = vec![];
         let all_graph_nodes: HashSet<String> = graph_nodes.into_iter().collect();
 
         match filter {
-            Some(edgeFilter) => {
-                let maybe_layers = edgeFilter.clone().layer_names.map(|l| l.contains);
+            Some(edge_filter) => {
+                let maybe_layers = edge_filter.clone().layer_names.map(|l| l.contains);
                 let fetched_edges =
                     get_expanded_edges(all_graph_nodes, self.vv.clone(), maybe_layers)
                         .iter()
                         .map(|ee| ee.clone().into())
                         .collect_vec();
-
-                res = fetched_edges
+                fetched_edges
                     .into_iter()
-                    .filter(|ev| edgeFilter.matches(ev))
-                    .collect();
+                    .filter(|ev| edge_filter.matches(ev))
+                    .collect()
             }
-            None => {
-                let fetched_edges = get_expanded_edges(all_graph_nodes, self.vv.clone(), None)
-                    .iter()
-                    .map(|ee| ee.clone().into())
-                    .collect_vec();
-
-                res = fetched_edges;
-            }
+            None => get_expanded_edges(all_graph_nodes, self.vv.clone(), None)
+                .iter()
+                .map(|ee| ee.clone().into())
+                .collect_vec(),
         }
-
-        return res;
     }
 
     async fn exploded_edges(&self, layer: Option<String>) -> Vec<Edge> {
