+++ docs/requirements.txt
@@ -8,7 +8,7 @@ Babel==2.12.1
 backcall==0.2.0
 beautifulsoup4==4.12.2
 bleach==6.0.0
-certifi==2023.5.7
+certifi==2023.7.22
 charset-normalizer==3.1.0
 contourpy==1.1.0
 cycler==0.11.0
