### 2023-07-26 e702d1f
doc cleanup and make signatures work
