+++ raphtory/src/db/task/vertex/window_eval_vertex.rs
@@ -3,21 +3,25 @@ use crate::{
         entities::VID,
         state::{accumulator_id::AccId, agg::Accumulator, compute_state::ComputeState, StateType},
         utils::time::IntoTime,
-        Direction, Prop,
+        Direction,
     },
     db::{
-        api::view::{
-            internal::{GraphPropertiesOps, GraphWindowOps},
-            GraphViewOps, TimeOps, VertexListOps, VertexViewOps,
+        api::{
+            properties::Properties,
+            view::{internal::GraphWindowOps, GraphViewOps, TimeOps, VertexListOps, VertexViewOps},
         },
-        graph::path::{Operations, PathFromVertex},
+        graph::{
+            path::{Operations, PathFromVertex},
+            vertex::VertexView,
+            views::window_graph::WindowedGraph,
+        },
         task::{
             edge::window_eval_edge::WindowEvalEdgeView, task_state::Local2,
             vertex::eval_vertex_state::EVState,
         },
     },
 };
-use std::{cell::RefCell, collections::HashMap, marker::PhantomData, rc::Rc};
+use std::{cell::RefCell, marker::PhantomData, rc::Rc};
 
 pub struct WindowEvalVertex<'a, G: GraphViewOps, CS: ComputeState, S: 'static> {
     ss: usize,
@@ -83,7 +87,7 @@ impl<'a, G: GraphViewOps, CS: ComputeState, S: 'static> TimeOps for WindowEvalVe
     fn window<T: IntoTime>(&self, t_start: T, t_end: T) -> Self::WindowedViewType {
         WindowEvalVertex {
             ss: self.ss,
-            vertex: self.vertex.clone(),
+            vertex: self.vertex,
             graph: self.graph,
             _local_state: None,
             local_state_prev: self.local_state_prev,
@@ -97,7 +101,7 @@ impl<'a, G: GraphViewOps, CS: ComputeState, S: 'static> TimeOps for WindowEvalVe
 impl<'a, G: GraphViewOps, CS: ComputeState, S: 'static> VertexViewOps
     for WindowEvalVertex<'a, G, CS, S>
 {
-    type Graph = G;
+    type Graph = WindowedGraph<G>;
     type ValueType<T> = T;
     type PathType<'b> = WindowEvalPathFromVertex<'a, G, CS, S> where Self: 'b;
     type EList = Box<dyn Iterator<Item = WindowEvalEdgeView<'a, G, CS, S>> + 'a>;
@@ -120,90 +124,17 @@ impl<'a, G: GraphViewOps, CS: ComputeState, S: 'static> VertexViewOps
             .vertex_latest_time_window(self.vertex, self.t_start, self.t_end)
     }
 
-    fn property(
-        &self,
-        name: String,
-        include_static: bool,
-    ) -> Self::ValueType<Option<crate::core::Prop>> {
-        let props = self.property_history(name.clone());
-        match props.last() {
-            None => {
-                if include_static {
-                    self.graph.static_vertex_prop(self.vertex, &name)
-                } else {
-                    None
-                }
-            }
-            Some((_, prop)) => Some(prop.clone()),
-        }
-    }
-
     fn history(&self) -> Self::ValueType<Vec<i64>> {
         self.graph
             .vertex_history_window(self.vertex, self.t_start..self.t_end)
     }
 
-    fn property_history(&self, name: String) -> Self::ValueType<Vec<(i64, crate::core::Prop)>> {
-        self.graph
-            .temporal_vertex_prop_vec_window(self.vertex, &name, self.t_start, self.t_end)
-    }
-
-    fn properties(
-        &self,
-        include_static: bool,
-    ) -> Self::ValueType<std::collections::HashMap<String, crate::core::Prop>> {
[…]
