### 2023-07-26 16bf231
Properties (#1103)
* cleanup

* changes to rebase

* skeleton for properties objects

* Change properties API to enable dynamic boxing and start implementing some python wrappers

* compiles but still work to do

* add properties object for edges and some other improvements

* implement properties for Graph

* Fix errors and warnings after merge

* minor cleanup of properties

* fix the tests

* add some python interface methods

* start changing the properties so that their is a top-level properties object with temporal and static as sub-parts

* Mostly working but still need to fix all the t
[…]
