+++ raphtory-graphql/src/model/graph/graph.rs
@@ -42,7 +42,8 @@ impl GraphMeta {
 
     async fn static_properties(&self) -> Vec<Property> {
         self.graph
-            .static_properties()
+            .properties()
+            .constant()
             .into_iter()
             .map(|(k, v)| Property::new(k, v))
             .collect()
@@ -85,7 +86,8 @@ impl GqlGraph {
 
     async fn static_properties(&self) -> Vec<Property> {
         self.graph
-            .static_properties()
+            .properties()
+            .constant()
             .into_iter()
             .map(|(k, v)| Property::new(k, v))
             .collect()
