+++ raphtory/src/db/graph/edge.rs
@@ -8,16 +8,24 @@
 use crate::{
     core::{
         entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef},
+        storage::locked_view::LockedView,
         utils::time::IntoTime,
     },
     db::{
-        api::view::{BoxedIter, EdgeViewInternalOps},
+        api::{
+            properties::{
+                internal::{
+                    ConstPropertiesOps, Key, TemporalPropertiesOps, TemporalPropertyViewOps,
+                },
+                Properties,
+            },
+            view::{internal::Static, BoxedIter, EdgeViewInternalOps},
+        },
         graph::{vertex::VertexView, views::window_graph::WindowedGraph},
     },
     prelude::*,
 };
 use std::{
-    collections::HashMap,
     fmt::{Debug, Formatter},
     iter,
 };
@@ -30,6 +38,8 @@ pub struct EdgeView<G: GraphViewOps> {
     /// A reference to the edge.
     pub edge: EdgeRef,
 }
+
+impl<G: GraphViewOps> Static for EdgeView<G> {}
 
 impl<G: GraphViewOps> EdgeView<G> {
     pub fn new(graph: G, edge: EdgeRef) -> Self {
@@ -55,6 +65,50 @@ impl<G: GraphViewOps> EdgeViewInternalOps<G, VertexView<G>> for EdgeView<G> {
             graph: self.graph(),
             edge: e,
         }
+    }
+}
+
+impl<G: GraphViewOps> ConstPropertiesOps for EdgeView<G> {
+    fn const_property_keys<'a>(&'a self) -> Box<dyn Iterator<Item = LockedView<'a, String>> + 'a> {
+        self.graph.static_edge_prop_names(self.edge)
+    }
+
+    fn get_const_property(&self, key: &str) -> Option<Prop> {
+        self.graph.static_edge_prop(self.edge, key)
+    }
+}
+
+impl<G: GraphViewOps> TemporalPropertyViewOps for EdgeView<G> {
+    fn temporal_history(&self, id: &Key) -> Vec<i64> {
+        self.graph
+            .temporal_edge_prop_vec(self.edge, id)
+            .into_iter()
+            .map(|(t, _)| t)
+            .collect()
+    }
+
+    fn temporal_values(&self, id: &String) -> Vec<Prop> {
+        self.graph
+            .temporal_edge_prop_vec(self.edge, id)
+            .into_iter()
+            .map(|(_, v)| v)
+            .collect()
+    }
+}
+
+impl<G: GraphViewOps> TemporalPropertiesOps for EdgeView<G> {
+    fn temporal_property_keys<'a>(
+        &'a self,
+    ) -> Box<(dyn Iterator<Item = LockedView<'a, String>> + 'a)> {
+        Box::new(
+            self.graph
+                .temporal_edge_prop_names(self.edge)
+                .filter(|k| self.get_temporal_property(k).is_some()),
+        )
+    }
+
+    fn get_temporal_property(&self, key: &str) -> Option<String> {
+        (!self.graph.temporal_edge_prop_vec(self.edge, key).is_empty()).then_some(key.to_owned())
     }
 }
 
@@ -129,40 +183,8 @@ impl<G: GraphViewOps> EdgeListOps for BoxedIter<EdgeView<G>> {
     /// Specifies the associated type for the iterator over edges.
     type IterType<T> = Box<dyn Iterator<Item = T> + Send>;
 
-    fn has_property(self, name: String, include_static: bool) -> BoxedIter<bool> {
-        Box::new(self.map(move |e| e.has_property(&name, include_static)))
-    }
-
-    fn property(self, name: String, include_static: bool) -> BoxedIter<Option<Prop>> {
[…]
