### 2023-07-26 fc34605
add Rust Benchmark (cargo) benchmark result for 16bf23100dc30df7aba4f2806aab9d8d44b2b403
