### 2023-07-26 3821cc3
make docstring signatures work with type hints
