### 2023-07-27 7a103c3
working version up to events happening at same time
