+++ raphtory/src/algorithms/motifs/three_node_local_redone.rs
@@ -1,22 +1,34 @@
 // Imports ///////////////////////////////////////////
 use crate::{
-    algorithms::{motifs::three_node_motifs::*, k_core::k_core_set},
-    core::state::{accumulator_id::{accumulators::{self, val}, AccId}, compute_state::ComputeStateVec, agg::ValDef},
+    algorithms::{k_core::k_core_set, motifs::three_node_motifs::*},
+    core::state::{
+        accumulator_id::{
+            accumulators::{self, val},
+            AccId,
+        },
+        agg::ValDef,
+        compute_state::ComputeStateVec,
+    },
     db::{
-        api::view::{GraphViewOps, VertexViewOps, *, internal::CoreGraphOps},
+        api::view::{internal::CoreGraphOps, GraphViewOps, VertexViewOps, *},
+        graph::views::vertex_subgraph::{self, VertexSubgraph},
         task::{
             context::Context,
             task::{ATask, Job, Step},
             task_runner::TaskRunner,
             vertex::eval_vertex::EvalVertexView,
-        }, graph::views::vertex_subgraph::{self, VertexSubgraph},
+        },
     },
 };
 
 use itertools::{enumerate, Itertools};
 use num_traits::Zero;
 use rustc_hash::FxHashSet;
-use std::{collections::{HashMap, HashSet}, ops::Add, slice::Iter};
+use std::{
+    collections::{HashMap, HashSet},
+    ops::Add,
+    slice::Iter,
+};
 ///////////////////////////////////////////////////////
 
 // State objects for three node motifs
@@ -120,7 +132,7 @@ pub fn star_motif_count<G: GraphViewOps>(
             }
         })
         .collect::<Vec<StarEvent>>();
-    exploded_edges.sort_by_key(|e| e.time);
+    exploded_edges.sort_by_key(|e| (e.time, e.dir));
     let mut star_count = init_star_count(neigh_map.len());
     star_count.execute(&exploded_edges, delta);
     star_count.return_counts()
@@ -138,7 +150,8 @@ pub fn twonode_motif_count<G: GraphViewOps>(
         let nb_id = nb.id();
         let out = graph.edge(evv.id(), nb_id, None);
         let inc = graph.edge(nb_id, evv.id(), None);
-        let mut all_exploded : Vec<TwoNodeEvent> = out.iter()
+        let mut all_exploded: Vec<TwoNodeEvent> = out
+            .iter()
             .flat_map(|e| e.explode())
             .chain(inc.iter().flat_map(|e| e.explode()))
             .map(|e| {
@@ -148,7 +161,7 @@ pub fn twonode_motif_count<G: GraphViewOps>(
                 )
             })
             .collect();
-        all_exploded.sort_by_key(|e| e.time);
+        all_exploded.sort_by_key(|e| (e.time, e.dir));
         let mut two_node_counter = init_two_node_count();
         two_node_counter.execute(&all_exploded, delta);
         let two_node_result = two_node_counter.return_counts();
@@ -161,194 +174,213 @@ pub fn twonode_motif_count<G: GraphViewOps>(
 
 ///////////////////////////////////////////////////////
 
-pub fn triangle_motifs<G: GraphViewOps>(graph: &G, delta:i64, motifs_count_id:AccId<MotifCounter, MotifCounter,MotifCounter,ValDef<MotifCounter>>, threads: Option<usize>) -> HashMap<String, [usize;8]>{
-    let vertex_set = k_core_set(graph,2,usize::MAX, None);
-    // vertex_set.sort();
-    // let vertex_set
+pub fn triangle_motifs<G: GraphViewOps>(
+    graph: &G,
+    delta: i64,
[…]
