### 2023-07-27 beb33ba
start trying to fix properties
