### 2023-07-27 84e01d5
add some comments to explain how this magic works
