### 2023-07-27 bd33fcf
make docstring signatures work with type hints (#1137)
* make docstring signatures work with type hints

* add some comments to explain how this magic works
