### 2023-07-27 422e328
add Rust Benchmark (cargo) benchmark result for 006d15051260df0e3b44149661e34f4052802908
