+++ raphtory/src/algorithms/k_core.rs
@@ -100,9 +100,10 @@ where
 mod k_core_test {
     use std::collections::HashSet;
 
-    use crate::algorithms::k_core::k_core_set;
-    use crate::algorithms::triangle_count::triangle_count;
-    use crate::prelude::*;
+    use crate::{
+        algorithms::{k_core::k_core_set, triangle_count::triangle_count},
+        prelude::*,
+    };
 
     #[test]
     fn k_core_2() {
