### 2023-07-28 80c2802
build infrastructure make temporal property merging possible
