+++ raphtory/src/db/graph/views/window_graph.rs
@@ -298,25 +298,18 @@ impl<G: GraphViewOps> TimeSemantics for WindowedGraph<G> {
         name: &str,
         t_start: i64,
         t_end: i64,
-        layer_ids: LayerIds,
     ) -> Vec<(i64, Prop)> {
         self.graph.temporal_edge_prop_vec_window(
             e,
             name,
             self.actual_start(t_start),
             self.actual_end(t_end),
-            layer_ids,
         )
     }
 
-    fn temporal_edge_prop_vec(
-        &self,
-        e: EdgeRef,
-        name: &str,
-        layer_ids: LayerIds,
-    ) -> Vec<(i64, Prop)> {
+    fn temporal_edge_prop_vec(&self, e: EdgeRef, name: &str) -> Vec<(i64, Prop)> {
         self.graph
-            .temporal_edge_prop_vec_window(e, name, self.t_start, self.t_end, layer_ids)
+            .temporal_edge_prop_vec_window(e, name, self.t_start, self.t_end)
     }
 
     fn edge_layers(&self, e: EdgeRef, layer_ids: LayerIds) -> BoxedIter<EdgeRef> {
