### 2023-07-28 602c847
add Rust Benchmark (cargo) benchmark result for b932685d360455e388094169ec64292493e08b3d
