+++ raphtory/src/db/internal/graph_ops.rs
@@ -13,12 +13,26 @@ use genawaiter::sync::GenBoxed;
 use std::ops::Deref;
 
 impl<const N: usize> GraphOps for InnerTemporalGraph<N> {
+    fn layer_ids(&self) -> LayerIds {
+        LayerIds::All
+    }
+
+    fn local_vertex_ref(&self, v: VertexRef) -> Option<VID> {
+        match v {
+            VertexRef::Local(l) => Some(l),
+            VertexRef::Remote(_) => {
+                let vid = self.inner().resolve_vertex_ref(&v)?;
+                Some(vid)
+            }
+        }
+    }
+
     fn find_edge_id(&self, e_id: EID) -> Option<EdgeRef> {
         let e_id_usize: usize = e_id.into();
         if e_id_usize >= self.num_edges() {
             return None;
         }
-        let edge_view = self.edge(e_id);
+        let edge_view = self.inner().edge(e_id);
         Some(EdgeRef::new_outgoing(
             e_id,
             edge_view.src_id(),
@@ -26,56 +40,49 @@ impl<const N: usize> GraphOps for InnerTemporalGraph<N> {
         ))
     }
 
-    fn local_vertex_ref(&self, v: VertexRef) -> Option<VID> {
-        match v {
-            VertexRef::Local(l) => Some(l),
-            VertexRef::Remote(_) => {
-                let vid = self.resolve_vertex_ref(&v)?;
-                Some(vid)
-            }
-        }
+    fn layer_ids_from_names(&self, key: Layer) -> LayerIds {
+        self.inner().layer_id(key)
     }
 
-    fn get_unique_layers_internal(&self) -> Vec<usize> {
-        self.get_all_layers()
-    }
-
-    fn get_layer_id(&self, key: Layer) -> Option<LayerIds> {
-        self.layer_id(key)
+    fn edge_layer_ids(&self, e_id: EID) -> LayerIds {
+        let edge = self.inner().edge(e_id);
+        edge.layer_ids()
     }
 
     fn vertices_len(&self) -> usize {
-        self.internal_num_vertices()
+        self.inner().internal_num_vertices()
     }
 
     fn edges_len(&self, layers: LayerIds) -> usize {
-        self.internal_num_edges(layers)
+        self.inner().internal_num_edges(layers)
     }
 
     fn degree(&self, v: VID, d: Direction, layers: LayerIds) -> usize {
-        self.degree(v, d, layers)
+        self.inner().degree(v, d, layers)
     }
 
     fn vertex_refs(&self) -> Box<dyn Iterator<Item = VID> + Send> {
-        Box::new(self.vertex_ids())
+        Box::new(self.inner().vertex_ids())
     }
 
     fn edge_ref(&self, src: VertexRef, dst: VertexRef, layer: LayerIds) -> Option<EdgeRef> {
-        let src = self.resolve_vertex_ref(&src)?;
-        let dst = self.resolve_vertex_ref(&dst)?;
+        let src = self.inner().resolve_vertex_ref(&src)?;
+        let dst = self.inner().resolve_vertex_ref(&dst)?;
 
-        self.find_edge(src, dst, layer)
+        self.inner()
+            .find_edge(src, dst, layer)
             .map(|e_id| EdgeRef::new_outgoing(e_id, src, dst))
     }
 
     fn edge_refs(&self, layers: LayerIds) -> Box<dyn Iterator<Item = EdgeRef> + Send> {
         match layers {
             LayerIds::All => {
-                let iter = self.locked_edges().map(|edge| edge.deref().into());
+                let iter = self.inner().locked_edges().map(|edge| edge.deref().into());
                 Box::new(iter)
             }
[…]
