### 2023-07-28 5fdf221
fix static properties
