+++ raphtory/src/db/task/edge/window_eval_edge.rs
@@ -108,7 +108,8 @@ impl<'a, G: GraphViewOps, CS: ComputeState, S: 'static> ConstPropertiesOps
     }
 
     fn get_const_property(&self, key: &str) -> Option<Prop> {
-        self.graph().static_edge_prop(self.ev, key)
+        self.graph()
+            .static_edge_prop(self.ev, key, self.g.layer_ids())
     }
 }
 
