### 2023-07-31 3658759
add Rust Benchmark (cargo) benchmark result for e0f53e8ce1dfa06490f302fd0b05a24536e9f98f
