### 2023-08-01 e757742
fix warnings and clean up property list
