+++ raphtory/src/db/graph/edge.rs
@@ -8,7 +8,7 @@
 use super::views::layer_graph::LayeredGraph;
 use crate::{
     core::{
-        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, LayerIds},
+        entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef},
         storage::locked_view::LockedView,
         utils::time::IntoTime,
     },
