### 2023-08-01 89b7e4d
start adding List and Map to TProp
