### 2023-08-01 de89e6e
remove stray println that got reintroduced
