+++ python/tests/test_graphdb.py
@@ -611,15 +611,15 @@ def test_edge_properties():
                                                          'prop 4': True}
 
     assert g.at(2).edge(1, 2).properties == {'prop 1': 2, 'prop 4': False, 'prop 2': 0.6, 'static prop': 123,
-                                               'prop 3': 'hi'}
+                                             'prop 3': 'hi'}
 
     # testing property histories
     assert g.edge(1, 2).properties.temporal == {'prop 3': [(1, 'hi'), (3, 'hello')], 'prop 1': [(1, 1), (2, 2)],
-                                                 'prop 4': [(1, True), (2, False), (3, True)],
-                                                 'prop 2': [(2, 0.6), (3, 0.9)]}
+                                                'prop 4': [(1, True), (2, False), (3, True)],
+                                                'prop 2': [(2, 0.6), (3, 0.9)]}
 
     assert g.at(2).edge(1, 2).properties.temporal == {'prop 2': [(2, 0.6)], 'prop 4': [(1, True), (2, False)],
-                                                       'prop 1': [(1, 1), (2, 2)], 'prop 3': [(1, 'hi')]}
+                                                      'prop 1': [(1, 1), (2, 2)], 'prop 3': [(1, 'hi')]}
 
     # testing property names
     assert sorted(g.edge(1, 2).properties.keys()) == sorted(['prop 4', 'prop 1', 'prop 2', 'prop 3', 'static prop'])
@@ -627,7 +627,7 @@ def test_edge_properties():
     assert sorted(g.edge(1, 2).properties.temporal.keys()) == sorted(['prop 4', 'prop 1', 'prop 2', 'prop 3'])
 
     assert sorted(g.at(1).edge(1, 2).properties.temporal.keys()) == sorted(['prop 4', 'prop 1',
-                                                                              'prop 3'])
+                                                                            'prop 3'])
 
     # testing has_property
     assert "prop 4" in g.edge(1, 2).properties
@@ -1052,7 +1052,9 @@ def test_algo_result():
     assert sorted_res == {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1, '6': 1, '7': 1, '8': 1}
     # algo str f64
     actual = algorithms.pagerank(g)
-    expected_result = {'3': 0.10274080842110422, '2': 0.10274080842110422, '4': 0.1615298183542792, '6': 0.14074777909144864, '1': 0.07209850165402759, '5': 0.1615298183542792, '7': 0.14074777909144864, '8': 0.11786468661230831}
+    expected_result = {'3': 0.10274080842110422, '2': 0.10274080842110422, '4': 0.1615298183542792,
+                       '6': 0.14074777909144864, '1': 0.07209850165402759, '5': 0.1615298183542792,
+                       '7': 0.14074777909144864, '8': 0.11786468661230831}
     assert actual.get_all() == expected_result
     assert len(actual.to_df()) == 8
     # algo str vector
@@ -1397,49 +1399,65 @@ def test_graphQL():
     import os
 
     g1 = Graph()
-    g1.add_edge(1,"ben","hamza")
-    g1.add_edge(2,"haaroon","hamza")
-    g1.add_edge(3,"ben","haaroon")
+    g1.add_edge(1, "ben", "hamza")
+    g1.add_edge(2, "haaroon", "hamza")
[…]
