### 2023-08-01 3de7cc7
Rename AsProp to IntoProp to fit rust conventions and add specialised converters for maps and lists to avoid ambiguous implementation errors due to rust ignoring associated types
