+++ raphtory/src/lib.rs
@@ -102,7 +102,7 @@ pub mod search;
 pub mod prelude {
     pub const NO_PROPS: [(&str, Prop); 0] = [];
     pub use crate::{
-        core::{AsProp, Prop, PropUnwrap},
+        core::{IntoProp, Prop, PropUnwrap},
         db::{
             api::{
                 mutation::{AdditionOps, DeletionOps, PropertyAdditionOps},
