+++ raphtory/src/db/graph/edge.rs
@@ -307,7 +307,7 @@ mod test_edge {
     #[test]
     fn test_properties() {
         let g = Graph::new();
-        let props = [("test".to_string(), "test".as_prop())];
+        let props = [("test".to_string(), "test".into_prop())];
         g.add_edge(0, 1, 2, NO_PROPS, None).unwrap();
         g.add_edge(2, 1, 2, props.clone(), None).unwrap();
 
