### 2023-08-01 7496ab5
add more tests for edge layer properties
