### 2023-08-02 2aca124
layered graph can inherit TimeSemantics again
