+++ raphtory/src/db/graph/views/layer_graph.rs
@@ -1,5 +1,3 @@
-use std::ops::Range;
-
 use crate::{
     core::{
         entities::{edges::edge_ref::EdgeRef, vertices::vertex_ref::VertexRef, LayerIds, EID, VID},
@@ -8,11 +6,11 @@ use crate::{
     db::api::{
         properties::internal::InheritPropertiesOps,
         view::{
-            internal::{Base, GraphOps, InheritCoreOps, InheritMaterialize, TimeSemantics},
-            BoxedIter, Layer,
+            internal::{Base, GraphOps, InheritCoreOps, InheritMaterialize, InheritTimeSemantics},
+            Layer,
         },
     },
-    prelude::{GraphViewOps, Prop},
+    prelude::GraphViewOps,
 };
 use itertools::Itertools;
 
@@ -32,7 +30,7 @@ impl<G: GraphViewOps> Base for LayeredGraph<G> {
     }
 }
 
-// impl<G: GraphViewOps> InheritTimeSemantics for LayeredGraph<G> {}
+impl<G: GraphViewOps> InheritTimeSemantics for LayeredGraph<G> {}
 
 impl<G: GraphViewOps> InheritCoreOps for LayeredGraph<G> {}
 
@@ -155,126 +153,5 @@ impl<G: GraphViewOps> GraphOps for LayeredGraph<G> {
         layer: LayerIds,
     ) -> Box<dyn Iterator<Item = VertexRef> + Send> {
         self.graph.neighbours(v, d, self.constrain(layer))
-    }
-}
-
-impl<G: GraphViewOps> TimeSemantics for LayeredGraph<G> {
-    fn include_vertex_window(&self, v: VID, w: Range<i64>) -> bool {
-        self.graph.include_vertex_window(v, w)
-    }
-
-    fn include_edge_window(&self, e: EdgeRef, w: Range<i64>, layer_ids: LayerIds) -> bool {
-        self.graph
-            .include_edge_window(e, w, self.constrain(layer_ids))
-    }
-
-    fn edge_t(&self, e: EdgeRef, layer_ids: LayerIds) -> BoxedIter<EdgeRef> {
-        self.graph.edge_t(e, self.constrain(layer_ids))
-    }
-
-    fn edge_layers(&self, e: EdgeRef, layer_ids: LayerIds) -> BoxedIter<EdgeRef> {
-        self.graph.edge_layers(e, self.constrain(layer_ids))
-    }
-
-    fn edge_window_t(&self, e: EdgeRef, w: Range<i64>, layer_ids: LayerIds) -> BoxedIter<EdgeRef> {
-        self.graph.edge_window_t(e, w, self.constrain(layer_ids))
-    }
-
-    fn edge_window_layers(
-        &self,
-        e: EdgeRef,
-        w: Range<i64>,
-        layer_ids: LayerIds,
-    ) -> BoxedIter<EdgeRef> {
-        self.graph
-            .edge_window_layers(e, w, self.constrain(layer_ids))
-    }
-
-    fn edge_earliest_time(&self, e: EdgeRef, layer_ids: LayerIds) -> Option<i64> {
-        self.graph.edge_earliest_time(e, self.constrain(layer_ids))
-    }
-
-    fn edge_earliest_time_window(
-        &self,
-        e: EdgeRef,
-        w: Range<i64>,
-        layer_ids: LayerIds,
-    ) -> Option<i64> {
-        self.graph
-            .edge_earliest_time_window(e, w, self.constrain(layer_ids))
-    }
-
-    fn edge_latest_time(&self, e: EdgeRef, layer_ids: LayerIds) -> Option<i64> {
-        self.graph.edge_latest_time(e, self.constrain(layer_ids))
-    }
-
-    fn edge_latest_time_window(
[…]
