+++ raphtory/src/db/internal/prop_add.rs
@@ -10,7 +10,7 @@ impl<const N: usize> InternalPropertyAdditionOps for InnerTemporalGraph<N> {
         v: u64,
         data: Vec<(String, Prop)>,
     ) -> Result<(), GraphError> {
-        self.add_vertex_properties_internal(v, data)
+        self.inner().add_vertex_properties_internal(v, data)
     }
 
     fn internal_add_properties(
@@ -18,11 +18,11 @@ impl<const N: usize> InternalPropertyAdditionOps for InnerTemporalGraph<N> {
         t: i64,
         props: Vec<(String, Prop)>,
     ) -> Result<(), GraphError> {
-        self.add_property(t, props)
+        self.inner().add_property(t, props)
     }
 
     fn internal_add_static_properties(&self, props: Vec<(String, Prop)>) -> Result<(), GraphError> {
-        self.add_static_property(props)
+        self.inner().add_static_property(props)
     }
 
     fn internal_add_edge_properties(
@@ -32,6 +32,7 @@ impl<const N: usize> InternalPropertyAdditionOps for InnerTemporalGraph<N> {
         props: Vec<(String, Prop)>,
         layer: Option<&str>,
     ) -> Result<(), GraphError> {
-        self.add_edge_properties_internal(src, dst, props, layer)
+        self.inner()
+            .add_edge_properties_internal(src, dst, props, layer)
     }
 }
