+++ raphtory/src/python/graph/vertex.rs
@@ -15,7 +15,7 @@ use crate::{
             path::{PathFromGraph, PathFromVertex},
             vertex::VertexView,
             vertices::Vertices,
-            views::window_graph::WindowedGraph,
+            views::{layer_graph::LayeredGraph, window_graph::WindowedGraph},
         },
     },
     python::{
@@ -363,10 +363,10 @@ impl PyVertex {
         self.vertex.default_layer().into()
     }
 
-    #[doc = layer_doc_string!()]
+    #[doc = layers_doc_string!()]
     #[pyo3(signature = (name))]
-    pub fn layer(&self, name: &str) -> Option<PyVertex> {
-        Some(self.vertex.layer(name)?.into())
+    pub fn layers(&self, name: Vec<String>) -> Option<VertexView<LayeredGraph<DynamicGraph>>> {
+        self.vertex.layer(name)
     }
 
     /// Returns the history of a vertex, including vertex additions and changes made to vertex.
@@ -480,7 +480,6 @@ impl PyVertices {
         (move || vertices.latest_time()).into()
     }
 
-    //Fixme: needs a view that allows indexing
     #[getter]
     fn properties(&self) -> PyPropsList {
         let vertices = self.vertices.clone();
@@ -660,10 +659,10 @@ impl PyVertices {
         self.vertices.default_layer().into()
     }
 
-    #[doc = layer_doc_string!()]
+    #[doc = layers_doc_string!()]
     #[pyo3(signature = (name))]
-    pub fn layer(&self, name: &str) -> Option<PyVertices> {
-        Some(self.vertices.layer(name)?.into())
+    pub fn layer(&self, name: &str) -> Option<Vertices<LayeredGraph<DynamicGraph>>> {
+        self.vertices.layer(name)
     }
 
     //****** Python *******
@@ -838,10 +837,10 @@ impl PyPathFromGraph {
         self.path.default_layer().into()
     }
 
-    #[doc = layer_doc_string!()]
+    #[doc = layers_doc_string!()]
     #[pyo3(signature = (name))]
-    pub fn layer(&self, name: &str) -> Option<Self> {
-        Some(self.path.layer(name)?.into())
+    pub fn layer(&self, name: &str) -> Option<PathFromGraph<LayeredGraph<DynamicGraph>>> {
+        self.path.layer(name)
     }
 
     fn __repr__(&self) -> String {
@@ -1031,10 +1030,10 @@ impl PyPathFromVertex {
         self.path.default_layer().into()
     }
 
-    #[doc = layer_doc_string!()]
+    #[doc = layers_doc_string!()]
     #[pyo3(signature = (name))]
-    pub fn layer(&self, name: &str) -> Option<Self> {
-        Some(self.path.layer(name)?.into())
+    pub fn layer(&self, name: &str) -> Option<PathFromVertex<LayeredGraph<DynamicGraph>>> {
+        self.path.layer(name)
     }
 
     fn __repr__(&self) -> String {
