+++ raphtory-graphql/src/model/graph/node.rs
@@ -80,9 +80,9 @@ impl Node {
     }
 
     async fn in_neighbours<'a>(&self, layer: Option<String>) -> Vec<Node> {
-        match layer {
+        match layer.as_deref() {
             None => self.vv.in_neighbours().iter().map(|vv| vv.into()).collect(),
-            Some(layer) => match self.vv.layer(layer.as_str()) {
+            Some(layer) => match self.vv.layer(layer) {
                 None => {
                     vec![]
                 }
@@ -92,14 +92,14 @@ impl Node {
     }
 
     async fn out_neighbours(&self, layer: Option<String>) -> Vec<Node> {
-        match layer {
+        match layer.as_deref() {
             None => self
                 .vv
                 .out_neighbours()
                 .iter()
                 .map(|vv| vv.into())
                 .collect(),
-            Some(layer) => match self.vv.layer(layer.as_str()) {
+            Some(layer) => match self.vv.layer(layer) {
                 None => {
                     vec![]
                 }
@@ -109,9 +109,9 @@ impl Node {
     }
 
     async fn neighbours<'a>(&self, layer: Option<String>) -> Vec<Node> {
-        match layer {
+        match layer.as_deref() {
             None => self.vv.neighbours().iter().map(|vv| vv.into()).collect(),
-            Some(layer) => match self.vv.layer(layer.as_str()) {
+            Some(layer) => match self.vv.layer(layer) {
                 None => {
                     vec![]
                 }
@@ -124,9 +124,9 @@ impl Node {
         match layers {
             None => self.vv.degree(),
             Some(layers) => layers
-                .into_iter()
+                .iter()
                 .map(|layer| {
-                    let degree = match self.vv.layer(layer.as_str()) {
+                    let degree = match self.vv.layer(layer) {
                         None => 0,
                         Some(vvv) => vvv.degree(),
                     };
@@ -137,9 +137,9 @@ impl Node {
     }
 
     async fn out_degree(&self, layer: Option<String>) -> usize {
-        match layer {
+        match layer.as_deref() {
             None => self.vv.out_degree(),
-            Some(layer) => match self.vv.layer(layer.as_str()) {
+            Some(layer) => match self.vv.layer(layer) {
                 None => 0,
                 Some(vvv) => vvv.out_degree(),
             },
@@ -147,9 +147,9 @@ impl Node {
     }
 
     async fn in_degree(&self, layer: Option<String>) -> usize {
-        match layer {
+        match layer.as_deref() {
             None => self.vv.in_degree(),
-            Some(layer) => match self.vv.layer(layer.as_str()) {
+            Some(layer) => match self.vv.layer(layer) {
                 None => 0,
                 Some(vvv) => vvv.in_degree(),
             },
@@ -157,9 +157,9 @@ impl Node {
     }
 
     async fn out_edges(&self, layer: Option<String>) -> Vec<Edge> {
-        match layer {
+        match layer.as_deref() {
             None => self.vv.out_edges().map(|ee| ee.into()).collect(),
-            Some(layer) => match self.vv.layer(layer.as_str()) {
+            Some(layer) => match self.vv.layer(layer) {
                 None => {
                     vec![]
                 }
@@ -169,9 +169,9 @@ impl Node {
     }
 
     async fn in_edges(&self, layer: Option<String>) -> Vec<Edge> {
-        match layer {
+        match layer.as_deref() {
             None => self.vv.in_edges().map(|ee| ee.into()).collect(),
-            Some(layer) => match self.vv.layer(layer.as_str()) {
+            Some(layer) => match self.vv.layer(layer) {
                 None => {
                     vec![]
[…]

[…]
