### 2023-08-02 8f18b50
add Rust Benchmark (cargo) benchmark result for a01ee30e42314d4783479b813e02e436eeb743cd
