### 2023-08-02 4026282
Fix __getitem__ for python TemporalProperties
