### 2023-08-02 73dd15c
Fix __getitem__ for python TemporalProperties (#1147)
* add test to confirm bug

* Fix __getitem__ for python TemporalProperties
