### 2023-08-02 dd472b2
add Rust Benchmark (cargo) benchmark result for a866fcab93382049905b109bed50942e9cb24b70
