+++ raphtory/src/python/graph/graph_with_deletions.rs
@@ -9,7 +9,7 @@ use crate::{
     core::{utils::errors::GraphError, Prop},
     db::{
         api::mutation::{AdditionOps, PropertyAdditionOps},
-        graph::views::deletion_graph::GraphWithDeletions,
+        graph::{vertex::VertexView, views::deletion_graph::GraphWithDeletions},
     },
     prelude::DeletionOps,
     python::{
@@ -100,7 +100,7 @@ impl PyGraphWithDeletions {
         timestamp: PyTime,
         id: PyInputVertex,
         properties: Option<HashMap<String, Prop>>,
-    ) -> Result<(), GraphError> {
+    ) -> Result<VertexView<GraphWithDeletions>, GraphError> {
         self.graph
             .add_vertex(timestamp, id, properties.unwrap_or_default())
     }
