### 2023-08-02 862b0ae
fix range bug for TimeIndexEntry
