### 2023-08-03 3db88fd
Vertex timestamps shouldn't have secondary indices
