+++ raphtory/src/core/storage/timeindex.rs
@@ -1,22 +1,27 @@
+use crate::{
+    core::{entities::LayerIds, utils::time::error::ParseTimeError},
+    db::api::mutation::{internal::InternalAdditionOps, InputTime, TryIntoInputTime},
+};
 use itertools::Itertools;
 use num_traits::Saturating;
 use serde::{Deserialize, Serialize};
 use std::{
     cmp::{max, min},
     collections::BTreeSet,
+    fmt::Debug,
     ops::Range,
     sync::Arc,
-};
-
-use crate::{
-    core::{entities::LayerIds, utils::time::error::ParseTimeError},
-    db::api::mutation::{internal::InternalAdditionOps, InputTime, TryIntoInputTime},
 };
 
 use super::locked_view::LockedView;
 
 #[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Ord, PartialOrd, Eq)]
 pub struct TimeIndexEntry(i64, usize);
+
+pub trait AsTime: Debug + Copy + Ord + Eq + Send + Sync {
+    fn t(&self) -> &i64;
+    fn range(w: Range<i64>) -> Range<Self>;
+}
 
 impl From<i64> for TimeIndexEntry {
     fn from(value: i64) -> Self {
@@ -47,33 +52,44 @@ impl TimeIndexEntry {
     pub fn end(t: i64) -> Self {
         Self(t.saturating_add(1), 0)
     }
+}
 
-    pub fn range(w: Range<i64>) -> Range<Self> {
-        Self::start(w.start)..Self::start(w.end)
+impl AsTime for i64 {
+    fn t(&self) -> &i64 {
+        self
     }
 
-    pub fn t(&self) -> &i64 {
+    fn range(w: Range<i64>) -> Range<Self> {
+        w
+    }
+}
+
+impl AsTime for TimeIndexEntry {
+    fn t(&self) -> &i64 {
         &self.0
+    }
+    fn range(w: Range<i64>) -> Range<Self> {
+        Self::start(w.start)..Self::start(w.end)
     }
 }
 
 #[derive(Default, Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum TimeIndex {
+pub enum TimeIndex<T: Ord + Eq + Copy + Debug> {
     #[default]
     Empty,
-    One(TimeIndexEntry),
-    Set(BTreeSet<TimeIndexEntry>),
+    One(T),
+    Set(BTreeSet<T>),
 }
 
-impl TimeIndex {
+impl<T: AsTime> TimeIndex<T> {
     pub fn is_empty(&self) -> bool {
         matches!(self, TimeIndex::Empty)
     }
 
-    pub fn one(ti: TimeIndexEntry) -> Self {
+    pub fn one(ti: T) -> Self {
         Self::One(ti)
     }
-    pub fn insert(&mut self, ti: TimeIndexEntry) -> bool {
+    pub fn insert(&mut self, ti: T) -> bool {
         match self {
             TimeIndex::Empty => {
                 *self = TimeIndex::One(ti);
@@ -95,7 +111,7 @@ impl TimeIndex {
         match self {
             TimeIndex::Empty => false,
             TimeIndex::One(t) => w.contains(t.t()),
-            TimeIndex::Set(ts) => ts.range(TimeIndexEntry::range(w)).next().is_some(),
+            TimeIndex::Set(ts) => ts.range(T::range(w)).next().is_some(),
         }
     }
 
@@ -112,7 +128,7 @@ impl TimeIndex {
                     Box::new(std::iter::empty())
                 }
             }
[…]
