### 2023-08-03 4017a76
IndexedGraph needs to implement GraphViewOps
