### 2023-08-03 58fb11e
clean up uses of VertexRef
