+++ raphtory/src/db/internal/time_semantics.rs
@@ -17,14 +17,14 @@ impl<const N: usize> TimeSemantics for InnerTemporalGraph<N> {
         self.inner()
             .node_entry(v)
             .value()
-            .and_then(|node| node.timestamps().first())
+            .and_then(|node| node.timestamps().first_t())
     }
 
     fn vertex_latest_time(&self, v: VID) -> Option<i64> {
         self.inner()
             .node_entry(v)
             .value()
-            .and_then(|node| node.timestamps().last())
+            .and_then(|node| node.timestamps().last_t())
     }
 
     fn view_start(&self) -> Option<i64> {
@@ -47,14 +47,14 @@ impl<const N: usize> TimeSemantics for InnerTemporalGraph<N> {
         self.inner()
             .node_entry(v)
             .value()
-            .and_then(|node| node.timestamps().range(t_start..t_end).first())
+            .and_then(|node| node.timestamps().range(t_start..t_end).first_t())
     }
 
     fn vertex_latest_time_window(&self, v: VID, t_start: i64, t_end: i64) -> Option<i64> {
         self.inner()
             .node_entry(v)
             .value()
-            .and_then(|node| node.timestamps().range(t_start..t_end).last())
+            .and_then(|node| node.timestamps().range(t_start..t_end).last_t())
     }
 
     fn include_vertex_window(&self, v: VID, w: Range<i64>) -> bool {
@@ -66,11 +66,15 @@ impl<const N: usize> TimeSemantics for InnerTemporalGraph<N> {
     }
 
     fn vertex_history(&self, v: VID) -> Vec<i64> {
-        self.vertex_additions(v).iter().copied().collect()
+        self.vertex_additions(v).iter_t().copied().collect()
     }
 
     fn vertex_history_window(&self, v: VID, w: Range<i64>) -> Vec<i64> {
-        self.vertex_additions(v).range(w).iter().copied().collect()
+        self.vertex_additions(v)
+            .range(w)
+            .iter_t()
+            .copied()
+            .collect()
     }
 
     fn edge_t(&self, e: EdgeRef, layer_ids: LayerIds) -> BoxedIter<EdgeRef> {
@@ -127,8 +131,8 @@ impl<const N: usize> TimeSemantics for InnerTemporalGraph<N> {
     }
 
     fn edge_earliest_time(&self, e: EdgeRef, layer_ids: LayerIds) -> Option<i64> {
-        e.time()
-            .or_else(|| self.edge_additions(e, layer_ids).first())
+        e.time_t()
+            .or_else(|| self.edge_additions(e, layer_ids).first_t())
     }
 
     fn edge_earliest_time_window(
@@ -137,13 +141,13 @@ impl<const N: usize> TimeSemantics for InnerTemporalGraph<N> {
         w: Range<i64>,
         layer_ids: LayerIds,
     ) -> Option<i64> {
-        e.time()
-            .or_else(|| self.edge_additions(e, layer_ids).range(w).first())
+        e.time_t()
+            .or_else(|| self.edge_additions(e, layer_ids).range(w).first_t())
     }
 
     fn edge_latest_time(&self, e: EdgeRef, layer_ids: LayerIds) -> Option<i64> {
-        e.time()
-            .or_else(|| self.edge_additions(e, layer_ids).last())
+        e.time_t()
+            .or_else(|| self.edge_additions(e, layer_ids).last_t())
     }
 
     fn edge_latest_time_window(
@@ -152,12 +156,15 @@ impl<const N: usize> TimeSemantics for InnerTemporalGraph<N> {
         w: Range<i64>,
         layer_ids: LayerIds,
     ) -> Option<i64> {
-        e.time()
[…]

[…]
