### 2023-08-03 d2019ad
IndexedGraph needs to support IntoDynamic
