### 2023-08-03 af5f27c
fix property values for exploded edges at same time point
