+++ raphtory/src/core/storage/sorted_vec_map.rs
@@ -1,6 +1,6 @@
 use serde::{ser::SerializeSeq, Deserialize, Serialize};
 use sorted_vector_map::SortedVectorMap;
-use std::ops::Range;
+use std::{borrow::Borrow, ops::Range};
 
 // wrapper for SortedVectorMap
 #[derive(Debug, PartialEq, Clone)]
@@ -15,6 +15,14 @@ impl<K: Ord, V> Default for SVM<K, V> {
 impl<K: Ord, V> SVM<K, V> {
     pub(crate) fn new() -> Self {
         Self(SortedVectorMap::new())
+    }
+
+    pub(crate) fn get<Q>(&self, k: &Q) -> Option<&V>
+    where
+        K: Borrow<Q>,
+        Q: Ord + ?Sized,
+    {
+        self.0.get(k)
     }
 
     pub(crate) fn insert(&mut self, k: K, v: V) -> Option<V> {
