### 2023-08-03 fe79c34
add test for exploding multilayer edge
