### 2023-08-03 adcaef6
fix explode for edges
