### 2023-08-03 277454a
edge should not take layer name as input any more
