+++ raphtory/src/db/internal/core_ops.rs
@@ -285,7 +285,7 @@ mod test_edges {
         g.add_edge_properties(1, 2, [("layer", 3)], Some("layer3"))
             .unwrap();
 
-        let e_all = g.edge(1, 2, Layer::All).unwrap();
+        let e_all = g.edge(1, 2).unwrap();
         assert_eq!(
             e_all.properties().constant().as_map(),
             HashMap::from([
@@ -301,7 +301,7 @@ mod test_edges {
             vec![0.into(), 1.into()]
         );
 
-        let e = g.edge(1, 2, "layer1").unwrap();
+        let e = g.edge(1, 2).unwrap().layer("layer1").unwrap();
         assert!(e.properties().constant().contains("layer1"));
     }
 }
