+++ raphtory/src/db/graph/graph.rs
@@ -20,10 +20,7 @@ use crate::{
     core::{entities::graph::tgraph::InnerTemporalGraph, utils::errors::GraphError},
     db::api::{
         mutation::internal::{InheritAdditionOps, InheritPropertyAdditionOps},
-        view::{
-            internal::{Base, DynamicGraph, InheritViewOps, IntoDynamic},
-            Layer,
-        },
+        view::internal::{Base, DynamicGraph, InheritViewOps, IntoDynamic},
     },
     prelude::*,
 };
@@ -47,7 +44,7 @@ pub fn graph_equal<G1: GraphViewOps, G2: GraphViewOps>(g1: &G1, g2: &G2) -> bool
             g1.edges().explode().count() == g2.edges().explode().count() && // same number of exploded edges
             g1.edges().explode().all(|e| { // all exploded edges exist in other
                 g2
-                    .edge(e.src().id(), e.dst().id(), Layer::All)
+                    .edge(e.src().id(), e.dst().id())
                     .filter(|ee| ee.active(e.time().expect("exploded")))
                     .is_some()
             })
@@ -153,7 +150,7 @@ mod db_tests {
     use super::*;
     use crate::{
         core::{
-            entities::{vertices::vertex_ref::VertexRef, LayerIds},
+            entities::LayerIds,
             utils::time::{error::ParseTimeError, TryIntoTime},
             Direction, Prop,
         },
@@ -252,7 +249,7 @@ mod db_tests {
 
         edges
             .iter()
-            .all(|&(_, src, dst)| g.edge(src, dst, Layer::All).is_some())
+            .all(|&(_, src, dst)| g.edge(src, dst).is_some())
     }
 
     #[test]
@@ -502,9 +499,9 @@ mod db_tests {
         let v11 = g.vertex(11).unwrap();
         let v22 = g.vertex(22).unwrap();
         let v33 = g.vertex(33).unwrap();
-        let edge1111 = g.edge(&v11, &v11, Layer::All).unwrap();
-        let edge2233 = g.edge(&v22, &v33, Layer::All).unwrap();
-        let edge3311 = g.edge(&v33, &v11, Layer::All).unwrap();
+        let edge1111 = g.edge(&v11, &v11).unwrap();
+        let edge2233 = g.edge(&v22, &v33).unwrap();
+        let edge3311 = g.edge(&v33, &v11).unwrap();
 
         g.add_vertex_properties(11, vec![("a", Prop::U64(11)), ("b", Prop::I64(11))])
             .unwrap();
@@ -590,7 +587,7 @@ mod db_tests {
         g.add_edge(1, 0, 1, vec![("distance".to_string(), Prop::U32(5))], None)
             .expect("add edge");
 
-        let e = g.edge(0, 1, Layer::All).unwrap();
+        let e = g.edge(0, 1).unwrap();
 
         let prop = e.properties().get("distance").unwrap();
         assert_eq!(prop, Prop::U32(5));
@@ -685,10 +682,10 @@ mod db_tests {
         assert!(!g.has_edge(11, 22, "layer2"));
         assert!(g.has_edge(11, 44, "layer2"));
 
-        assert!(g.edge(11, 22, Layer::All).is_some());
-        assert!(g.edge(11, 44, Layer::Default).is_none());
-        assert!(g.edge(11, 22, "layer2").is_none());
-        assert!(g.edge(11, 44, "layer2").is_some());
+        assert!(g.edge(11, 22).is_some());
+        assert!(g.layer(Layer::Default).unwrap().edge(11, 44).is_none());
+        assert!(g.edge(11, 22).unwrap().layer("layer2").is_none());
[…]

[…]
