### 2023-08-03 8dc6086
add exploded edge count to example
