### 2023-08-07 6c8a731
fix FromPyObject for Graph and add support for list and map properties in python
