### 2023-08-07 d66941d
test graph, list, and map properties
