### 2023-08-08 88ce99b
Change add_edge such that it returns an edge view and fix deadlocks because of non-recursive reads
