+++ raphtory/src/core/entities/graph/tgraph.rs
@@ -24,7 +24,7 @@ use crate::{
             Entry,
         },
         utils::{
-            errors::{GraphError, MutateGraphError},
+            errors::{GraphError, IllegalMutate, MutateGraphError},
             time::TryIntoTime,
         },
         Direction, Prop, PropUnwrap,
@@ -281,12 +281,7 @@ impl<const N: usize> TemporalGraph<N> {
         prop_id: usize,
         is_static: bool,
     ) -> Option<LockedView<String>> {
-        let out = self.edge_meta.reverse_prop_id(prop_id, is_static);
-        if out.is_none() {
-            println!("reverse_prop_id_map: {:?}", self.edge_meta);
-            println!("reverse_prop_id: {} -> {:?}", prop_id, out);
-        }
-        out
+        self.edge_meta.reverse_prop_id(prop_id, is_static)
     }
 }
 
@@ -346,13 +341,18 @@ impl<const N: usize> TemporalGraph<N> {
         node.update_time(time);
 
         // update the properties;
-        for (prop_id, _, prop) in &props {
-            node.add_prop(time, *prop_id, prop.clone());
+        for (prop_id, prop) in props {
+            node.add_prop(time, prop_id, prop);
         }
 
         // update node prop
-        for (prop_id, name, prop) in &name_prop {
-            node.add_static_prop(*prop_id, name, prop.clone())?;
+        for (prop_id, prop) in name_prop {
+            node.add_static_prop(prop_id, prop).map_err(|err| {
+                MutateGraphError::IllegalVertexPropertyChange {
+                    vertex_id: v,
+                    source: IllegalMutate::from_source(err, "_id"),
+                }
+            })?;
         }
 
         Ok(v_id.into())
@@ -383,11 +383,21 @@ impl<const N: usize> TemporalGraph<N> {
             let mut node = self.storage.get_node_mut(vid);
             for (prop_name, prop) in data {
                 let prop_id = self.vertex_meta.resolve_prop_id(&prop_name, true);
-                node.add_static_prop(prop_id, &prop_name, prop.clone())
-                    .map_err(|err| GraphError::FailedToMutateGraph { source: err })?;
+                node.add_static_prop(prop_id, prop).map_err(|err| {
+                    GraphError::FailedToMutateGraph {
+                        source: MutateGraphError::IllegalVertexPropertyChange {
+                            vertex_id: v,
+                            source: IllegalMutate::from_source(err, &prop_name),
+                        },
+                    }
+                })?;
             }
+            Ok(())
+        } else {
+            Err(GraphError::FailedToMutateGraph {
+                source: MutateGraphError::VertexNotFoundError { vertex_id: v },
+            })
         }
-        Ok(())
     }
 
     pub(crate) fn add_edge_properties_internal(
@@ -437,14 +447,27 @@ impl<const N: usize> TemporalGraph<N> {
 
         let props = self
             .edge_meta
-            .resolve_prop_ids(props.clone(), true)
+            .resolve_prop_ids(props, true)
             .collect::<Vec<_>>();
 
         let mut layer = edge.layer_mut(layer_id);
-        for (prop_id, prop_name, prop) in props {
-            layer
-                .add_static_prop(prop_id, &prop_name, prop)
-                .expect("you should not fail!");
+        for (prop_id, prop) in props {
+            layer.add_static_prop(prop_id, prop).map_err(|err| {
+                GraphError::FailedToMutateGraph {
+                    source: MutateGraphError::IllegalEdgePropertyChange {
+                        src_id: src,
+                        dst_id: dst,
+                        source: IllegalMutate {
+                            name: self
[…]
