+++ python/tests/test_graphdb.py
@@ -1490,8 +1490,46 @@ def test_load_from_pandas():
                      (5, 6, 5.0, "purple")]
 
 
+def test_load_from_pandas_into_existing_graph():
+    edges_df = pd.DataFrame({
+        "src": [1, 2, 3, 4, 5],
+        "dst": [2, 3, 4, 5, 6],
+        "time": [1, 2, 3, 4, 5],
+        "weight": [1.0, 2.0, 3.0, 4.0, 5.0],
+        "marbles": ["red", "blue", "green", "yellow", "purple"]
+    })
+
+    vertices_df = pd.DataFrame({
+        "id": [1, 2, 3, 4, 5, 6],
+        "name": ["Alice", "Bob", "Carol", "Dave", "Eve", "Frank"],
+        "time": [1, 2, 3, 4, 5, 6],
+    })
+
+    g = Graph()
+
+    g.load_vertices_from_pandas(vertices_df, "id", "time", ["name"])
+
+    g.load_edges_from_pandas(edges_df, "src", "dst", "time", ["weight", "marbles"])
+
+    assert g.vertices().id().collect() == [1, 2, 3, 4, 5, 6]
+    edges = []
+    for e in g.edges():
+        weight = e["weight"]
+        marbles = e["marbles"]
+        edges.append((e.src().id(), e.dst().id(), weight, marbles))
+
+    assert edges == [(1, 2, 1.0, "red"), (2, 3, 2.0, "blue"), (3, 4, 3.0, "green"), (4, 5, 4.0, "yellow"),
+                     (5, 6, 5.0, "purple")]
+
+    vertices = []
+    for v in g.vertices():
+        name = v["name"]
+        vertices.append((v.id(), name))
+
+    assert vertices == [(1, "Alice"), (2, "Bob"), (3, "Carol"), (4, "Dave"), (5, "Eve"), (6, "Frank")]
+
+
 def test_load_from_pandas_vertices():
-    import pandas as pd
     edges_df = pd.DataFrame({
         "src": [1, 2, 3, 4, 5],
         "dst": [2, 3, 4, 5, 6],
@@ -1527,46 +1565,53 @@ def test_load_from_pandas_vertices():
     assert vertices == [(1, "Alice"), (2, "Bob"), (3, "Carol"), (4, "Dave"), (5, "Eve"), (6, "Frank")]
 
 
-def test_hits_algorithm():
-    g = graph_loader.lotr_graph()
-    assert algorithms.hits(g).get('Aldor') == (0.0035840950440615416, 0.007476256228983402)
-
-
-def load_from_pandas_into_existing_graph():
-    import pandas as pd
+def test_load_from_pandas_with_types():
     edges_df = pd.DataFrame({
         "src": [1, 2, 3, 4, 5],
         "dst": [2, 3, 4, 5, 6],
         "time": [1, 2, 3, 4, 5],
         "weight": [1.0, 2.0, 3.0, 4.0, 5.0],
-        "marbles": ["red", "blue", "green", "yellow", "purple"]
+        "marbles": ["red", "blue", "green", "yellow", "purple"],
+        "layers":  ["layer 1", "layer 2", "layer 3", "layer 4", "layer 5"]
     })
-
     vertices_df = pd.DataFrame({
         "id": [1, 2, 3, 4, 5, 6],
         "name": ["Alice", "Bob", "Carol", "Dave", "Eve", "Frank"],
         "time": [1, 2, 3, 4, 5, 6],
+        "types":  ["Person 1", "Person 2", "Person 3", "Person 4", "Person 5", "Person 6"]
     })
+    g = Graph()
+    g.load_vertices_from_pandas(vertices_df, "id", "time", ["name"],vertex_type="Person")
[…]
