+++ raphtory/src/core/storage/timeindex.rs
@@ -1,75 +1,142 @@
+use crate::{
+    core::{entities::LayerIds, utils::time::error::ParseTimeError},
+    db::api::mutation::{internal::InternalAdditionOps, InputTime, TryIntoInputTime},
+};
 use itertools::Itertools;
+use num_traits::Saturating;
 use serde::{Deserialize, Serialize};
 use std::{
     cmp::{max, min},
     collections::BTreeSet,
+    fmt::Debug,
     ops::Range,
     sync::Arc,
 };
 
-use crate::core::entities::LayerIds;
-
 use super::locked_view::LockedView;
 
-#[derive(Default, Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum TimeIndex {
-    #[default]
-    Empty,
-    One(i64),
-    Set(BTreeSet<i64>),
+#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Ord, PartialOrd, Eq)]
+pub struct TimeIndexEntry(i64, usize);
+
+pub trait AsTime: Debug + Copy + Ord + Eq + Send + Sync {
+    fn t(&self) -> &i64;
+    fn range(w: Range<i64>) -> Range<Self>;
 }
 
-impl TimeIndex {
-    pub fn is_empty(&self) -> bool {
-        match self {
-            TimeIndex::Empty => true,
-            _ => false,
-        }
+impl From<i64> for TimeIndexEntry {
+    fn from(value: i64) -> Self {
+        Self::start(value)
+    }
+}
+
+impl TimeIndexEntry {
+    pub fn new(t: i64, s: usize) -> Self {
+        Self(t, s)
     }
 
-    pub fn one(t: i64) -> Self {
-        Self::One(t)
+    pub fn from_input<G: InternalAdditionOps, T: TryIntoInputTime>(
+        g: &G,
+        t: T,
+    ) -> Result<Self, ParseTimeError> {
+        let t = t.try_into_input_time()?;
+        Ok(match t {
+            InputTime::Simple(t) => Self::new(t, g.next_event_id()),
+            InputTime::Indexed(t, s) => Self::new(t, s),
+        })
     }
-    pub fn insert(&mut self, t: i64) -> bool {
+
+    pub fn start(t: i64) -> Self {
+        Self(t, 0)
+    }
+
+    pub fn end(t: i64) -> Self {
+        Self(t.saturating_add(1), 0)
+    }
+}
+
+impl AsTime for i64 {
+    fn t(&self) -> &i64 {
+        self
+    }
+
+    fn range(w: Range<i64>) -> Range<Self> {
+        w
+    }
+}
+
+impl AsTime for TimeIndexEntry {
+    fn t(&self) -> &i64 {
+        &self.0
+    }
+    fn range(w: Range<i64>) -> Range<Self> {
+        Self::start(w.start)..Self::start(w.end)
+    }
+}
+
+#[derive(Default, Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub enum TimeIndex<T: Ord + Eq + Copy + Debug> {
+    #[default]
+    Empty,
+    One(T),
+    Set(BTreeSet<T>),
+}
+
+impl<T: AsTime> TimeIndex<T> {
+    pub fn is_empty(&self) -> bool {
+        matches!(self, TimeIndex::Empty)
+    }
+
+    pub fn one(ti: T) -> Self {
+        Self::One(ti)
+    }
+    pub fn insert(&mut self, ti: T) -> bool {
         match self {
             TimeIndex::Empty => {
-                *self = TimeIndex::One(t);
+                *self = TimeIndex::One(ti);
                 true
             }
             TimeIndex::One(t0) => {
-                if *t0 == t {
+                if t0 == &ti {
                     false
                 } else {
[…]
