+++ raphtory/src/core/entities/properties/tprop.rs
@@ -4,7 +4,7 @@ use crate::{
             properties::{props::DictMapper, tcell::TCell},
             LayerIds,
         },
-        storage::locked_view::LockedView,
+        storage::{locked_view::LockedView, timeindex::TimeIndexEntry},
         Prop,
     },
     db::graph::graph::Graph,
@@ -35,24 +35,24 @@ pub enum TProp {
 }
 
 impl TProp {
-    pub(crate) fn from(t: i64, prop: &Prop) -> Self {
+    pub(crate) fn from(t: TimeIndexEntry, prop: Prop) -> Self {
         match prop {
-            Prop::Str(value) => TProp::Str(TCell::new(t, value.to_string())),
-            Prop::I32(value) => TProp::I32(TCell::new(t, *value)),
-            Prop::I64(value) => TProp::I64(TCell::new(t, *value)),
-            Prop::U32(value) => TProp::U32(TCell::new(t, *value)),
-            Prop::U64(value) => TProp::U64(TCell::new(t, *value)),
-            Prop::F32(value) => TProp::F32(TCell::new(t, *value)),
-            Prop::F64(value) => TProp::F64(TCell::new(t, *value)),
-            Prop::Bool(value) => TProp::Bool(TCell::new(t, *value)),
-            Prop::DTime(value) => TProp::DTime(TCell::new(t, *value)),
-            Prop::Graph(value) => TProp::Graph(TCell::new(t, value.clone())),
-            Prop::List(value) => TProp::List(TCell::new(t, value.clone())),
-            Prop::Map(value) => TProp::Map(TCell::new(t, value.clone())),
+            Prop::Str(value) => TProp::Str(TCell::new(t, value)),
+            Prop::I32(value) => TProp::I32(TCell::new(t, value)),
+            Prop::I64(value) => TProp::I64(TCell::new(t, value)),
+            Prop::U32(value) => TProp::U32(TCell::new(t, value)),
+            Prop::U64(value) => TProp::U64(TCell::new(t, value)),
+            Prop::F32(value) => TProp::F32(TCell::new(t, value)),
+            Prop::F64(value) => TProp::F64(TCell::new(t, value)),
+            Prop::Bool(value) => TProp::Bool(TCell::new(t, value)),
+            Prop::DTime(value) => TProp::DTime(TCell::new(t, value)),
+            Prop::Graph(value) => TProp::Graph(TCell::new(t, value)),
+            Prop::List(value) => TProp::List(TCell::new(t, value)),
+            Prop::Map(value) => TProp::Map(TCell::new(t, value)),
         }
     }
 
-    pub(crate) fn set(&mut self, t: i64, prop: &Prop) {
+    pub(crate) fn set(&mut self, t: TimeIndexEntry, prop: Prop) {
         // TODO: we should bubble up the result if we change the type of the property
         match self {
             TProp::Empty => {
@@ -65,60 +65,78 @@ impl TProp {
             }
             TProp::I32(cell) => {
                 if let Prop::I32(a) = prop {
-                    cell.set(t, *a);
+                    cell.set(t, a);
                 }
             }
             TProp::I64(cell) => {
                 if let Prop::I64(a) = prop {
-                    cell.set(t, *a);
+                    cell.set(t, a);
                 }
             }
             TProp::U32(cell) => {
                 if let Prop::U32(a) = prop {
-                    cell.set(t, *a);
+                    cell.set(t, a);
                 }
             }
[…]
