### 2023-08-08 6ee372c
add Rust Benchmark (cargo) benchmark result for 747a943fb39b9167576204d98f5917fc96a1db14
