### 2023-08-08 5ed0ec1
Changed to static props
