### 2023-08-08 699e528
Added functions for adding constant props only
