+++ raphtory/src/python/graph/pandas.rs
@@ -1,5 +1,4 @@
 use std::collections::HashMap;
-use std::hash::Hash;
 use arrow2::{
     array::{Array, BooleanArray, PrimitiveArray, Utf8Array},
     ffi,
@@ -10,7 +9,6 @@ use pyo3::{
 };
 
 use crate::{core::utils::errors::GraphError, prelude::*};
-use crate::db::api::mutation::CollectProperties;
 
 fn i64_opt_into_u64_opt(x: Option<&i64>) -> Option<u64> {
     x.map(|x| (*x).try_into().unwrap())
@@ -204,6 +202,160 @@ pub(crate) fn load_edges_from_df<'a, S: AsRef<str>>(
     }
     Ok(())
 }
+
+
+pub(crate) fn load_vertex_props_from_df<'a>(
+    df: &'a PretendDF,
+    vertex_id: &str,
+    const_props: Option<Vec<&str>>,
+    shared_const_props: Option<HashMap<String,Prop>>,
+    graph: &Graph,
+) -> Result<(), GraphError> {
+
+    let const_prop_iter = const_props
+        .unwrap_or_default()
+        .into_iter()
+        .map(|name| lift_property(name, &df))
+        .reduce(combine_prop_iters)
+        .unwrap_or_else(|| Box::new(std::iter::repeat(vec![])));
+
+    if let Some(vertex_id) = df.iter_col::<u64>(vertex_id)
+    {
+        let iter = vertex_id.map(|i| i.copied());
+        for (vertex_id, const_props) in iter.zip(const_prop_iter) {
+            if let Some(vertex_id) = vertex_id {
+                graph.add_vertex_properties(vertex_id, const_props)?;
+                if let Some(shared_const_props) = &shared_const_props {
+                    graph.add_vertex_properties(vertex_id, shared_const_props.iter())?;
+                }
+            }
+        }
+    } else if let Some(vertex_id) = df.iter_col::<i64>(vertex_id)
+    {
+        let iter = vertex_id.map(i64_opt_into_u64_opt);
+        for (vertex_id, const_props) in iter.zip(const_prop_iter) {
+            if let Some(vertex_id) = vertex_id {
+                graph.add_vertex_properties(vertex_id, const_props)?;
+                if let Some(shared_const_props) = &shared_const_props {
+                    graph.add_vertex_properties(vertex_id, shared_const_props.iter())?;
+                }
+            }
+        }
+
+    } else if let Some(vertex_id) = df.utf8::<i32>(vertex_id)
+    {
+        let iter = vertex_id.into_iter();
+        for (vertex_id, const_props) in iter.zip(const_prop_iter) {
+            if let Some(vertex_id)= vertex_id {
+                graph.add_vertex_properties(vertex_id, const_props)?;
+                if let Some(shared_const_props) = &shared_const_props {
+                    graph.add_vertex_properties(vertex_id, shared_const_props.iter())?;
+                }
+            }
+        }
+    } else if let Some(vertex_id) = df.utf8::<i64>(vertex_id)
+    {
+        let iter = vertex_id.into_iter();
+        for (vertex_id, const_props) in iter.zip(const_prop_iter) {
+            if let Some(vertex_id) = vertex_id {
+                graph.add_vertex_properties(vertex_id, const_props)?;
+                if let Some(shared_const_props) = &shared_const_props {
+                    graph.add_vertex_properties(vertex_id, shared_const_props.iter())?;
+                }
+            }
+        }
+    } else {
+        return Err(GraphError::LoadFailure(
+            "vertex id column must be either u64 or text, time column must be i64".to_string(),
+        ));
+    }
+    Ok(())
+}
+
+pub(crate) fn load_edges_props_from_df<'a, S: AsRef<str>>(
+    df: &'a PretendDF,
[…]
