+++ raphtory/src/python/graph/pandas.rs
@@ -1,4 +1,3 @@
-use std::collections::HashMap;
 use arrow2::{
     array::{Array, BooleanArray, PrimitiveArray, Utf8Array},
     ffi,
@@ -7,6 +6,7 @@ use arrow2::{
 use pyo3::{
     create_exception, exceptions::PyException, ffi::Py_uintptr_t, prelude::*, types::PyDict,
 };
+use std::collections::HashMap;
 
 use crate::{core::utils::errors::GraphError, prelude::*};
 
@@ -53,7 +53,7 @@ pub(crate) fn load_vertices_from_df<'a>(
     time: &str,
     props: Option<Vec<&str>>,
     const_props: Option<Vec<&str>>,
-    shared_const_props: Option<HashMap<String,Prop>>,
+    shared_const_props: Option<HashMap<String, Prop>>,
     graph: &Graph,
 ) -> Result<(), GraphError> {
     let prop_iter = props
@@ -70,17 +70,15 @@ pub(crate) fn load_vertices_from_df<'a>(
         .reduce(combine_prop_iters)
         .unwrap_or_else(|| Box::new(std::iter::repeat(vec![])));
 
-
-
     if let (Some(vertex_id), Some(time)) = (df.iter_col::<u64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(|i| i.copied()).zip(time);
-        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter,shared_const_props)?;
+        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter, shared_const_props)?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.iter_col::<i64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(i64_opt_into_u64_opt).zip(time);
-        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter,shared_const_props)?;
+        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter, shared_const_props)?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.utf8::<i32>(vertex_id), df.iter_col::<i64>(time))
     {
@@ -123,7 +121,7 @@ pub(crate) fn load_edges_from_df<'a, S: AsRef<str>>(
     time: &str,
     props: Option<Vec<&str>>,
     const_props: Option<Vec<&str>>,
-    shared_const_props: Option<HashMap<String,Prop>>,
+    shared_const_props: Option<HashMap<String, Prop>>,
     layer: Option<S>,
     layer_in_df: Option<S>,
     graph: &Graph,
@@ -153,7 +151,14 @@ pub(crate) fn load_edges_from_df<'a, S: AsRef<str>>(
             .map(|i| i.copied())
             .zip(dst.map(|i| i.copied()))
             .zip(time);
-        load_edges_from_num_iter(&graph, triplets, prop_iter, const_prop_iter, shared_const_props, layer)?;
+        load_edges_from_num_iter(
+            &graph,
+            triplets,
+            prop_iter,
+            const_prop_iter,
+            shared_const_props,
+            layer,
+        )?;
     } else if let (Some(src), Some(dst), Some(time)) = (
         df.iter_col::<i64>(src),
         df.iter_col::<i64>(dst),
@@ -163,19 +168,33 @@ pub(crate) fn load_edges_from_df<'a, S: AsRef<str>>(
             .map(i64_opt_into_u64_opt)
             .zip(dst.map(i64_opt_into_u64_opt))
             .zip(time);
-        load_edges_from_num_iter(&graph, triplets, prop_iter, const_prop_iter,shared_const_props, layer)?;
[…]
