### 2023-08-08 a433850
added tests for static props
