+++ raphtory/src/search/mod.rs
@@ -27,6 +27,7 @@ use crate::{
     },
     prelude::*,
 };
+use crate::search::fields::{EDGE_ID, VERTEX_ID};
 
 #[derive(Clone)]
 pub struct IndexedGraph<G> {
