+++ python/src/lib.rs
@@ -9,8 +9,8 @@ use raphtory_core::python::{
         edge::{PyEdge, PyEdges},
         graph::PyGraph,
         graph_with_deletions::PyGraphWithDeletions,
-        properties::{PyConstProperties, PyProperties, PyTemporalProp, PyTemporalProperties},
         index::GraphIndex,
+        properties::{PyConstProperties, PyProperties, PyTemporalProp, PyTemporalProperties},
         vertex::{PyVertex, PyVertices},
     },
     packages::{algorithms::*, graph_gen::*, graph_loader::*},
