### 2023-08-09 fb25726
add reproducing python test
