### 2023-08-09 2b13908
Bug/unsafe layer panic (#1166)
* add reproducing python test

* add reproducing rust test

* remove unsafe_layer completely and fix constant property retrieval for edges
