### 2023-08-09 5c1e38d
add Rust Benchmark (cargo) benchmark result for 47283e2db552e913566a082ca3b6d2d1177f72ce
