### 2023-08-09 a7d5b0c
Added constant props to pandas loaders (#1165)
* Added vertex type arguments to the pandas loader and tests for layers/types in pandas loader

* Changed to static props

* Added functions for adding constant props only

* fmt

* added tests for static props

* Finalised tests
