+++ raphtory/src/python/graph/pandas.rs
@@ -6,6 +6,7 @@ use arrow2::{
 use pyo3::{
     create_exception, exceptions::PyException, ffi::Py_uintptr_t, prelude::*, types::PyDict,
 };
+use std::collections::HashMap;
 
 use crate::{core::utils::errors::GraphError, prelude::*};
 
@@ -51,6 +52,8 @@ pub(crate) fn load_vertices_from_df<'a>(
     vertex_id: &str,
     time: &str,
     props: Option<Vec<&str>>,
+    const_props: Option<Vec<&str>>,
+    shared_const_props: Option<HashMap<String, Prop>>,
     graph: &Graph,
 ) -> Result<(), GraphError> {
     let prop_iter = props
@@ -60,31 +63,46 @@ pub(crate) fn load_vertices_from_df<'a>(
         .reduce(combine_prop_iters)
         .unwrap_or_else(|| Box::new(std::iter::repeat(vec![])));
 
+    let const_prop_iter = const_props
+        .unwrap_or_default()
+        .into_iter()
+        .map(|name| lift_property(name, &df))
+        .reduce(combine_prop_iters)
+        .unwrap_or_else(|| Box::new(std::iter::repeat(vec![])));
+
     if let (Some(vertex_id), Some(time)) = (df.iter_col::<u64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(|i| i.copied()).zip(time);
-        load_vertices_from_num_iter(graph, iter, prop_iter)?;
+        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter, shared_const_props)?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.iter_col::<i64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(i64_opt_into_u64_opt).zip(time);
-        load_vertices_from_num_iter(graph, iter, prop_iter)?;
+        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter, shared_const_props)?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.utf8::<i32>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.into_iter().zip(time);
-        for ((vertex_id, time), props) in iter.zip(prop_iter) {
+        for (((vertex_id, time), props), const_props) in iter.zip(prop_iter).zip(const_prop_iter) {
             if let (Some(vertex_id), Some(time)) = (vertex_id, time) {
                 graph.add_vertex(*time, vertex_id, props)?;
+                graph.add_vertex_properties(vertex_id, const_props)?;
+                if let Some(shared_const_props) = &shared_const_props {
+                    graph.add_vertex_properties(vertex_id, shared_const_props.iter())?;
+                }
             }
         }
     } else if let (Some(vertex_id), Some(time)) =
         (df.utf8::<i64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.into_iter().zip(time);
-        for ((vertex_id, time), props) in iter.zip(prop_iter) {
+        for (((vertex_id, time), props), const_props) in iter.zip(prop_iter).zip(const_prop_iter) {
             if let (Some(vertex_id), Some(time)) = (vertex_id, time) {
                 graph.add_vertex(*time, vertex_id, props)?;
+                graph.add_vertex_properties(vertex_id, const_props)?;
+                if let Some(shared_const_props) = &shared_const_props {
+                    graph.add_vertex_properties(vertex_id, shared_const_props)?;
+                }
             }
         }
     } else {
[…]
