### 2023-08-09 9d2d6fa
add Rust Benchmark (cargo) benchmark result for a105bafe3f600b45546932d5d625411b3a89e2c5
