### 2023-08-09 d216323
add Rust Benchmark (cargo) benchmark result for 1a0218901bd7ae89415d4107f5ec2ddcf88aad8b
