### 2023-08-09 863ccdf
add Rust Benchmark (cargo) benchmark result for 3ad0c02eef15dab8d44ea5fac0c6989b55c61475
