### 2023-08-09 56f4676
add Rust Benchmark (cargo) benchmark result for f9bff518105a7373ce6441097ab68812723cd131
