+++ raphtory/src/algorithms/motifs/three_node_eval_wip.rs
@@ -1,521 +1,0 @@
-use crate::{
-    algorithms::motifs::three_node_motifs::*,
-    core::state::{
-        accumulator_id::{accumulators::val, AccId},
-        agg::ValDef,
-        compute_state::ComputeStateVec,
-    },
-    db::{
-        api::view::{GraphViewOps, VertexViewOps, *},
-        task::{
-            context::Context,
-            task::{ATask, Job, Step},
-            task_runner::TaskRunner,
-            vertex::eval_vertex::EvalVertexView,
-        },
-    },
-};
-use num_traits::Zero;
-use std::{collections::HashMap, ops::Add, slice::Iter};
-
-pub fn star_motif_count<G: GraphViewOps>(
-    evv: &EvalVertexView<G, ComputeStateVec, MotifCounter>,
-    delta: i64,
-) -> [usize; 24] {
-    let neigh_map: HashMap<u64, usize> = evv
-        .neighbours()
-        .into_iter()
-        .enumerate()
-        .map(|(num, nb)| (nb.id(), num))
-        .into_iter()
-        .collect();
-    let mut exploded_edges = evv
-        .edges()
-        .explode()
-        .map(|edge| {
-            if edge.src().id() == evv.id() {
-                star_event(neigh_map[&edge.dst().id()], 1, edge.time().unwrap())
-            } else {
-                star_event(neigh_map[&edge.src().id()], 0, edge.time().unwrap())
-            }
-        })
-        .collect::<Vec<StarEvent>>();
-    exploded_edges.sort_by_key(|e| e.time);
-    let mut star_count = init_star_count(neigh_map.len());
-    star_count.execute(&exploded_edges, delta);
-    star_count.return_counts()
-}
-
-pub fn twonode_motif_count<G: GraphViewOps>(
-    graph: &G,
-    evv: &EvalVertexView<G, ComputeStateVec, MotifCounter>,
-    delta: i64,
-) -> [usize; 8] {
-    let mut counts = [0; 8];
-    for nb in evv.neighbours().into_iter() {
-        let nb_id = nb.id();
-        let out = graph.edge(evv.id(), nb_id, None);
-        let inc = graph.edge(nb_id, evv.id(), None);
-        let mut all_exploded = match (out, inc) {
-            (Some(o), Some(i)) => o
-                .explode()
-                .chain(i.explode())
-                .map(|e| {
-                    two_node_event(
-                        if e.src().id() == evv.id() { 1 } else { 0 },
-                        e.time().unwrap(),
-                    )
-                })
-                .collect::<Vec<TwoNodeEvent>>(),
-            (Some(o), None) => o
-                .explode()
-                .map(|e| two_node_event(1, e.time().unwrap()))
-                .collect::<Vec<TwoNodeEvent>>(),
-            (None, Some(i)) => i
-                .explode()
-                .map(|e| two_node_event(0, e.time().unwrap()))
-                .collect::<Vec<TwoNodeEvent>>(),
-            (None, None) => Vec::new(),
-        };
-        all_exploded.sort_by_key(|e| e.time);
-        let mut two_node_counter = init_two_node_count();
-        two_node_counter.execute(&all_exploded, delta);
-        let two_node_result = two_node_counter.return_counts();
-        for i in 0..8 {
-            counts[i] += two_node_result[i];
-        }
-    }
-    counts
-}
-
-pub fn triangle_motif_count<G: GraphViewOps>(
-    graph: &G,
-    evv: &EvalVertexView<G, ComputeStateVec, MotifCounter>,
-    delta: i64,
[…]
