### 2023-08-09 da345d6
update 2 new edge api
