### 2023-08-09 1aa7d59
add mini optimisation
