+++ raphtory/src/algorithms/motifs/mod.rs
@@ -1,3 +1,3 @@
 pub mod three_node_local_single_thread;
-pub mod three_node_temporal_motifs;
 pub mod three_node_motifs;
+pub mod three_node_temporal_motifs;
