### 2023-08-09 1e36192
fn for returning vertex subgraph as well as the set
