### 2023-08-10 #5633 @narnolddd
Just checking the pdf issue is now resolved @luizirber? Thanks.
