### 2023-08-10 b8ab8fb
Store graphs behind a RwLock and add mutation to read a new directory
