### 2023-08-10 84bd9b0
cleanup and add tests
