+++ raphtory-graphql/src/lib.rs
@@ -10,9 +10,12 @@ mod data;
 #[cfg(test)]
 mod graphql_test {
     use super::*;
-    use dynamic_graphql::{dynamic::DynamicRequestExt, App, FieldValue};
+    use crate::{data::Data, model::App};
+    use dynamic_graphql::Request;
     use raphtory::{db::api::view::internal::IntoDynamic, prelude::*};
+    use serde_json::json;
     use std::collections::HashMap;
+    use tempfile::tempdir;
 
     #[tokio::test]
     async fn search_for_gandalf_query() {
@@ -24,11 +27,8 @@ mod graphql_test {
             .add_vertex(0, "Frodo", [("kind".to_string(), Prop::str("Hobbit"))])
             .expect("Could not add vertex!");
 
-        let graphs = HashMap::from([("lotr".to_string(), graph.into_dynamic().into())]);
-        let data = data::Data { graphs };
-
-        #[derive(App)]
-        struct App(model::QueryRoot);
+        let graphs = HashMap::from([("lotr".to_string(), graph.into_dynamic())]);
+        let data = data::Data::from_map(graphs);
         let schema = App::create_schema().data(data).finish().unwrap();
 
         let query = r#"
@@ -40,16 +40,13 @@ mod graphql_test {
           }
         }
         "#;
-
-        let root = model::QueryRoot;
-        let req = dynamic_graphql::Request::new(query).root_value(FieldValue::owned_any(root));
-
+        let req = Request::new(query);
         let res = schema.execute(req).await;
         let data = res.data.into_json().unwrap();
 
         assert_eq!(
             data,
-            serde_json::json!({
+            json!({
                 "graph": {
                     "search": [
                         {
@@ -68,11 +65,9 @@ mod graphql_test {
             .add_vertex(0, 11, NO_PROPS)
             .expect("Could not add vertex!");
 
-        let graphs = HashMap::from([("lotr".to_string(), graph.into_dynamic().into())]);
-        let data = data::Data { graphs };
+        let graphs = HashMap::from([("lotr".to_string(), graph.into_dynamic())]);
+        let data = data::Data::from_map(graphs);
 
-        #[derive(App)]
-        struct App(model::QueryRoot);
         let schema = App::create_schema().data(data).finish().unwrap();
 
         let query = r#"
@@ -84,10 +79,7 @@ mod graphql_test {
           }
         }
         "#;
-
-        let root = model::QueryRoot;
-        let req = dynamic_graphql::Request::new(query).root_value(FieldValue::owned_any(root));
-
+        let req = Request::new(query);
         let res = schema.execute(req).await;
         let data = res.data.into_json().unwrap();
 
@@ -118,11 +110,9 @@ mod graphql_test {
             panic!("Could not add vertex! {:?}", err);
         }
 
-        let graphs = HashMap::from([("lotr".to_string(), graph.into_dynamic().into())]);
-        let data = data::Data { graphs };
+        let graphs = HashMap::from([("lotr".to_string(), graph.into_dynamic())]);
+        let data = Data::from_map(graphs);
 
-        #[derive(App)]
-        struct App(model::QueryRoot);
         let schema = App::create_schema().data(data).finish().unwrap();
 
         let gandalf_query = r#"
@@ -135,16 +125,13 @@ mod graphql_test {
         }
         "#;
 
-        let root = model::QueryRoot;
-        let req =
-            dynamic_graphql::Request::new(gandalf_query).root_value(FieldValue::owned_any(root));
-
+        let req = Request::new(gandalf_query);
         let res = schema.execute(req).await;
         let data = res.data.into_json().unwrap();
 
         assert_eq!(
             data,
-            serde_json::json!({
+            json!({
                 "graph": {
                     "nodes": [
[…]
