+++ raphtory-graphql/Cargo.toml
@@ -41,4 +41,5 @@ ordered-float = "3.7.0"
 
 [dev-dependencies]
 serde_json = "1.0"
+tempfile = "3.6.0"
 
