### 2023-08-11 1be01a7
start work on supporting sending graphs over graphql
