### 2023-08-11 a19646d
show what the path looks like for debugging
