### 2023-08-11 d307541
print query and response
