### 2023-08-11 5100f58
simplification of map filter
