### 2023-08-11 a9e4a5d
Feature kcore (#1123)
* add k core algorithm ahead of triangle optimisations

* optimisations to triangle_count

* something not right

* add test for triangle count on subgraph

* fix the issue with compute and subgraphs

* add k core algorithm ahead of triangle optimisations

* optimisations to triangle_count

* something not right

* add test

* triangle motifs first run

* almost working just got a borrow issue

* fixed borrow prob

* global works, local is mega busted

* test 4 global

* working version up to events happening at same time

* try fix fmt issue

* formatting changes

* fix 
[…]
