+++ raphtory/src/algorithms/k_core.rs
@@ -1,0 +1,167 @@
+use crate::{
+    core::{entities::VID, state::compute_state::ComputeStateVec},
+    db::{
+        api::view::{GraphViewOps, VertexViewOps},
+        graph::views::vertex_subgraph::VertexSubgraph,
+        task::{
+            context::Context,
+            task::{ATask, Job, Step},
+            task_runner::TaskRunner,
+            vertex::eval_vertex::EvalVertexView,
+        },
+    },
+};
+use std::{
+    cmp,
+    collections::{HashMap, HashSet},
+};
+
+#[derive(Clone, Debug)]
+struct KCoreState {
+    alive: bool,
+}
+
+impl KCoreState {
+    fn new() -> Self {
+        Self { alive: true }
+    }
+}
+
+/// Determines which nodes are in the k-core for a given value of k
+///
+/// # Arguments
+///
+/// * `g` - A reference to the graph
+/// * `k` - Value of k such that the returned vertices have degree > k (recursively)
+/// * `iter_count` - The number of iterations to run
+/// * `threads` - number of threads to run on
+///
+/// # Returns
+///
+/// A hash set of vertices in the k core
+///
+pub fn k_core_set<G>(graph: &G, k: usize, iter_count: usize, threads: Option<usize>) -> HashSet<VID>
+where
+    G: GraphViewOps,
+{
+    let ctx: Context<G, ComputeStateVec> = graph.into();
+
+    let step1 = ATask::new(move |vv| {
+        let deg = vv.degree();
+        let state: &mut KCoreState = vv.get_mut();
+        state.alive = deg >= k;
+        Step::Continue
+    });
+
+    let step2 = ATask::new(
+        move |vv: &mut EvalVertexView<'_, G, ComputeStateVec, KCoreState>| {
+            let prev: bool = vv.prev().alive;
+            if prev == true {
+                let current = vv
+                    .neighbours()
+                    .into_iter()
+                    .filter(|n| n.prev().alive)
+                    .count()
+                    >= k;
+                let state: &mut KCoreState = vv.get_mut();
+                if current != prev {
+                    state.alive = current;
+                    Step::Continue
+                } else {
+                    Step::Done
+                }
+            } else {
+                Step::Done
+            }
+        },
+    );
+
+    let mut runner: TaskRunner<G, _> = TaskRunner::new(ctx);
+
+    runner.run(
+        vec![Job::new(step1)],
+        vec![Job::read_only(step2)],
+        KCoreState::new(),
+        |_, _, _, local| {
+            local
+                .iter()
+                .enumerate()
+                .filter(|(_, state)| state.alive)
+                .map(|(v_ref, _)| v_ref.into())
+                .collect::<HashSet<VID>>()
+        },
+        threads,
+        iter_count,
+        None,
+        None,
+    )
+}
+
+pub fn k_core<G>(
+    graph: &G,
+    k: usize,
+    iter_count: usize,
+    threads: Option<usize>,
+) -> VertexSubgraph<G>
+where
+    G: GraphViewOps,
+{
+    let v_set = k_core_set(graph, k, iter_count, threads);
+    graph.subgraph(v_set)
+}
+
+mod k_core_test {
+    use std::collections::HashSet;
+
+    use crate::{
+        algorithms::{k_core::k_core_set, triangle_count::triangle_count},
+        prelude::*,
+    };
+
+    #[test]
+    fn k_core_2() {
+        let graph = Graph::new();
+
+        let edges = vec![
+            (1, 2, 1),
+            (1, 3, 2),
+            (1, 4, 3),
+            (3, 1, 4),
+            (3, 4, 5),
[…]
