### 2023-08-11 8588e55
add Rust Benchmark (cargo) benchmark result for 057df89c4682356d8ccfbe33cbbe84a40949d1dd
