### 2023-08-11 6699ea3
Make window etc. accept i64 if it is positive
