### 2023-08-11 8edb72f
fix Interval conversions
