+++ README.md
@@ -116,6 +116,110 @@ The weight of the edge between Bob and Charlie has changed by 20.0 pts
 The top node in the graph is Charlie with a score of 0.4744116163405977
 ```
 
+## GraphQL
+
+### Create/Load a graph
+
+Save a raphtory graph and set the `GRAPH_DIRECTORY` to point to the directory containing the graph.
+
+Alternatively you can run the example below to generate a graph.
+
+<details>
+
+```bash
+mkdir -p /tmp/graphs
+mkdir -p examples/rust/src/bin/lotr/data/
+tail -n +2 resource/lotr.csv > examples/rust/src/bin/lotr/data/lotr.csv
+
+cd examples/rust && cargo run --bin lotr -r
+
+cp examples/rust/src/bin/lotr/data/graphdb.bincode /tmp/graphs/lotr.bincode
+```
+
+</details>
+
+
+### Run the GraphQL server
+
+The code below will run graphql with a UI at port `1736` 
+
+GraphlQL will look for graph files in `/tmp/graphs` or in the path set in the `GRAPH_DIRECTORY` Environment variable. 
+
+```bash
+cd raphtory-graphql && cargo run -r 
+```
+
+<details>
+<summary>Following output upon a successful launch</summary>
+
+```bash
+warning: `raphtory` (lib) generated 17 warnings (run `cargo fix --lib -p raphtory` to apply 13 suggestions)
+    Finished release [optimized] target(s) in 0.91s
+     Running `/Users/haaroony/Documents/dev/UI/Raphtory/target/release/raphtory-graphql`
+loading graph from /tmp/graphs/lotr.bincode
+Playground: http://localhost:1736
+  2023-08-11T14:36:52.444203Z  INFO poem::server: listening, addr: socket://0.0.0.0:1736
+    at /Users/pometry/.cargo/registry/src/github.com-1ecc6299db9ec823/poem-1.3.56/src/server.rs:109
+
+  2023-08-11T14:36:52.444257Z  INFO poem::server: server started
+    at /Users/pometry/.cargo/registry/src/github.com-1ecc6299db9ec823/poem-1.3.56/src/server.rs:111
+```
+</details>
+
+
+### Execute a query
+
+Go to the Playground at `http://localhost:1736` and execute the following commands:
+
+Query:
+```bash
+    query GetNodes($graphName: String!) {
+        graph(name: $graphName) {
+            nodes {
+              name
+            }
+      }
+    }
+```
+
+Query Variables:
+```bash
+{
+  "graphName": "lotr.bincode"
+}
+```
+
+Expected Result:
+```bash
+{
+  "data": {
+    "graph": {
+      "nodes": [
+        {
+          "name": "Gandalf"
+        },
+        {
+          "name": "Elrond"
+        },
+        {
+          "name": "Frodo"
+        },
+        {
+          "name": "Bilbo"
+        },
+        ...
+```
+
+Screenshot
+
+<details>
+  <summary>GraphQL Server Screenshot</summary>
+  
+  ![GraphQL Query](resources/graphql-demo.jpg)
+  
+</details>
+
+
 
 ## Installing Raphtory 
 
