### 2023-08-11 4debc28
graph ql demo
