### 2023-08-11 6727646
add Rust Benchmark (cargo) benchmark result for 09ae218d2298e006f21d0e57b7dde809525846a1
