### 2023-08-11 3475e7f
add graph schema to graphql
