+++ raphtory-graphql/src/model/graph/graph.rs
@@ -3,6 +3,7 @@ use std::{
     ops::Deref,
 };
 
+use crate::model::schema::graph_schema::GraphSchema;
 use crate::model::{
     algorithm::Algorithms,
     filters::{edgefilter::EdgeFilter, nodefilter::NodeFilter},
@@ -84,6 +85,10 @@ impl GqlGraph {
         w.into()
     }
 
+    async fn layer_names(&self) -> Vec<String> {
+        self.graph.get_unique_layers()
+    }
+
     async fn static_properties(&self) -> Vec<Property> {
         self.graph
             .properties()
@@ -104,6 +109,10 @@ impl GqlGraph {
                 .collect(),
             None => self.graph.vertices().iter().map(|vv| vv.into()).collect(),
         }
+    }
+
+    async fn schema(&self) -> GraphSchema {
+        GraphSchema::new(&self.graph)
     }
 
     async fn search(&self, query: String, limit: usize, offset: usize) -> Vec<Node> {
