### 2023-08-14 d43e605
Bug/performance (#1169)
* fix performance regression in `active` method

* Make window etc. accept i64 if it is positive

* add graph ops benchmark on lotr to base so we catch performance regressions earlier next time

* fix Interval conversions
