### 2023-08-14 c8017a7
add Rust Benchmark (cargo) benchmark result for c5a01c23916799de260f6ef96838484e4eaa72d6
