### 2023-08-14 8901402
add Rust Benchmark (cargo) benchmark result for fcbf0175dc5bce688dc9d1785ed71e4b53032fb3
