### 2023-08-14 9c9a9d8
always inline base
