### 2023-08-14 c6b5656
add None option to edge_refs
