### 2023-08-14 824fc86
optimisations for degree
