### 2023-08-14 1a53d75
optimise find_edge
