### 2023-08-15 4b734e0
attempted optimisation of edgerefs
