### 2023-08-15 a7dc054
clean up the printlns
