### 2023-08-15 d89a62e
Make GraphQL server mutable (#1168)
* Store graphs behind a RwLock and add mutation to read a new directory

* add docs

* Add mutation that does not overwrite existing graphs

* cleanup and add tests

* show what the path looks like for debugging

* print query and response

* test escaping backslash in path

* clean up the printlns
