### 2023-08-15 51e5002
add Rust Benchmark (cargo) benchmark result for 2bb06bb27fa46ce628d68957bdae70f9a3ed1145
