+++ raphtory/src/core/storage/mod.rs
@@ -29,11 +29,11 @@ fn resolve<const N: usize>(index: usize) -> (usize, usize) {
 }
 
 #[derive(Debug, Serialize, Deserialize)]
-pub struct LockVec<T> {
-    data: Arc<parking_lot::RwLock<Vec<Option<T>>>>,
+pub struct LockVec<T: Default> {
+    data: Arc<parking_lot::RwLock<Vec<T>>>,
 }
 
-impl<T: PartialEq> PartialEq for LockVec<T> {
+impl<T: PartialEq + Default> PartialEq for LockVec<T> {
     fn eq(&self, other: &Self) -> bool {
         let a = self.data.read();
         let b = other.data.read();
@@ -41,27 +41,25 @@ impl<T: PartialEq> PartialEq for LockVec<T> {
     }
 }
 
-impl<T> LockVec<T> {
+impl<T: Default> LockVec<T> {
     pub fn new() -> Self {
         Self {
             data: Arc::new(parking_lot::RwLock::new(Vec::new())),
         }
     }
 
-    pub fn read_arc_lock(
-        &self,
-    ) -> lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<Option<T>>> {
+    pub fn read_arc_lock(&self) -> lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<T>> {
         RwLock::read_arc(&self.data)
     }
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct RawStorage<T, const N: usize> {
+pub struct RawStorage<T: Default, const N: usize> {
     pub(crate) data: Box<[LockVec<T>]>,
     len: AtomicUsize,
 }
 
-impl<T: PartialEq, const N: usize> PartialEq for RawStorage<T, N> {
+impl<T: PartialEq + Default, const N: usize> PartialEq for RawStorage<T, N> {
     fn eq(&self, other: &Self) -> bool {
         self.data.eq(&other.data)
     }
@@ -69,7 +67,7 @@ impl<T: PartialEq, const N: usize> PartialEq for RawStorage<T, N> {
 
 #[derive(Debug)]
 pub struct ReadLockedStorage<T, const N: usize> {
-    locks: Box<[lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<Option<T>>>; N]>,
+    locks: Box<[lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<T>>; N]>,
     len: usize,
 }
 
@@ -77,19 +75,17 @@ impl<T, const N: usize> ReadLockedStorage<T, N> {
     pub(crate) fn get(&self, index: usize) -> &T {
         let (bucket, offset) = resolve::<N>(index);
         let bucket = &self.locks[bucket];
-        bucket[offset].as_ref().unwrap()
+        &bucket[offset]
     }
 
     pub(crate) fn iter(&self) -> impl Iterator<Item = &T> + '_ {
-        self.locks
-            .iter()
-            .flat_map(|v| v.iter().map(|v| v.as_ref().unwrap()))
+        self.locks.iter().flat_map(|v| v.iter())
     }
 }
 
-impl<T, const N: usize> RawStorage<T, N> {
+impl<T: Default, const N: usize> RawStorage<T, N> {
     pub fn read_lock(&self) -> ReadLockedStorage<T, N> {
-        let guards: [lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<Option<T>>>; N] =
+        let guards: [lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<T>>; N] =
             std::array::from_fn(|i| self.data[i].read_arc_lock());
         ReadLockedStorage {
             locks: guards.into(),
[…]
