### 2023-08-15 2c9a9ed
randomly trying things to get it to run quicker
