+++ raphtory/src/core/entities/graph/tgraph_storage.rs
@@ -7,8 +7,12 @@ use crate::core::{
     storage::{self, ArcEntry, Entry, EntryMut, PairEntryMut},
     Direction,
 };
+use rayon::prelude::{ParallelBridge, ParallelIterator};
 use serde::{Deserialize, Serialize};
-use std::{ops::Deref, sync::Arc};
+use std::{
+    ops::{Deref, Range},
+    sync::Arc,
+};
 
 #[derive(Debug, Deserialize, Serialize, PartialEq)]
 pub(crate) struct GraphStorage<const N: usize> {
@@ -57,6 +61,7 @@ impl<const N: usize> GraphStorage<N> {
         self.edges.entry_arc(id)
     }
 
+    #[inline]
     pub(crate) fn get_edge(&self, id: usize) -> Entry<'_, EdgeStore<N>, N> {
         self.edges.entry(id)
     }
@@ -74,6 +79,14 @@ impl<const N: usize> GraphStorage<N> {
             LayerIds::All => self.edges.len(),
             _ => self.edges.iter().filter(|e| e.has_layer(&layers)).count(),
         }
+    }
+
+    pub(crate) fn edges_window_len(&self, layers: LayerIds, w: Range<i64>) -> usize {
+        self.edges
+            .iter()
+            .par_bridge()
+            .filter(|e| e.active(&layers, w.clone()))
+            .count()
     }
 
     fn lock(&self) -> LockedGraphStorage<N> {
