+++ raphtory/src/db/internal/graph_ops.rs
@@ -118,4 +118,8 @@ impl<const N: usize> GraphOps for InnerTemporalGraph<N> {
 
         Box::new(iter.into_iter())
     }
+
+    fn edges_len_window(&self, t_start: i64, t_end: i64, layers: LayerIds) -> usize {
+        self.inner().internal_num_edges_window(layers, t_start .. t_end)
+    }
 }
