### 2023-08-15 547e904
randomly trying things to get it to run quicker
