+++ raphtory/src/core/entities/graph/tgraph.rs
@@ -38,7 +38,7 @@ use serde::{Deserialize, Serialize};
 use std::{
     fmt::Debug,
     hash::BuildHasherDefault,
-    ops::Deref,
+    ops::{Deref, Range},
     path::Path,
     sync::{atomic::AtomicUsize, Arc},
 };
@@ -230,7 +230,7 @@ impl<const N: usize> TemporalGraph<N> {
 
     pub(crate) fn global_vertex_id(&self, v: VID) -> Option<u64> {
         let node = self.storage.get_node(v.into());
-        node.value().map(|n| n.global_id())
+        Some(node.value().global_id())
     }
 
     pub(crate) fn vertex_name(&self, v: VID) -> String {
@@ -292,6 +292,10 @@ impl<const N: usize> TemporalGraph<N> {
 
     pub(crate) fn internal_num_edges(&self, layers: LayerIds) -> usize {
         self.storage.edges_len(layers)
+    }
+
+    pub(crate) fn internal_num_edges_window(&self, layers:LayerIds, w: Range<i64>) -> usize{
+        self.storage.edges_window_len(layers, w)
     }
 
     pub(crate) fn degree(&self, v: VID, dir: Direction, layers: LayerIds) -> usize {
@@ -660,6 +664,7 @@ impl<const N: usize> TemporalGraph<N> {
         ArcEdge::from_entry(edge, self.edge_meta.clone())
     }
 
+    #[inline]
     pub(crate) fn edge(&self, e: EID) -> EdgeView<N> {
         let edge = self.storage.get_edge(e.into());
         EdgeView::from_entry(edge, self)

[…]
