### 2023-08-15 ab12f36
make num_edges parallel
