### 2023-08-15 4bb7126
make num_edges parallel
