+++ raphtory/src/core/entities/graph/tgraph_storage.rs
@@ -82,11 +82,14 @@ impl<const N: usize> GraphStorage<N> {
     }
 
     pub(crate) fn edges_window_len(&self, layers: LayerIds, w: Range<i64>) -> usize {
-        self.edges
-            .iter()
-            .par_bridge()
-            .filter(|e| e.active(&layers, w.clone()))
-            .count()
+        self.edges.count_with_filter(|e|{
+            e.active(&layers, w.clone())
+        })
+        // self.edges
+        //     .iter()
+        //     .par_bridge()
+        //     .filter(|e| e.active(&layers, w.clone()))
+        //     .count()
     }
 
     fn lock(&self) -> LockedGraphStorage<N> {
