### 2023-08-16 898d282
change save and load to always use a materialised graph (so the bincode includes the graph type)
