+++ raphtory/src/db/api/view/internal/materialize.rs
@@ -1,11 +1,15 @@
-use crate::db::{
-    api::view::internal::{Base, DynamicGraph, IntoDynamic},
-    graph::{
-        graph::{Graph, InternalGraph},
-        views::deletion_graph::GraphWithDeletions,
+use crate::{
+    core::utils::errors::GraphError,
+    db::{
+        api::view::internal::{Base, DynamicGraph, IntoDynamic},
+        graph::{
+            graph::{Graph, InternalGraph},
+            views::deletion_graph::GraphWithDeletions,
+        },
     },
 };
 use serde::{Deserialize, Serialize};
+use std::path::Path;
 
 #[derive(Serialize, Deserialize)]
 pub enum MaterializedGraph {
@@ -46,6 +50,18 @@ impl MaterializedGraph {
             MaterializedGraph::EventGraph(_) => None,
             MaterializedGraph::PersistentGraph(g) => Some(g),
         }
+    }
+
+    pub fn load_from_file<P: AsRef<Path>>(path: P) -> Result<Self, GraphError> {
+        let f = std::fs::File::open(path)?;
+        let mut reader = std::io::BufReader::new(f);
+        Ok(bincode::deserialize_from(&mut reader)?)
+    }
+
+    pub fn save_to_file<P: AsRef<Path>>(&self, path: P) -> Result<(), GraphError> {
+        let f = std::fs::File::create(path)?;
+        let mut writer = std::io::BufWriter::new(f);
+        Ok(bincode::serialize_into(&mut writer, self)?)
     }
 }
 
