+++ raphtory-graphql/src/model/schema/property_schema.rs
@@ -1,6 +1,5 @@
 use dynamic_graphql::SimpleObject;
-use raphtory::db::api::properties::Properties;
-use raphtory::db::api::view::internal::BoxableGraphView;
+use raphtory::db::api::{properties::Properties, view::internal::BoxableGraphView};
 use std::collections::HashSet;
 
 #[derive(SimpleObject)]
