+++ raphtory-graphql/src/model/mod.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, ops::Deref};
+use std::{collections::HashMap, io::BufReader, ops::Deref};
 
 use crate::{
     data::Data,
@@ -6,10 +6,14 @@ use crate::{
 };
 use async_graphql::Context;
 use dynamic_graphql::{
-    App, Mutation, MutationFields, MutationRoot, ResolvedObject, ResolvedObjectFields,
+    App, Mutation, MutationFields, MutationRoot, ResolvedObject, ResolvedObjectFields, Result,
+    Upload,
 };
 use itertools::Itertools;
-use raphtory::db::api::view::internal::IntoDynamic;
+use raphtory::{
+    db::api::view::internal::{DynamicGraph, IntoDynamic, MaterializedGraph},
+    search::IndexedGraph,
+};
 
 pub(crate) mod algorithm;
 pub(crate) mod filters;
@@ -69,6 +73,16 @@ impl Mut {
         let keys: Vec<_> = new_graphs.keys().cloned().collect();
         data.extend(new_graphs);
         keys
+    }
+
+    /// Use GQL multipart upload to send new graphs to server (Graphs should be of type MaterializedGraph)
+    async fn upload_graph<'a>(ctx: &Context<'a>, name: String, graph: Upload) -> Result<GqlGraph> {
+        let g: MaterializedGraph =
+            bincode::deserialize_from(BufReader::new(graph.value(ctx)?.content))?;
+        let gi: IndexedGraph<DynamicGraph> = g.into_dynamic().into();
+        let mut data = ctx.data_unchecked::<Data>().graphs.write();
+        data.insert(name, gi.clone());
+        Ok(GqlGraph::from(gi))
     }
 }
 
