+++ raphtory-graphql/Cargo.toml
@@ -14,28 +14,24 @@ homepage.workspace = true
 
 [dependencies]
 raphtory = { path = "../raphtory", version = "0.5.1" }
-
+bincode = "1"
 dotenv = "0.15.0"
 itertools = "0.10"
 serde = {version = "1.0.147", features = ["derive"]}
 once_cell = "1.17.2"
-
 poem = "1.3.48"
 tokio = {version = "1.18.2", features = ["full"]}
-
 async-graphql = {version = "5.0.5", features = ["dynamic-schema"]}
 dynamic-graphql = "0.7.3"
 async-graphql-poem = "5.0.5"
 parking_lot = { version = "0.12" , features = ["serde", "arc_lock", "send_guard"] }
 futures-util = "0.3.0"
 async-stream = "0.3.0"
-
 opentelemetry = {version = "0.18.0", features = ["rt-tokio"]}
 opentelemetry-jaeger = {version = "0.17.0", features = ["rt-tokio"]}
 tracing = "0.1.37"
 tracing-opentelemetry = "0.18.0"
 tracing-subscriber = {version = "0.3.16", features = ["std", "env-filter"]}
-
 walkdir = "2"
 ordered-float = "3.7.0"
 
