### 2023-08-16 6be52df
GraphQl send graph (#1181)
* start work on supporting sending graphs over graphql

* implement graph upload

* add test

* change save and load to always use a materialised graph (so the bincode includes the graph type)

* simplify the test
