### 2023-08-16 2a15466
add Rust Benchmark (cargo) benchmark result for 6be52dfb05debeb31670eab21d9116938954bdfa
