### 2023-08-16 81341f9
roundtrip support for sending and receiving graphs using base64-encoded strings
