### 2023-08-16 ee5637d
add graph schema to graphql (#1172)
* add graph schema to graphql

* add some docs

* run cargo fmt

---------

Co-authored-by: Ben Steer <b.a.steer@qmul.ac.uk>
