+++ raphtory-graphql/src/model/mod.rs
@@ -18,6 +18,7 @@ use raphtory::{
 pub(crate) mod algorithm;
 pub(crate) mod filters;
 pub(crate) mod graph;
+pub(crate) mod schema;
 
 #[derive(ResolvedObject)]
 #[graphql(root)]
@@ -29,7 +30,7 @@ impl QueryRoot {
         "Hello world from raphtory-graphql"
     }
 
-    /// Returns a view including all events between `t_start` (inclusive) and `t_end` (exclusive)
+    /// Returns a graph
     async fn graph<'a>(ctx: &Context<'a>, name: &str) -> Option<GqlGraph> {
         let data = ctx.data_unchecked::<Data>();
         let g = data.graphs.read().get(name).cloned()?;
