+++ raphtory-graphql/src/model/schema/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod graph_schema;
+pub(crate) mod node_schema;
+pub(crate) mod property_schema;
