### 2023-08-16 9dd0ce6
add graph schema to graphql (#1172)
* add graph schema to graphql

* add some docs

* run cargo fmt

---------

Co-authored-by: Ben Steer <b.a.steer@qmul.ac.uk>
