+++ raphtory-graphql/src/model/graph/node.rs
@@ -48,10 +48,12 @@ impl Node {
             .unwrap_or("NONE".to_string())
     }
 
-    async fn property_names<'a>(&self, _ctx: &Context<'a>) -> Vec<String> {
+    /// Returns all the property names this node has a value for
+    async fn property_names(&self) -> Vec<String> {
         self.vv.properties().keys().map_into().collect()
     }
 
+    /// Returns all the properties of the node
     async fn properties(&self) -> Option<Vec<Property>> {
         Some(
             self.vv
@@ -62,10 +64,12 @@ impl Node {
         )
     }
 
+    /// Returns the value for the property with name `name`
     async fn property(&self, name: &str) -> Option<String> {
         self.vv.properties().get(name).map(|v| v.to_string())
     }
 
+    /// Returns the history as a vector of updates for the property with name `name`
     async fn property_history(&self, name: String) -> Vec<PropertyUpdate> {
         self.vv
             .properties()
@@ -120,6 +124,7 @@ impl Node {
         }
     }
 
+    /// Returns the number of edges connected to the node
     async fn degree(&self, layers: Option<Vec<String>>) -> usize {
         match layers {
             None => self.vv.degree(),
@@ -136,6 +141,7 @@ impl Node {
         }
     }
 
+    /// Returns the number of outgoing edges connected to the node
     async fn out_degree(&self, layer: Option<String>) -> usize {
         match layer.as_deref() {
             None => self.vv.out_degree(),
@@ -146,6 +152,7 @@ impl Node {
         }
     }
 
+    /// Returns the number of incoming edges connected to the node
     async fn in_degree(&self, layer: Option<String>) -> usize {
         match layer.as_deref() {
             None => self.vv.in_degree(),
