### 2023-08-16 72646fd
Added grouping for colour onto the nodes in pyvis (#1183)
