### 2023-08-16 e38933f
Release v0.5.2 (#1184)
chore: Release

Co-authored-by: Haaroon <Haaroon@users.noreply.github.com>
