### 2023-08-16 234e767
add variant with sorted vector for TimeIndex
