### 2023-08-16 35ec18d
add Rust Benchmark (cargo) benchmark result for e38933ffcefadb461f26d4fe85b5117655e66795
