### 2023-08-16 c3dc68b
improve has_edge
