+++ raphtory/src/db/internal/core_ops.rs
@@ -52,10 +52,14 @@ impl<const N: usize> CoreGraphOps for InnerTemporalGraph<N> {
         vertex.additions().unwrap()
     }
 
+    fn localise_vertex(&self, v: VertexRef) -> Option<VID> {
+        self.inner().resolve_vertex_ref(v)
+    }
+
     fn localise_vertex_unchecked(&self, v: VertexRef) -> VID {
         match v {
             VertexRef::Local(l) => l,
-            VertexRef::Remote(_) => self.inner().resolve_vertex_ref(&v).unwrap(),
+            VertexRef::Remote(_) => self.inner().resolve_vertex_ref(v).unwrap(),
         }
     }
 
