### 2023-08-16 5485dd4
moved graphql tests to own folder, renamed graphlql to graphlqlserver, added a new graphqlclient with predefined mutation queries, added tests except for upload, added new [gql dep to toml
