+++ python/pyproject.toml
@@ -15,7 +15,8 @@ dependencies = [
     "networkx >= 2.6.3",
     "pandas >= 2.0.3",
     "pyarrow >= 12.0.1",
-    "requests >= 2.31.0"
+    "requests >= 2.31.0",
+    "gql[all] == 3.4.1"
 ]
 
 
