+++ docs/requirements.txt
@@ -77,7 +77,7 @@ sphinxcontrib-serializinghtml==1.1.5
 stack-data==0.6.2
 tinycss2==1.2.1
 tomli==2.0.1
-tornado==6.3.2
+tornado==6.3.3
 traitlets==5.9.0
 typing_extensions==4.7.1
 tzdata==2023.3
