### 2023-08-16 8f0c3e8
add Rust Benchmark (cargo) benchmark result for d6fb05ed9bfe80a885ebfe0ca120e969ced068b0
