### 2023-08-17 0fa937a
rename receive function
