### 2023-08-17 ed93c36
roundtrip support for sending and receiving graphs using base64-encoding (#1182)
* roundtrip support for sending and receiving graphs using base64-encoded strings

* print response for failed test

* fix the test

* rename receive function
