+++ raphtory-graphql/src/model/mod.rs
@@ -1,10 +1,17 @@
-use std::{collections::HashMap, io::BufReader, ops::Deref};
+use std::{
+    collections::HashMap,
+    error::Error,
+    fmt::{Display, Formatter},
+    io::BufReader,
+    ops::Deref,
+};
 
 use crate::{
     data::Data,
     model::graph::graph::{GqlGraph, GraphMeta},
 };
 use async_graphql::Context;
+use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
 use dynamic_graphql::{
     App, Mutation, MutationFields, MutationRoot, ResolvedObject, ResolvedObjectFields, Result,
     Upload,
@@ -12,6 +19,7 @@ use dynamic_graphql::{
 use itertools::Itertools;
 use raphtory::{
     db::api::view::internal::{DynamicGraph, IntoDynamic, MaterializedGraph},
+    prelude::GraphViewOps,
     search::IndexedGraph,
 };
 
@@ -19,6 +27,17 @@ pub(crate) mod algorithm;
 pub(crate) mod filters;
 pub(crate) mod graph;
 pub(crate) mod schema;
+
+#[derive(Debug)]
+pub struct MissingGraph;
+
+impl Display for MissingGraph {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Graph does not exist")
+    }
+}
+
+impl Error for MissingGraph {}
 
 #[derive(ResolvedObject)]
 #[graphql(root)]
@@ -45,6 +64,19 @@ impl QueryRoot {
             .map(|(name, g)| GraphMeta::new(name.clone(), g.deref().clone().into_dynamic()))
             .collect_vec()
     }
+
+    async fn receive_graph<'a>(ctx: &Context<'a>, name: &str) -> Result<String> {
+        let data = ctx.data_unchecked::<Data>();
+        let g = data
+            .graphs
+            .read()
+            .get(name)
+            .cloned()
+            .ok_or(MissingGraph)?
+            .materialize()?;
+        let bincode = bincode::serialize(&g)?;
+        Ok(URL_SAFE_NO_PAD.encode(bincode))
+    }
 }
 
 #[derive(MutationRoot)]
@@ -56,6 +88,9 @@ pub(crate) struct Mut(MutRoot);
 #[MutationFields]
 impl Mut {
     /// Load graphs from a directory of bincode files (existing graphs with the same name are overwritten)
+    ///
+    /// # Returns:
+    ///   list of names for newly added graphs
     async fn load_graphs_from_path<'a>(ctx: &Context<'a>, path: String) -> Vec<String> {
         let new_graphs = Data::load_from_file(&path);
         let keys: Vec<_> = new_graphs.keys().cloned().collect();
@@ -65,6 +100,9 @@ impl Mut {
     }
 
     /// Load new graphs from a directory of bincode files (existing graphs will not been overwritten)
+    ///
+    /// # Returns:
+    ///   list of names for newly added graphs
     async fn load_new_graphs_from_path<'a>(ctx: &Context<'a>, path: String) -> Vec<String> {
         let mut data = ctx.data_unchecked::<Data>().graphs.write();
         let new_graphs: HashMap<_, _> = Data::load_from_file(&path)
@@ -76,14 +114,28 @@ impl Mut {
         keys
     }
 
-    /// Use GQL multipart upload to send new graphs to server (Graphs should be of type MaterializedGraph)
-    async fn upload_graph<'a>(ctx: &Context<'a>, name: String, graph: Upload) -> Result<GqlGraph> {
+    /// Use GQL multipart upload to send new graphs to server
+    ///
+    /// # Returns:
+    ///    name of the new graph
[…]
