+++ raphtory-graphql/Cargo.toml
@@ -15,6 +15,8 @@ homepage.workspace = true
 [dependencies]
 raphtory = { path = "../raphtory", version = "0.5.2" }
 bincode = "1"
+base64 = "0.21.2"
+thiserror = "1.0.44"
 dotenv = "0.15.0"
 itertools = "0.10"
 serde = {version = "1.0.147", features = ["derive"]}
