### 2023-08-17 aa79cd8
add Rust Benchmark (cargo) benchmark result for ed93c36258dcdee8ef33d1f3062f206e638227f8
