### 2023-08-17 e8cc2cd
Added GraphQL Client  (#1185)
* moved graphql tests to own folder, renamed graphlql to graphlqlserver, added a new graphqlclient with predefined mutation queries, added tests except for upload, added new [gql dep to toml

* sorted the test result

* add support for upload and add tests, format

* better docs
