### 2023-08-17 1b38045
add Rust Benchmark (cargo) benchmark result for 64b95dec6781281154fe7a996f459e47871b435b
