### 2023-08-18 550e37f
add Rust Benchmark (cargo) benchmark result for a462cb039b14dc26d7eadd206a29c9c637e6253a
