### 2023-08-18 68ee485
Implement iterators that avoid genawaiter
