+++ raphtory/src/core/storage/mod.rs
@@ -7,11 +7,13 @@ pub mod sorted_vec_map;
 pub mod timeindex;
 
 use self::iter::Iter;
+use lock_api;
 use locked_view::LockedView;
 use parking_lot::{RwLock, RwLockReadGuard};
 use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
 use std::{
+    array,
     fmt::Debug,
     ops::{Deref, DerefMut},
     sync::{
@@ -19,6 +21,9 @@ use std::{
         Arc,
     },
 };
+use tantivy::directory::Lock;
+
+type ArcRwLockReadGuard<T> = lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, T>;
 
 #[inline]
 fn resolve<const N: usize>(index: usize) -> (usize, usize) {
@@ -47,7 +52,7 @@ impl<T: Default> LockVec<T> {
         }
     }
 
-    pub fn read_arc_lock(&self) -> lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<T>> {
+    pub fn read_arc_lock(&self) -> ArcRwLockReadGuard<Vec<T>> {
         RwLock::read_arc(&self.data)
     }
 }
@@ -66,7 +71,7 @@ impl<T: PartialEq + Default, const N: usize> PartialEq for RawStorage<T, N> {
 
 #[derive(Debug)]
 pub struct ReadLockedStorage<T, const N: usize> {
-    locks: Box<[lock_api::ArcRwLockReadGuard<parking_lot::RawRwLock, Vec<T>>; N]>,
+    locks: [ArcRwLockReadGuard<Vec<T>>; N],
     len: usize,
 }
 
@@ -80,20 +85,60 @@ impl<T, const N: usize> ReadLockedStorage<T, N> {
     pub(crate) fn iter(&self) -> impl Iterator<Item = &T> + '_ {
         self.locks.iter().flat_map(|v| v.iter())
     }
+
+    pub(crate) fn par_iter(&self) -> impl ParallelIterator<Item = &T> + '_
+    where
+        T: Send + Sync,
+    {
+        self.locks.par_iter().flat_map(|v| v.par_iter())
+    }
+
+    pub(crate) fn into_iter(self) -> impl Iterator<Item = LockedEntry<T, N>> + Send
+    where
+        T: Send + Sync + 'static,
+    {
+        self.locks
+            .into_iter()
+            .enumerate()
+            .flat_map(|(bucket, data)| {
+                let arc_data = Arc::new(data);
+                (0..arc_data.len()).map(move |offset| LockedEntry {
+                    guard: arc_data.clone(),
+                    offset,
+                    bucket,
+                })
+            })
+    }
+
+    pub(crate) fn into_par_iter(self) -> impl ParallelIterator<Item = LockedEntry<T, N>>
+    where
+        T: Send + Sync + 'static,
+    {
+        self.locks
+            .into_par_iter()
+            .enumerate()
+            .flat_map(|(bucket, data)| {
+                let arc_data = Arc::new(data);
+                (0..arc_data.len())
+                    .into_par_iter()
+                    .map(move |offset| LockedEntry {
+                        guard: arc_data.clone(),
+                        offset,
+                        bucket,
+                    })
+            })
+    }
 }
 
 impl<T: Default + Send + Sync, const N: usize> RawStorage<T, N> {
     pub fn count_with_filter<F: Fn(&T) -> bool + Send + Sync>(&self, f: F) -> usize {
-        self.data
-            .par_iter()
-            .map(|lock_vec| lock_vec.data.read().par_iter().filter(|x| f(x)).count())
[…]
