### 2023-08-18 97579e9
add cleaner test to show broken num_edges on window for GraphWithDeletions
