+++ raphtory/src/db/graph/views/deletion_graph.rs
@@ -414,6 +414,8 @@ mod test_deletions {
             vec![(0, 1)]
         );
 
+        assert_eq!(g.window(1, 2).num_edges(), 1);
+
         assert!(g.window(11, 12).is_empty());
 
         assert_eq!(
