### 2023-08-21 6e7bb21
add Rust Benchmark (cargo) benchmark result for c5930ada9ec4613928663ac00f69e9f4b87a01f2
