### 2023-08-21 2cd6f87
add schema for layers and edges
