### 2023-08-21 883aeb1
add docs for LayerSchema.name function
