### 2023-08-21 c340a39
Tidying lotr + structure
