+++ docs/source/getting_started/installation/index.rst
@@ -1,7 +1,7 @@
 .. _advanced-installation:
 
 =========================
-Advanced Installation
+Installing from Source
 =========================
 
 .. toctree::
