+++ docs/source/getting_started/faqs/index.rst
@@ -1,0 +1,13 @@
+{{ header }}
+
+.. _FQA_index:
+
+
+===============
+FAQ - Quick look at Raphtory functionality
+===============
+
+
+.. toctree::
+    :maxdepth: 1
+    faq_ingestion.rst
