+++ docs/source/getting_started/installation/index.rst
@@ -5,7 +5,7 @@ Installing from Source
 =========================
 
 .. toctree::
-    :maxdepth: 1
+    :maxdepth: 2
 
     install_python
     install_rust
