+++ docs/source/getting_started/faqs/faq_ingestion.rst
@@ -2,9 +2,9 @@
 
 {{ header }}
 
-=========
-FAQ
-=========
+=============================================
+Ingestion - Getting data into Raphtory
+=============================================
 
 
 
