### 2023-08-22 9f27d8c
Added to all pandas functions
