+++ raphtory/src/python/graph/pandas.rs
@@ -14,7 +14,7 @@ fn i64_opt_into_u64_opt(x: Option<&i64>) -> Option<u64> {
     x.map(|x| (*x).try_into().unwrap())
 }
 
-pub(crate) fn process_pandas_py_df(df: &PyAny, py: Python) -> PyResult<PretendDF> {
+pub(crate) fn process_pandas_py_df(df: &PyAny, py: Python,size:usize) -> PyResult<PretendDF> {
     let globals = PyDict::new(py);
     globals.set_item("df", df)?;
     let module = py.import("pyarrow")?;
@@ -49,6 +49,7 @@ pub(crate) fn process_pandas_py_df(df: &PyAny, py: Python) -> PyResult<PretendDF
 
 pub(crate) fn load_vertices_from_df<'a>(
     df: &'a PretendDF,
+    size: usize,
     vertex_id: &str,
     time: &str,
     props: Option<Vec<&str>>,
@@ -73,17 +74,18 @@ pub(crate) fn load_vertices_from_df<'a>(
     if let (Some(vertex_id), Some(time)) = (df.iter_col::<u64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(|i| i.copied()).zip(time);
-        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter, shared_const_props)?;
+        load_vertices_from_num_iter(graph, size,iter, prop_iter, const_prop_iter, shared_const_props)?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.iter_col::<i64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(i64_opt_into_u64_opt).zip(time);
-        load_vertices_from_num_iter(graph, iter, prop_iter, const_prop_iter, shared_const_props)?;
+        load_vertices_from_num_iter(graph, size,iter, prop_iter, const_prop_iter, shared_const_props)?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.utf8::<i32>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.into_iter().zip(time);
-        for (((vertex_id, time), props), const_props) in iter.zip(prop_iter).zip(const_prop_iter) {
+        for (((vertex_id, time), props), const_props)
+            in tqdm!(iter.zip(prop_iter).zip(const_prop_iter),desc="Loading vertices",total=size,animation= kdam::Animation::FillUp,unit_scale=true) {
             if let (Some(vertex_id), Some(time)) = (vertex_id, time) {
                 graph.add_vertex(*time, vertex_id, props)?;
                 graph.add_vertex_properties(vertex_id, const_props)?;
@@ -96,7 +98,8 @@ pub(crate) fn load_vertices_from_df<'a>(
         (df.utf8::<i64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.into_iter().zip(time);
-        for (((vertex_id, time), props), const_props) in iter.zip(prop_iter).zip(const_prop_iter) {
+        for (((vertex_id, time), props), const_props)
+            in tqdm!(iter.zip(prop_iter).zip(const_prop_iter),desc="Loading vertices",total=size,animation= kdam::Animation::FillUp,unit_scale=true) {
             if let (Some(vertex_id), Some(time)) = (vertex_id, time) {
                 graph.add_vertex(*time, vertex_id, props)?;
                 graph.add_vertex_properties(vertex_id, const_props)?;
@@ -116,6 +119,7 @@ pub(crate) fn load_vertices_from_df<'a>(
 
[…]
