+++ raphtory/src/python/graph/pandas.rs
@@ -1,20 +1,20 @@
+use crate::{core::utils::errors::GraphError, prelude::*};
 use arrow2::{
     array::{Array, BooleanArray, PrimitiveArray, Utf8Array},
     ffi,
     types::{NativeType, Offset},
 };
+use kdam::tqdm;
 use pyo3::{
     create_exception, exceptions::PyException, ffi::Py_uintptr_t, prelude::*, types::PyDict,
 };
 use std::collections::HashMap;
-use crate::{core::utils::errors::GraphError, prelude::*};
-use kdam::tqdm;
 
 fn i64_opt_into_u64_opt(x: Option<&i64>) -> Option<u64> {
     x.map(|x| (*x).try_into().unwrap())
 }
 
-pub(crate) fn process_pandas_py_df(df: &PyAny, py: Python,size:usize) -> PyResult<PretendDF> {
+pub(crate) fn process_pandas_py_df(df: &PyAny, py: Python, size: usize) -> PyResult<PretendDF> {
     let globals = PyDict::new(py);
     globals.set_item("df", df)?;
     let module = py.import("pyarrow")?;
@@ -30,9 +30,8 @@ pub(crate) fn process_pandas_py_df(df: &PyAny, py: Python,size:usize) -> PyResul
         vec![]
     };
 
-    let arrays = tqdm!(rb
-        .iter()
-        .map(|rb| {
+    let arrays = tqdm!(
+        rb.iter().map(|rb| {
             (0..names.len())
                 .map(|i| {
                     let array = rb.call_method1("column", (i,))?;
@@ -40,8 +39,13 @@ pub(crate) fn process_pandas_py_df(df: &PyAny, py: Python,size:usize) -> PyResul
                     Ok::<Box<dyn Array>, PyErr>(arr)
                 })
                 .collect::<Result<Vec<_>, PyErr>>()
-        }),desc="Converting dataframe to Arrow",total=size,animation= kdam::Animation::FillUp,unit_scale=true)
-        .collect::<Result<Vec<_>, PyErr>>()?;
+        }),
+        desc = "Converting dataframe to Arrow",
+        total = size,
+        animation = kdam::Animation::FillUp,
+        unit_scale = true
+    )
+    .collect::<Result<Vec<_>, PyErr>>()?;
 
     let df = PretendDF { names, arrays };
     Ok(df)
@@ -74,18 +78,37 @@ pub(crate) fn load_vertices_from_df<'a>(
     if let (Some(vertex_id), Some(time)) = (df.iter_col::<u64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(|i| i.copied()).zip(time);
-        load_vertices_from_num_iter(graph, size,iter, prop_iter, const_prop_iter, shared_const_props)?;
+        load_vertices_from_num_iter(
+            graph,
+            size,
+            iter,
+            prop_iter,
+            const_prop_iter,
+            shared_const_props,
+        )?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.iter_col::<i64>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.map(i64_opt_into_u64_opt).zip(time);
-        load_vertices_from_num_iter(graph, size,iter, prop_iter, const_prop_iter, shared_const_props)?;
+        load_vertices_from_num_iter(
+            graph,
+            size,
+            iter,
+            prop_iter,
+            const_prop_iter,
+            shared_const_props,
+        )?;
     } else if let (Some(vertex_id), Some(time)) =
         (df.utf8::<i32>(vertex_id), df.iter_col::<i64>(time))
     {
         let iter = vertex_id.into_iter().zip(time);
[…]
