### 2023-08-23 d33f7d5
make input argument mut in merge_schemas
