+++ raphtory/src/algorithms/k_core.rs
@@ -11,9 +11,7 @@ use crate::{
         },
     },
 };
-use std::{
-    collections::{HashSet},
-};
+use std::collections::HashSet;
 
 #[derive(Clone, Debug)]
 struct KCoreState {
@@ -112,10 +110,7 @@ where
 mod k_core_test {
     use std::collections::HashSet;
 
-    use crate::{
-        algorithms::{k_core::k_core_set},
-        prelude::*,
-    };
+    use crate::{algorithms::k_core::k_core_set, prelude::*};
 
     #[test]
     fn k_core_2() {
