### 2023-08-23 7b58b58
roll back changes in src/model/mod.rs
