### 2023-08-23 b22967d
add schema for layers and edges (#1196)
* add schema for layers and edges

* add docs for LayerSchema.name function

* run cargo fmt

* make input argument mut in merge_schemas

* roll back changes in src/model/mod.rs
