### 2023-08-23 cb93f74
add Rust Benchmark (cargo) benchmark result for c458c2bac40585c5d3ddebae4a5d21a047f9b8c4
