+++ raphtory/Cargo.toml
@@ -43,6 +43,7 @@ lock_api = { version = "0.4", features = ["arc_lock", "serde"] }
 dashmap = {version ="5", features = ["serde"] }
 serde_with = "1.12.0"
 enum_dispatch = "0.3"
+kdam = "0.4.1"
 
 # io optional dependencies
 csv = {version="1.1.6", optional=true}
