### 2023-08-23 0189db2
add Rust Benchmark (cargo) benchmark result for 58d8a08b207eed7b0ae7ab7ba2e2f254cb50af5a
