+++ dev/bench/data.js
@@ -1,5 +1,5 @@
 window.BENCHMARK_DATA = {
-  "lastUpdate": 1692785165518,
+  "lastUpdate": 1692795768956,
   "repoUrl": "https://github.com/Pometry/Raphtory",
   "entries": {
     "Rust Benchmark": [
@@ -8815,6 +8815,204 @@ window.BENCHMARK_DATA = {
             "name": "lotr_graph_window_10/max_degree",
             "value": 357166,
             "range": "± 3061",
+            "unit": "ns/iter"
+          }
+        ]
+      },
+      {
+        "commit": {
+          "author": {
+            "email": "b.a.steer@qmul.ac.uk",
+            "name": "Ben Steer",
+            "username": "miratepuffin"
+          },
+          "committer": {
+            "email": "noreply@github.com",
+            "name": "GitHub",
+            "username": "web-flow"
+          },
+          "distinct": true,
+          "id": "58d8a08b207eed7b0ae7ab7ba2e2f254cb50af5a",
+          "message": "Pandas loading bars (#1197)\n\n* tqdm working\r\n\r\n* cargo lock\r\n\r\n* Added to all pandas functions\r\n\r\n* fmt\r\n\r\n* Fixed test and warnings\r\n\r\n* fmt\r\n\r\n* Fixed warnings in grpahql test\r\n\r\n* trying a different len",
+          "timestamp": "2023-08-23T13:52:42+01:00",
+          "tree_id": "a01279c377c9bfd50f17ea828b57474552fa14dc",
+          "url": "https://github.com/Pometry/Raphtory/commit/58d8a08b207eed7b0ae7ab7ba2e2f254cb50af5a"
+        },
+        "date": 1692795768205,
+        "tool": "cargo",
+        "benches": [
+          {
+            "name": "large/1k fixed edge updates with varying time",
+            "value": 669719,
+            "range": "± 71229",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "large/1k fixed edge updates with varying time and numeric string input",
+            "value": 1054058,
+            "range": "± 3753",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "large/1k fixed edge updates with varying time and string input",
+            "value": 1100339,
+            "range": "± 2034",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "large/1k random edge additions",
+            "value": 1124861,
+            "range": "± 80737",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "large/1k random edge additions with numeric string input",
+            "value": 1723292,
+            "range": "± 81997",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "lotr_graph/num_edges",
+            "value": 8,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "lotr_graph/has_edge_existing",
+            "value": 95,
+            "range": "± 1",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "lotr_graph/has_edge_nonexisting",
+            "value": 84,
+            "range": "± 1",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "lotr_graph/num_vertices",
+            "value": 1,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "lotr_graph/has_vertex_existing",
+            "value": 23,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "lotr_graph/has_vertex_nonexisting",
+            "value": 21,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "lotr_graph/max_id",
+            "value": 5061,
+            "range": "± 24",
+            "unit": "ns/iter"
+          },
+          {
[…]
