+++ raphtory/src/algorithms/algorithm_result.rs
@@ -1,3 +1,7 @@
+use crate::{
+    core::entities::vertices::vertex_ref::VertexRef, db::graph::vertex::VertexView, prelude::*,
+};
+use itertools::Itertools;
 use ordered_float::OrderedFloat;
 use std::{collections::HashMap, fmt, fmt::Debug, hash::Hash};
 
@@ -9,63 +13,74 @@ use std::{collections::HashMap, fmt, fmt::Debug, hash::Hash};
 ///
 /// This `AlgorithmResult` is returned for all algorithms that return a HashMap
 ///
-pub struct AlgorithmResult<K, V>
+pub struct AlgorithmResult<V, G>
 where
-    K: Clone + Hash + Eq + Ord,
     V: Clone,
+    G: GraphViewOps,
 {
     /// The result hashmap that stores keys of type `H` and values of type `Y`.
-    pub result: HashMap<K, V>,
+    pub result: Vec<V>,
+    pub graph: G,
 }
 
-impl<K, V> AlgorithmResult<K, V>
+impl<V, G> AlgorithmResult<V, G>
 where
-    K: Clone + Hash + Eq + Ord,
     V: Clone,
+    G: GraphViewOps,
 {
     /// Creates a new instance of `AlgorithmResult` with the provided hashmap.
     ///
     /// # Arguments
     ///
     /// * `result`: A `HashMap` with keys of type `H` and values of type `Y`.
-    pub fn new(result: HashMap<K, V>) -> Self {
-        Self { result }
+    pub fn new(result: Vec<V>, graph: G) -> Self {
+        Self { result, graph }
     }
 
     /// Returns a reference to the entire `result` hashmap.
-    pub fn get_all(&self) -> &HashMap<K, V> {
+    pub fn get_all_values(&self) -> &Vec<V> {
         &self.result
+    }
+
+    pub fn get_all(&self) -> Vec<(VertexView<G>, V)> {
+        self.graph
+            .vertices()
+            .into_iter()
+            .map(|vertex| (vertex.clone(), self.get(vertex).unwrap().clone()))
+            .collect_vec()
     }
 
     /// Returns the value corresponding to the provided key in the `result` hashmap.
     ///
     /// # Arguments
     ///
-    /// * `key`: The key of type `H` for which the value is to be retrieved.
-    pub fn get(&self, key: &K) -> Option<&V> {
-        self.result.get(&key)
+    /// * `key`: The key of type `K` for which the value is to be retrieved.
+    pub fn get<K: Into<VertexRef>>(&self, vertex: K) -> Option<&V> {
+        let vert = self.graph.local_vertex_ref(vertex.into())?;
+        let bb: usize = vert.into();
+        self.result.get(bb)
     }
 }
 
-pub struct AlgorithmResultIterator<'a, K, V> {
-    iter: std::collections::hash_map::Iter<'a, K, V>,
+pub struct AlgorithmResultIterator<'a, V> {
+    iter: std::slice::Iter<'a, V>,
 }
 
-impl<'a, K, V> Iterator for AlgorithmResultIterator<'a, K, V> {
-    type Item = (&'a K, &'a V);
+impl<'a, V> Iterator for AlgorithmResultIterator<'a, V> {
+    type Item = (&'a V);
 
     fn next(&mut self) -> Option<Self::Item> {
         self.iter.next()
     }
 }
 
-impl<'a, K, V> IntoIterator for &'a AlgorithmResult<K, V>
+impl<'a, V, G> IntoIterator for &'a AlgorithmResult<V, G>
 where
-    K: Clone + Hash + Ord,
     V: Clone,
+    G: GraphViewOps,
 {
[…]
