+++ raphtory/src/algorithms/temporal_reachability.rs
@@ -69,7 +69,7 @@ pub fn temporally_reachable_nodes<G: GraphViewOps, T: InputVertex>(
     start_time: i64,
     seed_nodes: Vec<T>,
     stop_nodes: Option<Vec<T>>,
-) -> AlgorithmResult<String, Vec<(i64, String)>> {
+) -> AlgorithmResult<Vec<(i64, String)>, G> {
     let mut ctx: Context<G, ComputeStateVec> = g.into();
 
     let infected_nodes = seed_nodes.into_iter().map(|n| n.id()).collect_vec();

[…]
