+++ docs/source/getting_started/faqs/faq_ingestion.rst
@@ -3,7 +3,7 @@
 {{ header }}
 
 =============================================
-Ingestion - Getting data into Raphtory
+FAQ - Getting data into Raphtory
 =============================================
 
 
