### 2023-08-23 0a554e3
Small changes to docs before the big update (#1199)
* Tided the lotr examples

* Tidying lotr + structure

* Removed overview till we write a nicer one

* Updating the faq

* Finalised this version of docs

* Added a couple bits to the reddit example
