+++ docs/source/api/index.rst
@@ -9,7 +9,6 @@ API
 .. toctree::
    :maxdepth: 2
 
-
    raphtory
    rust
 
