+++ docs/source/getting_started/intro_tutorials/index.rst
@@ -10,5 +10,3 @@ Tutorials
     :maxdepth: 1
 
     lotr.ipynb
-    graphql
-
