### 2023-08-23 cc24093
add Rust Benchmark (cargo) benchmark result for 9ca9166b455304d616b3ee817f5f416fdbca6b93
