### 2023-08-23 5b08bad
sum weight for a vertex algorithm
