+++ raphtory/src/algorithms/mod.rs
@@ -41,3 +41,4 @@ pub mod reciprocity;
 pub mod temporal_reachability;
 pub mod triangle_count;
 pub mod triplet_count;
+pub mod weight_accum;
