### 2023-08-23 a9c5d6a
add in out and both
