+++ raphtory/src/algorithms/weight_accum.rs
@@ -4,9 +4,12 @@
 //! in a graph.
 use crate::{
     algorithms::algorithm_result::AlgorithmResult,
-    core::state::{
-        accumulator_id::accumulators::sum,
-        compute_state::{ComputeState, ComputeStateVec},
+    core::{
+        state::{
+            accumulator_id::accumulators::sum,
+            compute_state::{ComputeState, ComputeStateVec},
+        },
+        Direction,
     },
     db::{
         api::view::GraphViewOps,
@@ -21,27 +24,42 @@ use crate::{
 };
 use ordered_float::OrderedFloat;
 
-/// Computes the net sum of weights for a given vertex.
+/// Computes the net sum of weights for a given vertex based on edge direction.
 ///
-/// For every edge connected to the vertex, if the source of the edge is the vertex itself,
-/// the weight is treated as negative. Otherwise, it's treated as positive.
+/// For every edge connected to the vertex, this function checks the source of the edge
+/// against the vertex itself to determine the directionality. The weight can be treated
+/// as negative or positive based on the edge's source and the specified direction:
+///
+/// - If the edge's source is the vertex itself and the direction is either `OUT` or `BOTH`,
+///   the weight is treated as negative.
+/// - If the edge's source is not the vertex and the direction is either `IN` or `BOTH`,
+///   the weight is treated as positive.
+/// - In all other cases, the weight contribution is zero.
 ///
 /// # Parameters
 /// - `v`: The vertex for which we want to compute the weight sum.
 /// - `name`: The name of the property which holds the edge weight.
+/// - `direction`: Specifies the direction of edges to consider (`IN`, `OUT`, or `BOTH`).
 ///
 /// # Returns
-/// Returns a `f64` which is the net sum of weights for the vertex.
+/// Returns a `f64` which is the net sum of weights for the vertex considering the specified direction.
 fn sum_weight_for_vertex<G: GraphViewOps, CS: ComputeState>(
     v: &EvalVertexView<G, CS, ()>,
     name: String,
+    direction: Direction,
 ) -> f64 {
     v.edges()
         .map(|edge| {
-            if edge.src().name() == v.name() {
+            if edge.src().name() == v.name()
+                && ((direction == Direction::OUT) | (direction == Direction::BOTH))
+            {
                 -edge.properties().get(name.clone()).unwrap_f64()
-            } else {
+            } else if edge.src().name() != v.name()
+                && ((direction == Direction::IN) | (direction == Direction::BOTH))
+            {
                 edge.properties().get(name.clone()).unwrap_f64()
+            } else {
+                0.0
             }
         })
         .sum()
@@ -50,6 +68,7 @@ fn sum_weight_for_vertex<G: GraphViewOps, CS: ComputeState>(
 /// Computes the sum of weights for all vertices in the graph.
 ///
 /// This function iterates over all vertices and calculates the net sum of weights.
+/// Incoming edges have a positive sum and outgoing edges have a negative sum
 /// It uses a compute context and tasks to achieve this.
 ///
 /// # Parameters
@@ -60,16 +79,17 @@ fn sum_weight_for_vertex<G: GraphViewOps, CS: ComputeState>(
 ///
 /// # Returns
 /// Returns an `AlgorithmResult` which maps each vertex to its corresponding net weight sum.
-pub fn sum_weights<G: GraphViewOps>(
+pub fn sum_weights_edges<G: GraphViewOps>(
     graph: &G,
     name: String,
+    direction: Direction,
     threads: Option<usize>,
[…]
