### 2023-08-23 93b74e7
yugioh mode, implement pydirection, implement python algorithm + tests
