### 2023-08-24 4d2c82d
implemented many functions upon the list, listlist, pyiterable, sum, count,
