+++ python/tests/test_graphdb.py
@@ -271,6 +271,7 @@ def test_getitem():
 
     assert g.vertex(1).properties.temporal.get("cost") == g.vertex(1).properties.temporal["cost"]
 
+
 def test_graph_properties():
     g = create_graph()
 
@@ -1000,7 +1001,7 @@ def test_edge_history():
     view = g.window(1, 5)
 
     assert (g.edge(1, 2).history() == [1, 3])
-    assert(view.edge(1, 4).history() == [4])
+    assert (view.edge(1, 4).history() == [4])
 
 
 def test_lotr_edge_history():
@@ -1412,7 +1413,6 @@ def test_deletions():
     assert list(g.edge(edges[0][1], edges[0][2]).explode().latest_time()) == [10]
 
 
-
 def test_load_from_pandas():
     import pandas as pd
     df = pd.DataFrame({
@@ -1519,60 +1519,87 @@ def test_load_from_pandas_with_types():
         "weight": [1.0, 2.0, 3.0, 4.0, 5.0],
         "marbles": ["red", "blue", "green", "yellow", "purple"],
         "marbles_const": ["red", "blue", "green", "yellow", "purple"],
-        "layers":  ["layer 1", "layer 2", "layer 3", "layer 4", "layer 5"]
+        "layers": ["layer 1", "layer 2", "layer 3", "layer 4", "layer 5"]
     })
     vertices_df = pd.DataFrame({
         "id": [1, 2, 3, 4, 5, 6],
         "name": ["Alice", "Bob", "Carol", "Dave", "Eve", "Frank"],
         "time": [1, 2, 3, 4, 5, 6],
-        "type":  ["Person 1", "Person 2", "Person 3", "Person 4", "Person 5", "Person 6"]
+        "type": ["Person 1", "Person 2", "Person 3", "Person 4", "Person 5", "Person 6"]
     })
     g = Graph()
-    g.load_vertices_from_pandas(vertices_df, "id", "time", ["name"],shared_const_props={"type": "Person", "tag": "test_tag"})
-    assert g.vertices().properties.constant.get("type").collect() == ["Person", "Person", "Person", "Person", "Person", "Person"]
-    assert g.vertices().properties.constant.get("tag").collect() == ["test_tag", "test_tag", "test_tag", "test_tag", "test_tag", "test_tag"]
+    g.load_vertices_from_pandas(vertices_df, "id", "time", ["name"],
+                                shared_const_props={"type": "Person", "tag": "test_tag"})
+    assert g.vertices().properties.constant.get("type").collect() == ["Person", "Person", "Person", "Person", "Person",
+                                                                      "Person"]
+    assert g.vertices().properties.constant.get("tag").collect() == ["test_tag", "test_tag", "test_tag", "test_tag",
+                                                                     "test_tag", "test_tag"]
 
     g = Graph()
-    g.load_vertices_from_pandas(vertices_df, "id", "time", ["name"],const_props=["type"])
-    assert g.vertices().properties.constant.get("type").collect() == ["Person 1", "Person 2", "Person 3", "Person 4", "Person 5", "Person 6"]
+    g.load_vertices_from_pandas(vertices_df, "id", "time", ["name"], const_props=["type"])
+    assert g.vertices().properties.constant.get("type").collect() == ["Person 1", "Person 2", "Person 3", "Person 4",
+                                                                      "Person 5", "Person 6"]
 
     g = Graph()
-    g.load_edges_from_pandas(edges_df, "src", "dst", "time", ["weight", "marbles"], const_props=["marbles_const"], shared_const_props={"type": "Edge", "tag": "test_tag"}, layer="test_layer")
[…]
