+++ raphtory/src/core/mod.rs
@@ -66,6 +66,20 @@ impl Prop {
     pub fn str(s: &str) -> Prop {
         Prop::Str(s.to_string())
     }
+
+    pub fn add(self, other: Prop) -> Option<Prop> {
+        match (self, other) {
+            (Prop::I32(a), Prop::I32(b)) => Some(Prop::I32(a + b)),
+            (Prop::I64(a), Prop::I64(b)) => Some(Prop::I64(a + b)),
+            (Prop::U32(a), Prop::U32(b)) => Some(Prop::U32(a + b)),
+            (Prop::U64(a), Prop::U64(b)) => Some(Prop::U64(a + b)),
+            (Prop::F32(a), Prop::F32(b)) => Some(Prop::F32(a + b)),
+            (Prop::F64(a), Prop::F64(b)) => Some(Prop::F64(a + b)),
+            (Prop::Str(a), Prop::Str(b)) => Some(Prop::Str(a + &b)),
+            // You can add more combinations here as needed.
+            _ => None,
+        }
+    }
 }
 
 pub trait PropUnwrap: Sized {
