### 2023-08-24 3d25343
fixed getter for edges with small annotation, added sum for valuelist and valuelistlist, fixed internal error with empty result, fixed none appearing in empty lists when doing a sum
