### 2023-10-10 #5633 @narnolddd
@luizirber just to confirm, the DOIs seem to be okay now.
