### 2023-10-11 #5940 @editorialbot
[REVIEW]: Raphtory: The temporal graph engine for Rust and Python
**Submitting author:** @narnolddd (<a href="http://orcid.org/0000-0001-6396-4788">Naomi Arnold</a>)
**Repository:** https://github.com/Pometry/Raphtory
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.7.0
