### 2023-10-14 Pometry/Raphtory#1332 @abhishektiwari
Add details on supported Python and Rust versions
Provide details on supported Python and Rust versions in documentation and `README`. Trying to run Python examples with `3.11` and they don't work out of the box.

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5940)
