### 2023-10-14 Pometry/Raphtory#1333 @abhishektiwari
Broken LOTR Example
Running LOTR example (`Raphtory/examples/py/lotr `) using Python `3.11.4` with `raphtory==0.6.1`. Example worked after several changes based on latest API documentation.

## 1

**Code Block:**
```
graph = Graph(1)
```

```

**Error:**
Cell In[1], line 7
      5 structure_file = "lotr.csv"
      6 properties_file = "lotr_properties.csv"
----> 7 graph = Graph(1)

TypeError: Graph.__new__() takes 0 positional arguments but 1 was given
```
## 2
**Code Block:**

```
print("Number of edges: %i" % len(graph.edges))
```

**Error:**

```
Cell In[30], line 5
      3 print("Number of vertices: %i" % len(graph.vertices))
      4 print(graph.edges)
----> 5 print("Number of edges: %i" % len(graph.edges))

TypeError: object of type 'builtins.Edges' has no len()
```

Example should be using `graph.count_edges()`

## 3
**Code Block:**

```
with open(properties_file, 'r') as csvfile:
    datareader = csv.reader(csvfile)
    for row in datareader:
        graph.add_vertex_properties(row[0], {"race": row[1],"gender": row[2]})
```

**Error:**

```
Cell In[28], line 4
      2 datareader = csv.reader(csvfile)
      3 for row in datareader:
----> 4     graph.add_vertex_properties(row[0], {"race": row[1],"gender": row[2]})

AttributeError: 'builtins.Graph' object has no attribute 'add_vertex_properties'
```

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5940)
