### 2023-10-14 #5940 @abhishektiwari
@narnolddd please see initial feedback.

## General checks
### Contribution and authorship
— I was unable to find GitHub commits from Matt Barnes on Raphtory. Can you please comment on Matt's contribution for this paper and software?

## Documentation
###  Installation instructions
Please add guidance for Python and Rust version supported. Also specify version requirements for `examples` either in folder readme or as an inline comment.

### Example usage
A tried many of the examples, and they seem to be outdated and broken ([1](https://github.com/Pometry/Raphtory/issues/1332)).

For instance, due to unclear guidance on what Python versions are supported, I started with `3.11.4` and have to make several modifications to get `NFT` Python example working. Opened a few tickets based on examples [2](https://github.com/Pometry/Raphtory/issues/1332). I would like to see at-least Python `LOTR` example working.

## Software paper
### Quality of writing
Small corrections,

— `platform`  not `platformß`
— ` Similarly, DyNetX` not  `Similarly DyNetX`

### State of field
I am curious how Raphtory compares with graph databases such as arangodb/neo4j/MemGraph? I can see comparison benchmark data and scripts providing data points against MemGraph but no mention in the paper itself how Raphtory stack up against graph databases.
