### 2023-10-16 Pometry/Raphtory#1332 @miratepuffin
Hi @abhishektiwari - the examples in the documentation (https://www.raphtory.com/user-guide/installation/) are all working.

Are you discussing the [examples](https://github.com/Pometry/Raphtory/tree/master/examples) within the repo? Some of these are indeed a little old and need to be cleaned up.  Or are you reffering to the code snippets within the paper? Our API has changed slightly since submission.
