### 2023-10-16 Pometry/Raphtory#1333 @miratepuffin
Hi @abhishektiwari - as mentioned in #1332, since creating the documentation on raphtory.com this examples folder has become more of a test space than explicit using facing examples. I shall tidy this up today.
