### 2023-10-30 #5940 @arashbm
Just a quick feedback on the paper: Great work, smooth installation, and enjoyable use. The manuscript is concise and checks all the right boxes. My only gripe is with the examples.

* The examples are not self-contained. I understand that more involved use cases might be hard to condense to a few lines of code but I think they should still be run-able on their own. For example, it is not immediately clear where `enron` in example one and `global_temporal_three_node_motifs` in the third example are imported from. Example three, however, imports  `local_temporal_three_node_motifs` and never uses it.
* It would be better if the examples would become more readable. I leave that to the discretion of the authors, but I would recommend smaller output figures and paddings in exchange for larger font sizes. Remember that for someone browsing the paper that might be their first point of contact with the software. It is a chance for you to show its elegance and ease of use.
