### 2023-11-26 #5940 @arashbm
@narnolddd I don't seem to be able to run `from raphtory.plottingutils import global_motif_heatplot` from the examples. Is this the API for the next version or am I missing something on my setup?
