### 2023-11-27 #5940 @narnolddd
Hi @arashbm , so sorry I should have added a comment earlier, when I regenerated the pdf it was because I was struggling a bit with the package to do it locally (have managed to do it now) and wasn't ready to read. Hope I didn't waste too much of your time with this! The code referenced in the example is at the moment just on a branch which is going to be merged today and will be put into the next release. I also have a reply to both reviews which I will share once that's in. Apologies again for this and thank you for all your time reading/testing the code and the feedback you've given already.
