### 2024-01-05 #5940 @narnolddd
# Raphtory review response

Hi @arashbm @abhishektiwari Happy New Year and thank you very much for your time and feedback on the paper. Apologies for the delay in getting back to you, there were a few other changes (unrelated to this submission) that needed to be made/fixed in the repository before we could push a release. This has now been done (Raphtory 0.7.0) as well as the requested changes to the manuscript and README. We have addressed the comments point by point below, but in summary have:
* removed references to outdated examples in the readme and brought the existing examples up to date
* made the examples in the paper more self-contained, easier to follow and with more generous font sizes
* added instructions for the correct Python/Rust versions
* fixed some small grammatical errors/typos

## Response to comments from Abhishek

> I was unable to find GitHub commits from Matt Barnes on Raphtory. Can you please comment on Matt's contribution for this paper and software?

Matt may not have committed directly to Raphtory but has been working on some of the academic/research use cases for the software and his input has been crucial for the design of new features (as is the case for some other authors on the paper).

> Please add guidance for Python and Rust version supported. Also specify version requirements for examples either in folder readme or as an inline comment.

Thanks, we have now added the Python and Rust version requirement in the README (>=3.8 and >=1.75 respectively).

> I tried many of the examples, and they seem to be outdated and broken (1).
> For instance, due to unclear guidance on what Python versions are supported, I started with 3.11.4 and have to make several modifications to get NFT Python example working. Opened a few tickets based on examples 2. I would like to see at-least Python LOTR example working.

Apologies for this, they had become out of date and should have been deleted as we had got new versions working elsewhere. We now have one self-contained and comprehensive example (Sociopatterns baboon dataset) for Python.

> Small corrections

Thank you for spotting those typos, these have now been fixed.

> I am curious how Raphtory compares with graph databases such as arangodb/neo4j/MemGraph? I can see comparison benchmark data and scripts providing data points against MemGraph but no mention in the paper itself how Raphtory stack up against graph databases.

This is definitely something would like to address in a longer-form paper to give it an in-depth treatment, both due to the space resrictions of the JOSS format and because JOSS' scope seemed to be more focused on highlighting features and general usefulness than performance testing.

## Response to comments from Arash

> The examples are not self-contained. I understand that more involved use cases might be hard to condense to a few lines of code but I think they should still be run-able on their own. For example, it is not immediately clear where enron in example one and global_temporal_three_node_motifs in the third example are imported from. Example three, however, imports local_temporal_three_node_motifs and never uses it.

Thanks for your feedback. We have tried to make them a little more contained -- they are not entirely independent but assuming the user is familiar with Pandas and Matplotlib it should be clearer now where things are coming from:
* In the first example, we have shown that `enron` is a dataframe read from a csv with given column names.
* We made a small python module called `plottingutils` within Raphtory so that some of the more complicated plots like the motif one with the pictures on the axes can be done out of the box. For example 3 we have imported some functions from there so it is hopefully a bit clearer.
* The only bit we have left the user to fill in the blanks with is `plotting_function` in example 2 due to space reasons, this is some boilerplate pandas/matplotlib code for plotting multiple columns of a dataframe which is simple enough but takes up a lot of space and this was the only way we could think of to get around it. Hoping it might be obvious enough for a user to proceed but would welcome any suggestions.

> It would be better if the examples would become more readable. I leave that to the discretion of the authors, but I would recommend smaller output figures and paddings in exchange for larger font sizes. Remember that for someone browsing the paper that might be their first point of contact with the software. It is a chance for you to show its elegance and ease of use.

Thanks, we agree. We have tried to adjust the spacing a bit to make the fonts larger and allow for some gaps between logical sections of the code (balanced against keping the examples self-contained).
