### 2024-01-18 #5940 @abhishektiwari
@narnolddd Revised manuscript and updates look great. Look forward to Raphtory being used both in academia and industry. Thanks to everyone who has work on the software and paper.

@luizirber Review checklist is now complete from my side. Please let me know if anything else needed.
