### 2024-01-18 #5940 @luizirber
@narnolddd I posted the post-review checklist, I checked the repo and didn't see the zenodo integration, so can you set that up? More info here: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
You can also use figshare, or other archival solutions that offer a DOI.
