### 2024-01-19 #5940 @narnolddd
Hi @luizirber @arashbm @abhishektiwari thank you so much again for the great reviews and for your time editing this. We have made a Zenodo archive of Raphtory 0.7.0 which is the version we released with our response to the review, here is the link https://zenodo.org/records/10530614 and DOI [10.5281/zenodo.10530613](https://zenodo.org/doi/10.5281/zenodo.10530613). Is this ok? It's not integrated with the Github repo, we had just recently done a Raphtory release and it looked like we would need to do another one for the integrated one to work. If not we can do another release on the weekend.
