### 2024-01-22 #5940 @narnolddd
Thanks @luizirber for that! I did a check of authors; all names and affiliations are correct and in the same order as the Zenodo archive. All authors who have an Orcid ID now have it in their metadata. Both the repo and the archive have the gnu GPL v3 license.
