### 2024-02-03 #5940 @luizirber
Hey @narnolddd! The ref check didn't pick up these three cases, some suggestions to improve them:

> Hagberg, A., Swart, P., & S Chult, D. (2008). Exploring network structure, dynamics, and function using NetworkX. Los Alamos National Lab.(LANL), Los Alamos, NM (United States)

This one doesn't have a DOI, but has a (probably?) stable URI: https://www.osti.gov/biblio/960616

> Yousaf, H., Arnold, N. A., Lambiotte, R., LaRock, T., Clegg, R. G., Zhong, P., Alnaimi, A., & Steer, B. (2023). Non-Markovian paths and cycles in NFT trades. arXiv Preprint arXiv:2303.11181

set this DOI to https://doi.org/10.48550/arXiv.2303.11181

> Zhang, W. (2015). Dynamic network analysis in Julia

Again no DOI, but (hopefully) stable URI: https://eprints.maths.manchester.ac.uk/2376/

---

For 1 and 3 maybe use [`howpublished`](https://github.com/Pometry/Raphtory/blob/fd2ef291c5d6f886c337774a8d566b79f9569eb3/paper/joss-raphtory.bib#L267) like you did with `RecallGraph` so the stable-ish URI also show in the paper? Would this be an acceptable solution to cite resources without DOIs, @openjournals/dsais-eics ?
