### 2024-02-13 #5940 @luizirber
:point_up: pinging @arfon for the question on citing resources without DOIs =]
