### 2024-02-13 #5940 @narnolddd
Hi there @luizirber, apologies for the late reply, just got back from holiday. I've added the missing DOI and added a URL to the other two refs as suggested so that a link is visible on the pdf for each citation.
