### 2024-02-23 #5940 @narnolddd
Hey @luizirber just checking if you had chance to take a look at the references we've revised? Out of curiosity I took a look at the first few papers on the JOSS landing page and saw some examples of papers with URLs/URIs where a DOI was not possible:

[https://joss.theoj.org/papers/10.21105/joss.05566](https://joss.theoj.org/papers/10.21105/joss.05566)
[https://joss.theoj.org/papers/10.21105/joss.06270](https://joss.theoj.org/papers/10.21105/joss.06270)
[https://joss.theoj.org/papers/10.21105/joss.06011](https://joss.theoj.org/papers/10.21105/joss.06011)

Thanks and have a great weekend!
