### 2024-03-06 #5940 @arfon
> ☝️ pinging @arfon for the question on citing resources without DOIs =]

Sorry I missed this. I think this is fine to move forward @luizirber.
