### 2024-03-27 #5940 @arfon
@narnolddd – could you please merge this PR? https://github.com/Pometry/Raphtory/pull/1551
